+++ core/module_shared.h
@@ -1,5 +1,5 @@
 /* **********************************************************
- * Copyright (c) 2011-2016 Google, Inc.  All rights reserved.
+ * Copyright (c) 2011-2019 Google, Inc.  All rights reserved.
  * Copyright (c) 2008-2010 VMware, Inc.  All rights reserved.
  * **********************************************************/
 
@@ -536,6 +536,12 @@ typedef enum {
     /* These are for use with dr_map_executable_file(): */
     MODLOAD_NOT_PRIVLIB = 0x0002,
     MODLOAD_SKIP_WRITABLE = 0x0004, /* ignored on Windows */
+    /* For use with privload_map_and_relocate() or elf_loader_map_phdrs().
+     * Places an extra no-access page after .bss.
+     */
+    MODLOAD_SEPARATE_BSS = 0x0008,
+    /* Indicates that the module is being loaded in another process. */
+    MODLOAD_SEPARATE_PROCESS = 0x0010,
 } modload_flags_t;
 
 /* This function is used for loading non-private libs as well as private. */
