+++ core/unix/module.h
@@ -1,5 +1,5 @@
 /* **********************************************************
- * Copyright (c) 2011-2018 Google, Inc.  All rights reserved.
+ * Copyright (c) 2011-2019 Google, Inc.  All rights reserved.
  * Copyright (c) 2008-2010 VMware, Inc.  All rights reserved.
  * **********************************************************/
 
@@ -195,11 +195,6 @@ get_private_library_address(app_pc modbase, const char *name);
 
 bool
 get_private_library_bounds(IN app_pc modbase, OUT byte **start, OUT byte **end);
-
-typedef byte *(*map_fn_t)(file_t f, size_t *size INOUT, uint64 offs, app_pc addr,
-                          uint prot /*MEMPROT_*/, map_flags_t map_flags);
-typedef bool (*unmap_fn_t)(byte *map, size_t size);
-typedef bool (*prot_fn_t)(byte *map, size_t size, uint prot /*MEMPROT_*/);
 
 #ifdef MACOS
 /* module_macho.c */
