+++ core/unix/module_macho.c
@@ -1,5 +1,5 @@
 /* *******************************************************************************
- * Copyright (c) 2013-2015 Google, Inc.  All rights reserved.
+ * Copyright (c) 2013-2019 Google, Inc.  All rights reserved.
  * *******************************************************************************/
 
 /*
@@ -61,80 +61,6 @@
 #include <string.h> /* strcmp */
 #include <dlfcn.h>
 
-#ifdef NOT_DYNAMORIO_CORE_PROPER
-#    undef LOG
-#    define LOG(...)         /* nothing */
-#    include <sys/utsname.h> /* for kernel_is_64bit */
-#    undef ASSERT
-#    define ASSERT(x) ((void)0)
-#    undef ASSERT_NOT_IMPLEMENTED
-#    define ASSERT_NOT_IMPLEMENTED(x) ((void)0)
-
-/* Compatibility layer
- * XXX i#1409: really we should split out a lib shared w/ non-core from os.c.
- */
-bool
-kernel_is_64bit(void)
-{
-    struct utsname uinfo;
-    if (uname(&uinfo) != 0)
-        return true; /* guess */
-    return (strncmp(uinfo.machine, "x86_64", sizeof("x86_64")) == 0);
-}
-
-bool
-safe_read(const void *base, size_t size, void *out_buf)
-{
-    memcpy(out_buf, base, size);
-    return true;
-}
-#endif
-
-/* XXX i#1345: support mixed-mode 32-bit and 64-bit in one process.
- * There is no official support for that on Linux or Windows and for now we do
- * not support it either, especially not mixing libraries.
- */
-#ifdef X64
-typedef struct mach_header_64 mach_header_t;
-typedef struct segment_command_64 segment_command_t;
-typedef struct nlist_64 nlist_t;
-#else
-typedef struct mach_header mach_header_t;
-typedef struct segment_command segment_command_t;
-typedef struct nlist nlist_t;
-#endif
-
-/* Like is_elf_so_header(), if size == 0 then safe-reads the header; else
- * assumes that [base, base+size) is readable.
- */
-static bool
-is_macho_header(app_pc base, size_t size)
-{
-    struct mach_header hdr_safe;
-    struct mach_header *hdr;
-    if (base == NULL)
-        return false;
-    if (size >= sizeof(hdr_safe)) {
-        hdr = (struct mach_header *)base;
-    } else {
-        if (!safe_read(base, sizeof(hdr_safe), &hdr_safe))
-            return false;
-        hdr = &hdr_safe;
-    }
-    ASSERT(offsetof(struct mach_header, filetype) ==
-           offsetof(struct mach_header_64, filetype));
-    if ((hdr->magic == MH_MAGIC && hdr->cputype == CPU_TYPE_X86) ||
-        (hdr->magic == MH_MAGIC_64 && hdr->cputype == CPU_TYPE_X86_64)) {
-        /* We shouldn't see MH_PRELOAD as it can't be loaded by the kernel */
-        if (hdr->filetype == MH_EXECUTE || hdr->filetype == MH_DYLIB ||
-            hdr->filetype == MH_BUNDLE || hdr->filetype == MH_DYLINKER ||
-            hdr->filetype == MH_FVMLIB) {
-            return true;
-        }
-    }
-    return false;
-}
-
 bool
 module_file_has_module_header(const char *filename)
 {
@@ -157,7 +83,6 @@ module_is_partial_map(app_pc base, size_t size, uint memprot)
     return false;
 }
 
-#ifndef NOT_DYNAMORIO_CORE_PROPER
 bool
 module_walk_program_headers(app_pc base, size_t view_size, bool at_map, bool dyn_reloc,
                             OUT app_pc *out_base /* relative pc */,
@@ -318,7 +243,6 @@ module_num_program_headers(app_pc base)
     ASSERT_NOT_IMPLEMENTED(false); /* FIXME i#58: implement MachO support */
     return 0;
 }
-#endif /* !NOT_DYNAMORIO_CORE_PROPER */
 
 bool
 module_read_program_header(app_pc base, uint segment_num,
@@ -386,7 +310,6 @@ module_is_executable(app_pc base)
     return (hdr->filetype == MH_EXECUTE);
 }
 
-#ifndef NOT_DYNAMORIO_CORE_PROPER
 /* ULEB128 is a little-endian 128-base encoding where msb is set if there's
  * another byte of data to add to the integer represented.
  */
@@ -547,91 +470,12 @@ get_proc_address(module_base_t lib, const char *name)
 {
     return get_proc_address_ex(lib, name, NULL);
 }
-#endif /* !NOT_DYNAMORIO_CORE_PROPER */
 
 size_t
 module_get_header_size(app_pc module_base)
 {
     ASSERT_NOT_IMPLEMENTED(false); /* FIXME i#58: implement MachO support */
     return 0;
-}
-
-static bool
-platform_from_macho(file_t f, dr_platform_t *platform)
-{
-    struct mach_header mach_hdr;
-    if (os_read(f, &mach_hdr, sizeof(mach_hdr)) != sizeof(mach_hdr))
-        return false;
-    if (!is_macho_header((app_pc)&mach_hdr, sizeof(mach_hdr)))
-        return false;
-    switch (mach_hdr.cputype) {
-    case CPU_TYPE_X86_64: *platform = DR_PLATFORM_64BIT; break;
-    case CPU_TYPE_X86: *platform = DR_PLATFORM_32BIT; break;
-    default: return false;
-    }
-    return true;
-}
-
-bool
-module_get_platform(file_t f, dr_platform_t *platform, dr_platform_t *alt_platform)
-{
-    struct fat_header fat_hdr;
-    /* Both headers start with a 32-bit magic # */
-    uint32_t magic;
-    if (os_read(f, &magic, sizeof(magic)) != sizeof(magic) || !os_seek(f, 0, SEEK_SET))
-        return false;
-    if (magic == FAT_CIGAM) { /* big-endian */
-        /* This is a "fat" or "universal" binary */
-        struct fat_arch arch;
-        uint num, i;
-        bool found_main = false, found_alt = false;
-        dr_platform_t local_alt = DR_PLATFORM_NONE;
-        int64 cur_pos;
-        if (os_read(f, &fat_hdr, sizeof(fat_hdr)) != sizeof(fat_hdr))
-            return false;
-        /* OSSwapInt32 is a macro, so there's no lib dependence here */
-        num = OSSwapInt32(fat_hdr.nfat_arch);
-        for (i = 0; i < num; i++) {
-            if (os_read(f, &arch, sizeof(arch)) != sizeof(arch))
-                return false;
-            cur_pos = os_tell(f);
-            /* The primary platform is the one that will be used on an execve,
-             * which is the one that matches the kernel's bitwidth.
-             */
-            if ((kernel_is_64bit() && OSSwapInt32(arch.cputype) == CPU_TYPE_X86_64) ||
-                (!kernel_is_64bit() && OSSwapInt32(arch.cputype) == CPU_TYPE_X86)) {
-                /* Line up right before the Mach-O header */
-                if (!os_seek(f, OSSwapInt32(arch.offset), SEEK_SET))
-                    return false;
-                if (!platform_from_macho(f, platform))
-                    return false;
-                found_main = true;
-                if (found_alt)
-                    break;
-            } else if (OSSwapInt32(arch.cputype) == CPU_TYPE_X86_64 ||
-                       OSSwapInt32(arch.cputype) == CPU_TYPE_X86) {
-                /* Line up right before the Mach-O header */
-                if (!os_seek(f, OSSwapInt32(arch.offset), SEEK_SET))
-                    return false;
-                if (platform_from_macho(f, &local_alt)) {
-                    found_alt = true;
-                    if (found_main)
-                        break;
-                }
-            }
-            if (!os_seek(f, cur_pos, SEEK_SET))
-                return false;
-        }
-        if (!found_main && found_alt) {
-            *platform = local_alt;
-            local_alt = DR_PLATFORM_NONE;
-        }
-        if (alt_platform != NULL)
-            *alt_platform = local_alt;
-        return (found_main || found_alt);
-    } else if (alt_platform != NULL)
-        *alt_platform = DR_PLATFORM_NONE;
-    return platform_from_macho(f, platform);
 }
 
 bool
@@ -656,7 +500,6 @@ module_read_os_data(app_pc base, bool dyn_reloc, OUT ptr_int_t *load_delta,
     return false;
 }
 
-#ifndef NOT_DYNAMORIO_CORE_PROPER
 char *
 get_shared_lib_name(app_pc map)
 {
@@ -683,13 +526,12 @@ module_get_os_privmod_data(app_pc base, size_t size, bool dyn_reloc,
 byte *
 module_dynamorio_lib_base(void)
 {
-#    if defined(STATIC_LIBRARY) || defined(STANDALONE_UNIT_TEST)
+#if defined(STATIC_LIBRARY) || defined(STANDALONE_UNIT_TEST)
     return (byte *)&_mh_execute_header;
-#    else
+#else
     return (byte *)&_mh_dylib_header;
-#    endif
+#endif
 }
-#endif /* !NOT_DYNAMORIO_CORE_PROPER */
 
 ptr_uint_t
 module_get_text_section(app_pc file_map, size_t file_size)
@@ -698,7 +540,6 @@ module_get_text_section(app_pc file_map, size_t file_size)
     return 0;
 }
 
-#ifndef NOT_DYNAMORIO_CORE_PROPER
 bool
 module_dyld_shared_region(app_pc *start OUT, app_pc *end OUT)
 {
@@ -803,4 +644,3 @@ module_walk_dyld_list(app_pc dyld_base)
                         modinfo->imageFilePath, 0);
     }
 }
-#endif /* !NOT_DYNAMORIO_CORE_PROPER */
