+++ core/unix/os.c
@@ -1,5 +1,5 @@
 /* *******************************************************************************
- * Copyright (c) 2010-2018 Google, Inc.  All rights reserved.
+ * Copyright (c) 2010-2019 Google, Inc.  All rights reserved.
  * Copyright (c) 2011 Massachusetts Institute of Technology  All rights reserved.
  * Copyright (c) 2000-2010 VMware, Inc.  All rights reserved.
  * *******************************************************************************/
@@ -117,10 +117,6 @@ typedef struct rlimit64 rlimit64_t;
 #include "module.h" /* elf */
 #include "tls.h"
 
-#ifdef LINUX
-#    include "module_private.h" /* for ELF_AUXV_TYPE and AT_PAGESZ */
-#endif
-
 #if defined(X86) && defined(DEBUG)
 #    include "os_asm_defines.asm" /* for TLS_SELF_OFFSET_ASM */
 #endif
@@ -203,17 +199,6 @@ char **our_environ;
 #    include "instrument.h"
 #endif
 
-/* Cross arch syscall nums for use with struct stat64. */
-#ifdef X64
-#    ifdef SYS_stat
-#        define SYSNUM_STAT SYS_stat
-#    endif
-#    define SYSNUM_FSTAT SYS_fstat
-#else
-#    define SYSNUM_STAT SYS_stat64
-#    define SYSNUM_FSTAT SYS_fstat64
-#endif
-
 #ifdef MACOS
 #    define SYSNUM_EXIT_PROCESS SYS_exit
 #    define SYSNUM_EXIT_THREAD SYS_bsdthread_terminate
@@ -228,21 +213,6 @@ char **our_environ;
 #    define PR_SET_VMA_ANON_NAME 0
 #endif
 
-#ifdef NOT_DYNAMORIO_CORE_PROPER
-#    undef ASSERT
-#    undef ASSERT_NOT_IMPLEMENTED
-#    undef ASSERT_NOT_TESTED
-#    undef ASSERT_CURIOSITY
-#    define ASSERT(x)                 /* nothing */
-#    define ASSERT_NOT_IMPLEMENTED(x) /* nothing */
-#    define ASSERT_NOT_TESTED(x)      /* nothing */
-#    define ASSERT_CURIOSITY(x)       /* nothing */
-#    undef LOG
-#    undef DOSTATS
-#    define LOG(...)     /* nothing */
-#    define DOSTATS(...) /* nothing */
-#else /* !NOT_DYNAMORIO_CORE_PROPER: around most of file, to exclude preload */
-
 /* Guards data written by os_set_app_thread_area(). */
 DECLARE_CXTSWPROT_VAR(static mutex_t set_thread_area_lock,
                       INIT_LOCK_FREE(set_thread_area_lock));
@@ -252,12 +222,12 @@ static bool last_thread_tls_exited;
 
 tls_type_t tls_global_type;
 
-#    ifndef HAVE_TLS
+#ifndef HAVE_TLS
 /* We use a table lookup to find a thread's dcontext */
 /* Our only current no-TLS target, VMKernel (VMX86_SERVER), doesn't have apps with
  * tons of threads anyway
  */
-#        define MAX_THREADS 512
+#    define MAX_THREADS 512
 typedef struct _tls_slot_t {
     thread_id_t tid;
     dcontext_t *dcontext;
@@ -266,19 +236,19 @@ typedef struct _tls_slot_t {
 static tls_slot_t *tls_table;
 /* not static so deadlock_avoidance_unlock() can look for it */
 DECLARE_CXTSWPROT_VAR(mutex_t tls_lock, INIT_LOCK_FREE(tls_lock));
-#    endif
+#endif
 
-#    ifdef CLIENT_INTERFACE
+#ifdef CLIENT_INTERFACE
 /* Should we place this in a client header?  Currently mentioned in
  * dr_raw_tls_calloc() docs.
  */
 static bool client_tls_allocated[MAX_NUM_CLIENT_TLS];
 DECLARE_CXTSWPROT_VAR(static mutex_t client_tls_lock, INIT_LOCK_FREE(client_tls_lock));
-#    endif
+#endif
 
-#    include <stddef.h> /* for offsetof */
+#include <stddef.h> /* for offsetof */
 
-#    include <sys/utsname.h> /* for struct utsname */
+#include <sys/utsname.h> /* for struct utsname */
 
 /* forward decl */
 static void
@@ -287,11 +257,11 @@ static bool
 os_switch_lib_tls(dcontext_t *dcontext, bool to_app);
 static bool
 os_switch_seg_to_context(dcontext_t *dcontext, reg_id_t seg, bool to_app);
-#    ifdef X86
+#ifdef X86
 static bool
 os_set_dr_tls_base(dcontext_t *dcontext, os_local_state_t *tls, byte *base);
-#    endif
-#    ifdef LINUX
+#endif
+#ifdef LINUX
 static bool
 handle_app_mremap(dcontext_t *dcontext, byte *base, size_t size, byte *old_base,
                   size_t old_size, uint old_prot, uint old_type);
@@ -304,7 +274,7 @@ static bool
 handle_restartable_region_syscall_pre(dcontext_t *dcontext);
 static void
 handle_restartable_region_syscall_post(dcontext_t *dcontext, bool success);
-#    endif
+#endif
 
 /* full path to our own library, used for execve */
 static char dynamorio_library_path[MAXIMUM_PATH]; /* just dir */
@@ -313,8 +283,8 @@ static char dynamorio_library_filepath[MAXIMUM_PATH];
 static char dynamorio_alt_arch_path[MAXIMUM_PATH];
 static char dynamorio_alt_arch_filepath[MAXIMUM_PATH]; /* just dir */
 /* Makefile passes us LIBDIR_X{86,64} defines */
-#    define DR_LIBDIR_X86 STRINGIFY(LIBDIR_X86)
-#    define DR_LIBDIR_X64 STRINGIFY(LIBDIR_X64)
+#define DR_LIBDIR_X86 STRINGIFY(LIBDIR_X86)
+#define DR_LIBDIR_X64 STRINGIFY(LIBDIR_X64)
 
 /* pc values delimiting dynamo dll image */
 static app_pc dynamo_dll_start = NULL;
@@ -336,9 +306,9 @@ pid_t pid_cached;
 
 static bool fault_handling_initialized;
 
-#    ifdef PROFILE_RDTSC
+#ifdef PROFILE_RDTSC
 uint kilo_hertz; /* cpu clock speed */
-#    endif
+#endif
 
 /* Xref PR 258731, dup of STDOUT/STDERR in case app wants to close them. */
 DR_API file_t our_stdout = STDOUT_FILENO;
@@ -354,29 +324,29 @@ static int min_dr_fd;
  * the table key is the fd and the payload is the set of DR_FILE_* flags.
  */
 static generic_table_t *fd_table;
-#    define INIT_HTABLE_SIZE_FD 6 /* should remain small */
-#    ifdef DEBUG
+#define INIT_HTABLE_SIZE_FD 6 /* should remain small */
+#ifdef DEBUG
 static int num_fd_add_pre_heap;
-#    endif
+#endif
 
-#    ifdef LINUX
+#ifdef LINUX
 /* i#1004: brk emulation */
 static byte *app_brk_map;
 static byte *app_brk_cur;
 static byte *app_brk_end;
-#    endif
+#endif
 
-#    ifdef MACOS
+#ifdef MACOS
 /* xref i#1404: we should expose these via the dr_get_os_version() API */
 static int macos_version;
-#        define MACOS_VERSION_HIGH_SIERRA 17
-#        define MACOS_VERSION_SIERRA 16
-#        define MACOS_VERSION_EL_CAPITAN 15
-#        define MACOS_VERSION_YOSEMITE 14
-#        define MACOS_VERSION_MAVERICKS 13
-#        define MACOS_VERSION_MOUNTAIN_LION 12
-#        define MACOS_VERSION_LION 11
-#    endif
+#    define MACOS_VERSION_HIGH_SIERRA 17
+#    define MACOS_VERSION_SIERRA 16
+#    define MACOS_VERSION_EL_CAPITAN 15
+#    define MACOS_VERSION_YOSEMITE 14
+#    define MACOS_VERSION_MAVERICKS 13
+#    define MACOS_VERSION_MOUNTAIN_LION 12
+#    define MACOS_VERSION_LION 11
+#endif
 
 static bool
 is_readable_without_exception_internal(const byte *pc, size_t size, bool query_os);
@@ -385,10 +355,10 @@ static void
 process_mmap(dcontext_t *dcontext, app_pc base, size_t size, uint prot,
              uint flags _IF_DEBUG(const char *map_type));
 
-#    ifdef LINUX
+#ifdef LINUX
 static char *
 read_proc_self_exe(bool ignore_cache);
-#    endif
+#endif
 
 /* Libc independent directory iterator, similar to readdir.  If we ever need
  * this on Windows we should generalize it and export it to clients.
@@ -422,19 +392,19 @@ app_pc vsyscall_syscall_end_pc = NULL;
 app_pc vsyscall_sysenter_return_pc = NULL;
 /* pc where our hook-displaced code was copied */
 app_pc vsyscall_sysenter_displaced_pc = NULL;
-#    define VSYSCALL_PAGE_START_HARDCODED ((app_pc)(ptr_uint_t)0xffffe000)
-#    ifdef X64
+#define VSYSCALL_PAGE_START_HARDCODED ((app_pc)(ptr_uint_t)0xffffe000)
+#ifdef X64
 /* i#430, in Red Hat Enterprise Server 5.6, vsyscall region is marked
  * not executable
  * ffffffffff600000-ffffffffffe00000 ---p 00000000 00:00 0  [vsyscall]
  */
-#        define VSYSCALL_REGION_MAPS_NAME "[vsyscall]"
-#    endif
+#    define VSYSCALL_REGION_MAPS_NAME "[vsyscall]"
+#endif
 /* i#1908: vdso and vsyscall are now split */
 app_pc vdso_page_start = NULL;
 size_t vdso_size = 0;
 
-#    if !defined(STANDALONE_UNIT_TEST) && !defined(STATIC_LIBRARY)
+#if !defined(STANDALONE_UNIT_TEST) && !defined(STATIC_LIBRARY)
 /* The pthreads library keeps errno in its pthread_descr data structure,
  * which it looks up by dispatching on the stack pointer.  This doesn't work
  * when within dynamo.  Thus, we define our own __errno_location() for use both
@@ -462,9 +432,9 @@ __errno_location(void)
         return &(dcontext->upcontext_ptr->dr_errno);
     }
 }
-#    endif /* !STANDALONE_UNIT_TEST && !STATIC_LIBRARY */
+#endif /* !STANDALONE_UNIT_TEST && !STATIC_LIBRARY */
 
-#    if defined(HAVE_TLS) && defined(CLIENT_INTERFACE)
+#if defined(HAVE_TLS) && defined(CLIENT_INTERFACE)
 /* i#598
  * (gdb) x/20i (*(errno_loc_t)0xf721e413)
  * 0xf721e413 <__errno_location>:       push   %ebp
@@ -502,7 +472,7 @@ our_libc_errno_loc(void)
         return NULL;
     return (int *)(app_tls + libc_errno_tls_offs);
 }
-#    endif
+#endif
 
 /* i#238/PR 499179: libc errno preservation
  *
@@ -542,7 +512,7 @@ get_libc_errno_location(bool do_init)
                 ASSERT(libc_errno_loc != NULL);
                 LOG(GLOBAL, LOG_THREADS, 2, "libc errno loc func: " PFX "\n",
                     libc_errno_loc);
-#    ifdef CLIENT_INTERFACE
+#ifdef CLIENT_INTERFACE
                 /* Currently, the DR is loaded by system loader and hooked up
                  * to app's libc.  So right now, we still need this routine.
                  * we can remove this after libc independency and/or
@@ -554,13 +524,13 @@ get_libc_errno_location(bool do_init)
                         found = false;
                     release_recursive_lock(&privload_lock);
                 }
-#    endif
+#endif
                 if (found)
                     break;
             }
         }
         module_iterator_stop(mi);
-#    if defined(HAVE_TLS) && defined(CLIENT_INTERFACE)
+#if defined(HAVE_TLS) && defined(CLIENT_INTERFACE)
         /* i#598: init the libc errno's offset.  If we didn't find libc above,
          * then we don't need to do this.
          */
@@ -570,7 +540,7 @@ get_libc_errno_location(bool do_init)
             libc_errno_tls_offs = (void *)libc_errno_loc() - priv_lib_tls_base;
             libc_errno_loc = &our_libc_errno_loc;
         }
-#    endif
+#endif
     }
     return libc_errno_loc;
 }
@@ -582,14 +552,14 @@ get_libc_errno_location(bool do_init)
 int
 get_libc_errno(void)
 {
-#    if defined(STANDALONE_UNIT_TEST) && (defined(MACOS) || defined(ANDROID))
+#if defined(STANDALONE_UNIT_TEST) && (defined(MACOS) || defined(ANDROID))
     return errno;
-#    else
-#        ifdef STANDALONE_UNIT_TEST
+#else
+#    ifdef STANDALONE_UNIT_TEST
     errno_loc_t func = __errno_location;
-#        else
+#    else
     errno_loc_t func = get_libc_errno_location(false);
-#        endif
+#    endif
     if (func == NULL) {
         /* libc hasn't been loaded yet or we're doing early injection. */
         return 0;
@@ -601,7 +571,7 @@ get_libc_errno(void)
             return *loc;
     }
     return 0;
-#    endif
+#endif
 }
 
 /* N.B.: pthreads has two other locations it keeps on a per-thread basis:
@@ -721,30 +691,30 @@ our_init(int argc, char **argv, char **envp)
      */
     extern void dynamorio_app_take_over(void);
     bool takeover = false;
-#    ifdef INIT_TAKE_OVER
+#ifdef INIT_TAKE_OVER
     takeover = true;
-#    endif
-#    ifdef VMX86_SERVER
+#endif
+#ifdef VMX86_SERVER
     /* PR 391765: take over here instead of using preload */
     takeover = os_in_vmkernel_classic();
-#    endif
-#    ifndef STATIC_LIBRARY
+#endif
+#ifndef STATIC_LIBRARY
     if (our_environ != NULL) {
         /* Set by dynamorio_set_envp above.  These should agree. */
         ASSERT(our_environ == envp);
     } else {
         our_environ = envp;
     }
-#    endif
+#endif
     /* if using preload, no -early_inject */
-#    ifdef STATIC_LIBRARY
+#ifdef STATIC_LIBRARY
     if (!takeover) {
         const char *takeover_env = getenv("DYNAMORIO_TAKEOVER_IN_INIT");
         if (takeover_env != NULL && strcmp(takeover_env, "1") == 0) {
             takeover = true;
         }
     }
-#    endif
+#endif
     if (takeover) {
         if (dynamorio_app_init() == 0 /* success */) {
             dynamorio_app_take_over();
@@ -753,26 +723,26 @@ our_init(int argc, char **argv, char **envp)
     return 0;
 }
 
-#    if defined(STATIC_LIBRARY) || defined(STANDALONE_UNIT_TEST)
+#if defined(STATIC_LIBRARY) || defined(STANDALONE_UNIT_TEST)
 /* If we're getting linked into a binary that already has an _init definition
  * like the app's exe or unit_tests, we add a pointer to our_init() to the
  * .init_array section.  We can't use the constructor attribute because not all
  * toolchains pass the args and environment to the constructor.
  */
 static init_fn_t
-#        ifdef MACOS
+#    ifdef MACOS
     __attribute__((section("__DATA,__mod_init_func"), aligned(sizeof(void *)), used))
-#        else
-    __attribute__((section(".init_array"), aligned(sizeof(void *)), used))
-#        endif
-    init_array[] = { our_init };
 #    else
+    __attribute__((section(".init_array"), aligned(sizeof(void *)), used))
+#    endif
+    init_array[] = { our_init };
+#else
 /* If we're a normal shared object, then we override _init.
  */
 int
 _init(int argc, char **argv, char **envp)
 {
-#        ifdef ANDROID
+#    ifdef ANDROID
     /* i#1862: the Android loader passes *nothing* to lib init routines.  We
      * rely on DR being listed before libc so we can read the TLS slot the
      * kernel set up.
@@ -784,10 +754,10 @@ _init(int argc, char **argv, char **envp)
         envp = NULL;
     }
     ASSERT_MESSAGE(CHKLVL_ASSERTS, "failed to find envp", envp != NULL);
-#        endif
+#    endif
     return our_init(argc, argv, envp);
 }
-#    endif
+#endif
 
 bool
 kernel_is_64bit(void)
@@ -795,7 +765,7 @@ kernel_is_64bit(void)
     return kernel_64bit;
 }
 
-#    ifdef MACOS
+#ifdef MACOS
 /* XXX: if we get enough of these, move to os_macos.c or sthg */
 static bool
 sysctl_query(int level0, int level1, void *buf, size_t bufsz)
@@ -808,7 +778,7 @@ sysctl_query(int level0, int level1, void *buf, size_t bufsz)
     res = dynamorio_syscall(SYS___sysctl, 6, &name, 2, buf, &len, NULL, 0);
     return (res >= 0);
 }
-#    endif
+#endif
 
 static void
 get_uname(void)
@@ -817,7 +787,7 @@ get_uname(void)
      * or .data unprot
      */
     static struct utsname uinfo; /* can be large, avoid stack overflow */
-#    ifdef MACOS
+#ifdef MACOS
     if (!sysctl_query(CTL_KERN, KERN_OSTYPE, &uinfo.sysname, sizeof(uinfo.sysname)) ||
         !sysctl_query(CTL_KERN, KERN_HOSTNAME, &uinfo.nodename, sizeof(uinfo.nodename)) ||
         !sysctl_query(CTL_KERN, KERN_OSRELEASE, &uinfo.release, sizeof(uinfo.release)) ||
@@ -826,11 +796,11 @@ get_uname(void)
         ASSERT(false && "sysctl queries failed");
         return;
     }
-#    else
+#else
     DEBUG_DECLARE(int res =)
     dynamorio_syscall(SYS_uname, 1, (ptr_uint_t)&uinfo);
     ASSERT(res >= 0);
-#    endif
+#endif
     LOG(GLOBAL, LOG_TOP, 1, "uname:\n\tsysname: %s\n", uinfo.sysname);
     LOG(GLOBAL, LOG_TOP, 1, "\tnodename: %s\n", uinfo.nodename);
     LOG(GLOBAL, LOG_TOP, 1, "\trelease: %s\n", uinfo.release);
@@ -838,14 +808,14 @@ get_uname(void)
     LOG(GLOBAL, LOG_TOP, 1, "\tmachine: %s\n", uinfo.machine);
     if (strncmp(uinfo.machine, "x86_64", sizeof("x86_64")) == 0)
         kernel_64bit = true;
-#    ifdef MACOS
+#ifdef MACOS
     /* XXX: I would skip these checks for standalone so we don't have to set env
      * vars for frontends to see the options but I'm still afraid of some syscall
      * crash with no output: I'd rather have two messages than silent crashing.
      */
     if (DYNAMO_OPTION(max_supported_os_version) != 0) { /* 0 disables */
         /* We only support OSX 10.7.5 - 10.9.1.  That means kernels 11.x-13.x. */
-#        define MIN_DARWIN_VERSION_SUPPORTED 11
+#    define MIN_DARWIN_VERSION_SUPPORTED 11
         int kernel_major;
         if (sscanf(uinfo.release, "%d", &kernel_major) != 1 ||
             kernel_major > DYNAMO_OPTION(max_supported_os_version) ||
@@ -856,7 +826,7 @@ get_uname(void)
         }
         macos_version = kernel_major;
     }
-#    endif
+#endif
 }
 
 /* os-specific initializations */
@@ -875,14 +845,14 @@ os_init(void)
      * or whether getpid suffices.  even 2.4 kernels have gettid
      * (maps to getpid), don't have an old enough target to test this.
      */
-#    ifdef MACOS
+#ifdef MACOS
     kernel_thread_groups = (dynamorio_syscall(SYS_thread_selfid, 0) >= 0);
-#    else
+#else
     kernel_thread_groups = (dynamorio_syscall(SYS_gettid, 0) >= 0);
-#    endif
+#endif
     LOG(GLOBAL, LOG_TOP | LOG_STATS, 1, "thread id is from %s\n",
         kernel_thread_groups ? "gettid" : "getpid");
-#    ifdef MACOS
+#ifdef MACOS
     /* SYS_thread_selfid was added in 10.6.  We have no simple way to get the
      * thread id on 10.5, so we don't support it.
      */
@@ -890,15 +860,15 @@ os_init(void)
         SYSLOG(SYSLOG_WARNING, UNSUPPORTED_OS_VERSION, 3, get_application_name(),
                get_application_pid(), "Mac OSX 10.5 or earlier");
     }
-#    else
+#else
     ASSERT_CURIOSITY(kernel_thread_groups);
-#    endif
+#endif
 
     pid_cached = get_process_id();
 
-#    ifdef VMX86_SERVER
+#ifdef VMX86_SERVER
     vmk_init();
-#    endif
+#endif
 
     signal_init();
     /* We now set up an early fault handler for safe_read() (i#350) */
@@ -906,13 +876,13 @@ os_init(void)
 
     memquery_init();
 
-#    ifdef PROFILE_RDTSC
+#ifdef PROFILE_RDTSC
     if (dynamo_options.profile_times) {
         ASSERT_NOT_TESTED();
         kilo_hertz = get_timer_frequency();
         LOG(GLOBAL, LOG_TOP | LOG_STATS, 1, "CPU MHz is %d\n", kilo_hertz / 1000);
     }
-#    endif /* PROFILE_RDTSC */
+#endif /* PROFILE_RDTSC */
     /* Needs to be after heap_init */
     IF_NO_MEMQUERY(memcache_init());
 
@@ -920,26 +890,26 @@ os_init(void)
     fd_table = generic_hash_create(
         GLOBAL_DCONTEXT, INIT_HTABLE_SIZE_FD, 80 /* load factor: not perf-critical */,
         HASHTABLE_SHARED | HASHTABLE_PERSISTENT, NULL _IF_DEBUG("fd table"));
-#    ifdef DEBUG
+#ifdef DEBUG
     if (GLOBAL != INVALID_FILE)
         fd_table_add(GLOBAL, OS_OPEN_CLOSE_ON_FORK);
-#    endif
+#endif
 
     /* Ensure initialization */
     get_dynamorio_dll_start();
 
-#    ifdef LINUX
+#ifdef LINUX
     if (DYNAMO_OPTION(emulate_brk))
         init_emulated_brk(NULL);
-#    endif
+#endif
 
-#    ifdef ANDROID
+#ifdef ANDROID
     /* This must be set up earlier than privload_tls_init, and must be set up
      * for non-client-interface as well, as this initializes DR_TLS_BASE_OFFSET
      * (i#1931).
      */
     init_android_version();
-#    endif
+#endif
 }
 
 /* called before any logfiles are opened */
@@ -1025,28 +995,28 @@ set_executable_path(const char *exe_path)
  * On recent XNU versions, the kernel now prefixes the executable path
  * with the string executable_path= so it can be parsed getenv style.
  */
-#    ifdef MACOS
-#        define EXECUTABLE_KEY "executable_path="
-#    endif
+#ifdef MACOS
+#    define EXECUTABLE_KEY "executable_path="
+#endif
 /* i#189: we need to re-cache after a fork */
 static char *
 get_application_name_helper(bool ignore_cache, bool full_path)
 {
     if (!executable_path[0] || ignore_cache) {
-#    ifdef VMX86_SERVER
+#ifdef VMX86_SERVER
         if (os_in_vmkernel_userworld()) {
             vmk_getnamefrompid(pid, executable_path, sizeof(executable_path));
         } else
-#    endif
+#endif
             if (DYNAMO_OPTION(early_inject)) {
             ASSERT(executable_path[0] != '\0' &&
                    "i#907: Can't read /proc/self/exe for early injection");
         } else {
-#    ifdef LINUX
+#ifdef LINUX
             /* Populate cache from /proc/self/exe link. */
             strncpy(executable_path, read_proc_self_exe(ignore_cache),
                     BUFFER_SIZE_ELEMENTS(executable_path));
-#    else
+#else
             /* OSX kernel puts full app exec path above envp */
             char *c, **env = our_environ;
             do {
@@ -1076,7 +1046,7 @@ get_application_name_helper(bool ignore_cache, bool full_path)
                          len > 0 ? "/" : "", c);
             } else
                 strncpy(executable_path, c, BUFFER_SIZE_ELEMENTS(executable_path));
-#    endif
+#endif
             NULL_TERMINATE_BUFFER(executable_path);
             /* FIXME: Fall back on /proc/self/cmdline and maybe argv[0] from
              * _init().
@@ -1114,9 +1084,9 @@ get_application_short_name(void)
 }
 
 /* Processor information provided by kernel */
-#    define PROC_CPUINFO "/proc/cpuinfo"
-#    define CPUMHZ_LINE_LENGTH 64
-#    define CPUMHZ_LINE_FORMAT "cpu MHz\t\t: %lu.%03lu\n"
+#define PROC_CPUINFO "/proc/cpuinfo"
+#define CPUMHZ_LINE_LENGTH 64
+#define CPUMHZ_LINE_FORMAT "cpu MHz\t\t: %lu.%03lu\n"
 /* printed in /usr/src/linux-2.4/arch/i386/kernel/setup.c calibrated in time.c */
 /* seq_printf(m, "cpu MHz\t\t: %lu.%03lu\n", cpu_khz / 1000, (cpu_khz % 1000)) */
 /* e.g. cpu MHz           : 1594.851 */
@@ -1162,18 +1132,18 @@ get_timer_frequency_cpuinfo(void)
 timestamp_t
 get_timer_frequency()
 {
-#    ifdef VMX86_SERVER
+#ifdef VMX86_SERVER
     if (os_in_vmkernel_userworld()) {
         return vmk_get_timer_frequency();
     }
-#    endif
+#endif
     return get_timer_frequency_cpuinfo();
 }
 
 /* DR has standardized on UTC time which counts from since Jan 1, 1601.
  * That's the Windows standard.  But Linux uses the Epoch of Jan 1, 1970.
  */
-#    define UTC_TO_EPOCH_SECONDS 11644473600
+#define UTC_TO_EPOCH_SECONDS 11644473600
 
 /* seconds since 1601 */
 uint
@@ -1181,14 +1151,14 @@ query_time_seconds(void)
 {
     struct timeval current_time;
     uint64 val = dynamorio_syscall(SYS_gettimeofday, 2, &current_time, NULL);
-#    ifdef MACOS
+#ifdef MACOS
     /* MacOS before Sierra returns usecs:secs and does not set the timeval struct. */
     if (macos_version < MACOS_VERSION_SIERRA) {
         if ((int)val < 0)
             return 0;
         return (uint)val + UTC_TO_EPOCH_SECONDS;
     }
-#    endif
+#endif
     if ((int)val >= 0) {
         return current_time.tv_sec + UTC_TO_EPOCH_SECONDS;
     } else {
@@ -1203,7 +1173,7 @@ query_time_millis()
 {
     struct timeval current_time;
     uint64 val = dynamorio_syscall(SYS_gettimeofday, 2, &current_time, NULL);
-#    ifdef MACOS
+#ifdef MACOS
     /* MacOS before Sierra returns usecs:secs and does not set the timeval struct. */
     if (macos_version < MACOS_VERSION_SIERRA) {
         if ((int)val > 0) {
@@ -1211,7 +1181,7 @@ query_time_millis()
             current_time.tv_usec = (uint)(val >> 32);
         }
     }
-#    endif
+#endif
     if ((int)val >= 0) {
         uint64 res =
             (((uint64)current_time.tv_sec) * 1000) + (current_time.tv_usec / 1000);
@@ -1229,7 +1199,7 @@ query_time_micros()
 {
     struct timeval current_time;
     uint64 val = dynamorio_syscall(SYS_gettimeofday, 2, &current_time, NULL);
-#    ifdef MACOS
+#ifdef MACOS
     /* MacOS before Sierra returns usecs:secs and does not set the timeval struct. */
     if (macos_version < MACOS_VERSION_SIERRA) {
         if ((int)val > 0) {
@@ -1237,7 +1207,7 @@ query_time_micros()
             current_time.tv_usec = (uint)(val >> 32);
         }
     }
-#    endif
+#endif
     if ((int)val >= 0) {
         uint64 res = (((uint64)current_time.tv_sec) * 1000000) + current_time.tv_usec;
         res += UTC_TO_EPOCH_SECONDS * 1000000;
@@ -1248,7 +1218,7 @@ query_time_micros()
     }
 }
 
-#    ifdef RETURN_AFTER_CALL
+#ifdef RETURN_AFTER_CALL
 /* Finds the bottom of the call stack, presumably at program startup. */
 /* This routine is a copycat of internal_dump_callstack and makes
    assumptions about program state, i.e. that frame pointers are valid
@@ -1276,7 +1246,7 @@ find_stack_bottom()
     }
     return retaddr;
 }
-#    endif /* RETURN_AFTER_CALL */
+#endif /* RETURN_AFTER_CALL */
 
 /* os-specific atexit cleanup */
 void
@@ -1295,9 +1265,9 @@ os_slow_exit(void)
     }
 
     DELETE_LOCK(set_thread_area_lock);
-#    ifdef CLIENT_INTERFACE
+#ifdef CLIENT_INTERFACE
     DELETE_LOCK(client_tls_lock);
-#    endif
+#endif
     IF_NO_MEMQUERY(memcache_exit());
 }
 
@@ -1376,24 +1346,24 @@ os_timeout(int time_in_milliseconds)
  * transparency becomes more of a problem.
  */
 
-#    define TLS_LOCAL_STATE_OFFSET (offsetof(os_local_state_t, state))
+#define TLS_LOCAL_STATE_OFFSET (offsetof(os_local_state_t, state))
 
 /* offset from top of page */
-#    define TLS_OS_LOCAL_STATE 0x00
+#define TLS_OS_LOCAL_STATE 0x00
 
-#    define TLS_SELF_OFFSET (TLS_OS_LOCAL_STATE + offsetof(os_local_state_t, self))
-#    define TLS_THREAD_ID_OFFSET (TLS_OS_LOCAL_STATE + offsetof(os_local_state_t, tid))
-#    define TLS_DCONTEXT_OFFSET (TLS_OS_LOCAL_STATE + TLS_DCONTEXT_SLOT)
-#    ifdef X86
-#        define TLS_MAGIC_OFFSET (TLS_OS_LOCAL_STATE + offsetof(os_local_state_t, magic))
-#    endif
+#define TLS_SELF_OFFSET (TLS_OS_LOCAL_STATE + offsetof(os_local_state_t, self))
+#define TLS_THREAD_ID_OFFSET (TLS_OS_LOCAL_STATE + offsetof(os_local_state_t, tid))
+#define TLS_DCONTEXT_OFFSET (TLS_OS_LOCAL_STATE + TLS_DCONTEXT_SLOT)
+#ifdef X86
+#    define TLS_MAGIC_OFFSET (TLS_OS_LOCAL_STATE + offsetof(os_local_state_t, magic))
+#endif
 
 /* they should be used with os_tls_offset, so do not need add TLS_OS_LOCAL_STATE here
  */
-#    define TLS_APP_LIB_TLS_BASE_OFFSET (offsetof(os_local_state_t, app_lib_tls_base))
-#    define TLS_APP_ALT_TLS_BASE_OFFSET (offsetof(os_local_state_t, app_alt_tls_base))
-#    define TLS_APP_LIB_TLS_REG_OFFSET (offsetof(os_local_state_t, app_lib_tls_reg))
-#    define TLS_APP_ALT_TLS_REG_OFFSET (offsetof(os_local_state_t, app_alt_tls_reg))
+#define TLS_APP_LIB_TLS_BASE_OFFSET (offsetof(os_local_state_t, app_lib_tls_base))
+#define TLS_APP_ALT_TLS_BASE_OFFSET (offsetof(os_local_state_t, app_alt_tls_base))
+#define TLS_APP_LIB_TLS_REG_OFFSET (offsetof(os_local_state_t, app_lib_tls_reg))
+#define TLS_APP_ALT_TLS_REG_OFFSET (offsetof(os_local_state_t, app_alt_tls_reg))
 
 /* N.B.: imm and offs are ushorts!
  * We use %c[0-9] to get gcc to emit an integer constant without a leading $ for
@@ -1405,108 +1375,102 @@ os_timeout(int time_in_milliseconds)
  * precise constraint, then the compiler would be able to optimize better.  See
  * glibc comments on THREAD_SELF.
  */
-#    ifdef X86
-#        define WRITE_TLS_SLOT_IMM(imm, var)       \
-            IF_NOT_HAVE_TLS(ASSERT_NOT_REACHED()); \
-            ASSERT(sizeof(var) == sizeof(void *)); \
-            asm volatile("mov %0, %" ASM_SEG ":%c1" : : "r"(var), "i"(imm));
+#ifdef X86
+#    define WRITE_TLS_SLOT_IMM(imm, var)       \
+        IF_NOT_HAVE_TLS(ASSERT_NOT_REACHED()); \
+        ASSERT(sizeof(var) == sizeof(void *)); \
+        asm volatile("mov %0, %" ASM_SEG ":%c1" : : "r"(var), "i"(imm));
 
-#        define READ_TLS_SLOT_IMM(imm, var)        \
-            IF_NOT_HAVE_TLS(ASSERT_NOT_REACHED()); \
-            ASSERT(sizeof(var) == sizeof(void *)); \
-            asm volatile("mov %" ASM_SEG ":%c1, %0" : "=r"(var) : "i"(imm));
+#    define READ_TLS_SLOT_IMM(imm, var)        \
+        IF_NOT_HAVE_TLS(ASSERT_NOT_REACHED()); \
+        ASSERT(sizeof(var) == sizeof(void *)); \
+        asm volatile("mov %" ASM_SEG ":%c1, %0" : "=r"(var) : "i"(imm));
 
-#        define WRITE_TLS_INT_SLOT_IMM(imm, var)   \
-            IF_NOT_HAVE_TLS(ASSERT_NOT_REACHED()); \
-            ASSERT(sizeof(var) == sizeof(int));    \
-            asm volatile("movl %0, %" ASM_SEG ":%c1" : : "r"(var), "i"(imm));
+#    define WRITE_TLS_INT_SLOT_IMM(imm, var)   \
+        IF_NOT_HAVE_TLS(ASSERT_NOT_REACHED()); \
+        ASSERT(sizeof(var) == sizeof(int));    \
+        asm volatile("movl %0, %" ASM_SEG ":%c1" : : "r"(var), "i"(imm));
 
-#        define READ_TLS_INT_SLOT_IMM(imm, var)    \
-            IF_NOT_HAVE_TLS(ASSERT_NOT_REACHED()); \
-            ASSERT(sizeof(var) == sizeof(int));    \
-            asm volatile("movl %" ASM_SEG ":%c1, %0" : "=r"(var) : "i"(imm));
+#    define READ_TLS_INT_SLOT_IMM(imm, var)    \
+        IF_NOT_HAVE_TLS(ASSERT_NOT_REACHED()); \
+        ASSERT(sizeof(var) == sizeof(int));    \
+        asm volatile("movl %" ASM_SEG ":%c1, %0" : "=r"(var) : "i"(imm));
 
 /* FIXME: need dedicated-storage var for _TLS_SLOT macros, can't use expr */
-#        define WRITE_TLS_SLOT(offs, var)                       \
-            IF_NOT_HAVE_TLS(ASSERT_NOT_REACHED());              \
-            ASSERT(sizeof(var) == sizeof(void *));              \
-            ASSERT(sizeof(offs) == 2);                          \
-            asm("mov %0, %%" ASM_XAX : : "m"((var)) : ASM_XAX); \
-            asm("movzw" IF_X64_ELSE("q", "l") " %0, %%" ASM_XDX \
-                :                                               \
-                : "m"((offs))                                   \
-                : ASM_XDX);                                     \
-            asm("mov %%" ASM_XAX ", %" ASM_SEG ":(%%" ASM_XDX ")" : : : ASM_XAX, ASM_XDX);
+#    define WRITE_TLS_SLOT(offs, var)                                                   \
+        IF_NOT_HAVE_TLS(ASSERT_NOT_REACHED());                                          \
+        ASSERT(sizeof(var) == sizeof(void *));                                          \
+        ASSERT(sizeof(offs) == 2);                                                      \
+        asm("mov %0, %%" ASM_XAX : : "m"((var)) : ASM_XAX);                             \
+        asm("movzw" IF_X64_ELSE("q", "l") " %0, %%" ASM_XDX : : "m"((offs)) : ASM_XDX); \
+        asm("mov %%" ASM_XAX ", %" ASM_SEG ":(%%" ASM_XDX ")" : : : ASM_XAX, ASM_XDX);
 
-#        define READ_TLS_SLOT(offs, var)                                       \
-            ASSERT(sizeof(var) == sizeof(void *));                             \
-            ASSERT(sizeof(offs) == 2);                                         \
-            asm("movzw" IF_X64_ELSE("q", "l") " %0, %%" ASM_XAX                \
-                :                                                              \
-                : "m"((offs))                                                  \
-                : ASM_XAX);                                                    \
-            asm("mov %" ASM_SEG ":(%%" ASM_XAX "), %%" ASM_XAX : : : ASM_XAX); \
-            asm("mov %%" ASM_XAX ", %0" : "=m"((var)) : : ASM_XAX);
-#    elif defined(AARCHXX)
+#    define READ_TLS_SLOT(offs, var)                                                    \
+        ASSERT(sizeof(var) == sizeof(void *));                                          \
+        ASSERT(sizeof(offs) == 2);                                                      \
+        asm("movzw" IF_X64_ELSE("q", "l") " %0, %%" ASM_XAX : : "m"((offs)) : ASM_XAX); \
+        asm("mov %" ASM_SEG ":(%%" ASM_XAX "), %%" ASM_XAX : : : ASM_XAX);              \
+        asm("mov %%" ASM_XAX ", %0" : "=m"((var)) : : ASM_XAX);
+#elif defined(AARCHXX)
 /* Android needs indirection through a global.  The Android toolchain has
  * trouble with relocations if we use a global directly in asm, so we convert to
  * a local variable in these macros.  We pay the cost of the extra instructions
  * for Linux ARM to share the code.
  */
-#        define WRITE_TLS_SLOT_IMM(imm, var)                                            \
-            do {                                                                        \
-                uint _base_offs = DR_TLS_BASE_OFFSET;                                   \
-                __asm__ __volatile__("mov " ASM_R2 ", %0 \n\t" READ_TP_TO_R3_DISP_IN_R2 \
-                                     "str %1, [" ASM_R3 ", %2] \n\t"                    \
-                                     :                                                  \
-                                     : "r"(_base_offs), "r"(var), "i"(imm)              \
-                                     : "memory", ASM_R2, ASM_R3);                       \
-            } while (0)
-#        define READ_TLS_SLOT_IMM(imm, var)                                             \
-            do {                                                                        \
-                uint _base_offs = DR_TLS_BASE_OFFSET;                                   \
-                __asm__ __volatile__("mov " ASM_R2 ", %1 \n\t" READ_TP_TO_R3_DISP_IN_R2 \
-                                     "ldr %0, [" ASM_R3 ", %2] \n\t"                    \
-                                     : "=r"(var)                                        \
-                                     : "r"(_base_offs), "i"(imm)                        \
-                                     : ASM_R2, ASM_R3);                                 \
-            } while (0)
-#        define WRITE_TLS_INT_SLOT_IMM WRITE_TLS_SLOT_IMM /* b/c 32-bit */
-#        define READ_TLS_INT_SLOT_IMM READ_TLS_SLOT_IMM   /* b/c 32-bit */
-#        define WRITE_TLS_SLOT(offs, var)                                               \
-            do {                                                                        \
-                uint _base_offs = DR_TLS_BASE_OFFSET;                                   \
-                __asm__ __volatile__("mov " ASM_R2 ", %0 \n\t" READ_TP_TO_R3_DISP_IN_R2 \
-                                     "add " ASM_R3 ", " ASM_R3 ", %2 \n\t"              \
-                                     "str %1, [" ASM_R3 "]   \n\t"                      \
-                                     :                                                  \
-                                     : "r"(_base_offs), "r"(var), "r"(offs)             \
-                                     : "memory", ASM_R2, ASM_R3);                       \
-            } while (0)
-#        define READ_TLS_SLOT(offs, var)                                                \
-            do {                                                                        \
-                uint _base_offs = DR_TLS_BASE_OFFSET;                                   \
-                __asm__ __volatile__("mov " ASM_R2 ", %1 \n\t" READ_TP_TO_R3_DISP_IN_R2 \
-                                     "add " ASM_R3 ", " ASM_R3 ", %2 \n\t"              \
-                                     "ldr %0, [" ASM_R3 "]   \n\t"                      \
-                                     : "=r"(var)                                        \
-                                     : "r"(_base_offs), "r"(offs)                       \
-                                     : ASM_R2, ASM_R3);                                 \
-            } while (0)
-#    endif /* X86/ARM */
+#    define WRITE_TLS_SLOT_IMM(imm, var)                                            \
+        do {                                                                        \
+            uint _base_offs = DR_TLS_BASE_OFFSET;                                   \
+            __asm__ __volatile__("mov " ASM_R2 ", %0 \n\t" READ_TP_TO_R3_DISP_IN_R2 \
+                                 "str %1, [" ASM_R3 ", %2] \n\t"                    \
+                                 :                                                  \
+                                 : "r"(_base_offs), "r"(var), "i"(imm)              \
+                                 : "memory", ASM_R2, ASM_R3);                       \
+        } while (0)
+#    define READ_TLS_SLOT_IMM(imm, var)                                             \
+        do {                                                                        \
+            uint _base_offs = DR_TLS_BASE_OFFSET;                                   \
+            __asm__ __volatile__("mov " ASM_R2 ", %1 \n\t" READ_TP_TO_R3_DISP_IN_R2 \
+                                 "ldr %0, [" ASM_R3 ", %2] \n\t"                    \
+                                 : "=r"(var)                                        \
+                                 : "r"(_base_offs), "i"(imm)                        \
+                                 : ASM_R2, ASM_R3);                                 \
+        } while (0)
+#    define WRITE_TLS_INT_SLOT_IMM WRITE_TLS_SLOT_IMM /* b/c 32-bit */
+#    define READ_TLS_INT_SLOT_IMM READ_TLS_SLOT_IMM   /* b/c 32-bit */
+#    define WRITE_TLS_SLOT(offs, var)                                               \
+        do {                                                                        \
+            uint _base_offs = DR_TLS_BASE_OFFSET;                                   \
+            __asm__ __volatile__("mov " ASM_R2 ", %0 \n\t" READ_TP_TO_R3_DISP_IN_R2 \
+                                 "add " ASM_R3 ", " ASM_R3 ", %2 \n\t"              \
+                                 "str %1, [" ASM_R3 "]   \n\t"                      \
+                                 :                                                  \
+                                 : "r"(_base_offs), "r"(var), "r"(offs)             \
+                                 : "memory", ASM_R2, ASM_R3);                       \
+        } while (0)
+#    define READ_TLS_SLOT(offs, var)                                                \
+        do {                                                                        \
+            uint _base_offs = DR_TLS_BASE_OFFSET;                                   \
+            __asm__ __volatile__("mov " ASM_R2 ", %1 \n\t" READ_TP_TO_R3_DISP_IN_R2 \
+                                 "add " ASM_R3 ", " ASM_R3 ", %2 \n\t"              \
+                                 "ldr %0, [" ASM_R3 "]   \n\t"                      \
+                                 : "=r"(var)                                        \
+                                 : "r"(_base_offs), "r"(offs)                       \
+                                 : ASM_R2, ASM_R3);                                 \
+        } while (0)
+#endif /* X86/ARM */
 
-#    ifdef X86
+#ifdef X86
 /* We use this at thread init and exit to make it easy to identify
  * whether TLS is initialized (i#2089).
  * We assume alignment does not matter.
  */
 static os_local_state_t uninit_tls; /* has .magic == 0 */
-#    endif
+#endif
 
 static bool
 is_thread_tls_initialized(void)
 {
-#    ifdef X86
+#ifdef X86
     if (INTERNAL_OPTION(safe_read_tls_init)) {
         /* Avoid faults during early init or during exit when we have no handler.
          * It's not worth extending the handler as the faults are a perf hit anyway.
@@ -1540,7 +1504,7 @@ is_thread_tls_initialized(void)
             /* XXX: make this a safe read: but w/o dcontext we need special asm support */
             READ_TLS_SLOT_IMM(TLS_SELF_OFFSET, os_tls);
         }
-#        ifdef X64
+#    ifdef X64
         if (os_tls == NULL && tls_dr_using_msr()) {
             /* When the MSR is used, the selector in the register remains 0.
              * We can't clear the MSR early in a new thread and then look for
@@ -1556,7 +1520,7 @@ is_thread_tls_initialized(void)
                 os_tls = (os_local_state_t *)base;
             }
         }
-#        endif
+#    endif
         if (os_tls != NULL) {
             return (os_tls->tid == get_sys_thread_id() ||
                     /* The child of a fork will initially come here */
@@ -1565,7 +1529,7 @@ is_thread_tls_initialized(void)
         } else
             return false;
     }
-#    elif defined(AARCHXX)
+#elif defined(AARCHXX)
     byte **dr_tls_base_addr;
     if (tls_global_type == TLS_TYPE_NONE)
         return false;
@@ -1581,7 +1545,7 @@ is_thread_tls_initialized(void)
      * which comes here.
      */
     return true;
-#    endif
+#endif
 }
 
 bool
@@ -1592,11 +1556,11 @@ is_DR_segment_reader_entry(app_pc pc)
      * It's a little ugly but it's not clear there's a better solution.
      * See the discussion in i#2463 c#2.
      */
-#    ifdef X86
+#ifdef X86
     if (INTERNAL_OPTION(safe_read_tls_init)) {
         return pc == (app_pc)safe_read_tls_magic || pc == (app_pc)safe_read_tls_self;
     }
-#    endif
+#endif
     /* XXX i#2463: for ARM and for -no_safe_read_tls_init it may be
      * more complicated as the PC may not be a function entry but the
      * start of a bb after a branch in our C code that uses inline asm
@@ -1605,11 +1569,11 @@ is_DR_segment_reader_entry(app_pc pc)
     return false;
 }
 
-#    if defined(X86) || defined(DEBUG)
+#if defined(X86) || defined(DEBUG)
 static bool
 is_thread_tls_allocated(void)
 {
-#        ifdef X86
+#    ifdef X86
     if (INTERNAL_OPTION(safe_read_tls_init)) {
         /* We use this routine to allow currently-native threads, for which
          * is_thread_tls_initialized() (and thus is_thread_initialized()) will
@@ -1622,10 +1586,10 @@ is_thread_tls_allocated(void)
         magic = safe_read_tls_magic();
         return magic == TLS_MAGIC_VALID || magic == TLS_MAGIC_INVALID;
     }
-#        endif
+#    endif
     return is_thread_tls_initialized();
 }
-#    endif
+#endif
 
 /* converts a local_state_t offset to a segment offset */
 ushort
@@ -1694,7 +1658,7 @@ get_os_tls_from_dc(dcontext_t *dcontext)
     return (os_local_state_t *)(local_state - offsetof(os_local_state_t, state));
 }
 
-#    ifdef AARCHXX
+#ifdef AARCHXX
 bool
 os_set_app_tls_base(dcontext_t *dcontext, reg_id_t reg, void *base)
 {
@@ -1718,7 +1682,7 @@ os_set_app_tls_base(dcontext_t *dcontext, reg_id_t reg, void *base)
     ASSERT_NOT_REACHED();
     return false;
 }
-#    endif
+#endif
 
 void *
 os_get_app_tls_base(dcontext_t *dcontext, reg_id_t reg)
@@ -1757,7 +1721,7 @@ os_get_app_tls_base_offset(reg_id_t reg)
     return 0;
 }
 
-#    ifdef X86
+#ifdef X86
 ushort
 os_get_app_tls_reg_offset(reg_id_t reg)
 {
@@ -1770,7 +1734,7 @@ os_get_app_tls_reg_offset(reg_id_t reg)
     ASSERT_NOT_REACHED();
     return 0;
 }
-#    endif
+#endif
 
 void *
 get_tls(ushort tls_offs)
@@ -1794,18 +1758,18 @@ set_tls(ushort tls_offs, void *value)
 byte *
 get_segment_base(uint seg)
 {
-#    ifdef X86
+#ifdef X86
     if (seg == SEG_CS || seg == SEG_SS || seg == SEG_DS || seg == SEG_ES)
         return NULL;
-#        ifdef HAVE_TLS
+#    ifdef HAVE_TLS
     return tls_get_fs_gs_segment_base(seg);
-#        else
+#    else
     return (byte *)POINTER_MAX;
-#        endif /* HAVE_TLS */
-#    elif defined(AARCHXX)
+#    endif /* HAVE_TLS */
+#elif defined(AARCHXX)
     /* XXX i#1551: should we rename/refactor to avoid "segment"? */
     return (byte *)read_thread_register(seg);
-#    endif
+#endif
 }
 
 /* i#572: handle opnd_compute_address to return the application
@@ -1814,10 +1778,10 @@ get_segment_base(uint seg)
 byte *
 get_app_segment_base(uint seg)
 {
-#    ifdef X86
+#ifdef X86
     if (seg == SEG_CS || seg == SEG_SS || seg == SEG_DS || seg == SEG_ES)
         return NULL;
-#    endif /* X86 */
+#endif /* X86 */
     if (IF_CLIENT_INTERFACE_ELSE(INTERNAL_OPTION(private_loader), false)) {
         return get_tls(os_get_app_tls_base_offset(seg));
     }
@@ -1836,23 +1800,23 @@ get_local_state_extended()
 local_state_t *
 get_local_state()
 {
-#    ifdef HAVE_TLS
+#ifdef HAVE_TLS
     return (local_state_t *)get_local_state_extended();
-#    else
+#else
     return NULL;
-#    endif
+#endif
 }
 
-#    ifdef DEBUG
+#ifdef DEBUG
 void
 os_enter_dynamorio(void)
 {
-#        ifdef ARM
+#    ifdef ARM
     /* i#1578: check that app's tls value doesn't match our sentinel */
     ASSERT(*(byte **)get_dr_tls_base_addr() != TLS_SLOT_VAL_EXITED);
-#        endif
-}
 #    endif
+}
+#endif
 
 /* i#107: handle segment register usage conflicts between app and dr:
  * os_handle_mov_seg updates the app's tls selector maintained by DR.
@@ -1861,7 +1825,7 @@ os_enter_dynamorio(void)
 void
 os_handle_mov_seg(dcontext_t *dcontext, byte *pc)
 {
-#    ifdef X86
+#ifdef X86
     instr_t instr;
     opnd_t opnd;
     reg_id_t seg;
@@ -1912,10 +1876,10 @@ os_handle_mov_seg(dcontext_t *dcontext, byte *pc)
         "app lib tls base: " PFX ", alt tls base: " PFX "\n",
         get_thread_id(), reg_names[seg], sel, os_tls->app_lib_tls_base,
         os_tls->app_alt_tls_base);
-#    elif defined(ARM)
+#elif defined(ARM)
     /* FIXME i#1551: NYI on ARM */
     ASSERT_NOT_REACHED();
-#    endif /* X86/ARM */
+#endif /* X86/ARM */
 }
 
 /* Initialization for TLS mangling (-mangle_app_seg on x86).
@@ -1925,13 +1889,13 @@ static void
 os_tls_app_seg_init(os_local_state_t *os_tls, void *segment)
 {
     app_pc app_lib_tls_base, app_alt_tls_base;
-#    ifdef X86
+#ifdef X86
     int i, index;
     our_modify_ldt_t *desc;
 
     os_tls->app_lib_tls_reg = read_thread_register(TLS_REG_LIB);
     os_tls->app_alt_tls_reg = read_thread_register(TLS_REG_ALT);
-#    endif
+#endif
     app_lib_tls_base = get_segment_base(TLS_REG_LIB);
     app_alt_tls_base = get_segment_base(TLS_REG_ALT);
 
@@ -1943,7 +1907,7 @@ os_tls_app_seg_init(os_local_state_t *os_tls, void *segment)
     os_tls->app_alt_tls_base =
         is_dynamo_address(app_alt_tls_base) ? NULL : app_alt_tls_base;
 
-#    ifdef X86
+#ifdef X86
     /* get all TLS thread area value */
     /* XXX: is get_thread_area supported in 64-bit kernel?
      * It has syscall number 211.
@@ -1956,7 +1920,7 @@ os_tls_app_seg_init(os_local_state_t *os_tls, void *segment)
     for (i = 0; i < GDT_NUM_TLS_SLOTS; i++) {
         tls_get_descriptor(i + index, &desc[i]);
     }
-#    endif /* X86 */
+#endif /* X86 */
 
     os_tls->os_seg_info.dr_tls_base = segment;
     os_tls->os_seg_info.priv_alt_tls_base = IF_X86_ELSE(segment, NULL);
@@ -1967,11 +1931,11 @@ os_tls_app_seg_init(os_local_state_t *os_tls, void *segment)
             os_tls->app_lib_tls_base, privload_tls_init(os_tls->app_lib_tls_base));
     }
 
-#    ifdef X86
+#ifdef X86
     LOG(THREAD_GET, LOG_THREADS, 1,
         "thread " TIDFMT " app lib tls reg: 0x%x, alt tls reg: 0x%x\n", get_thread_id(),
         os_tls->app_lib_tls_reg, os_tls->app_alt_tls_reg);
-#    endif
+#endif
     LOG(THREAD_GET, LOG_THREADS, 1,
         "thread " TIDFMT " app lib tls base: " PFX ", alt tls base: " PFX "\n",
         get_thread_id(), os_tls->app_lib_tls_base, os_tls->app_alt_tls_base);
@@ -1985,11 +1949,11 @@ os_tls_app_seg_init(os_local_state_t *os_tls, void *segment)
 void
 os_tls_init(void)
 {
-#    ifdef X86
+#ifdef X86
     ASSERT(TLS_MAGIC_OFFSET_ASM == TLS_MAGIC_OFFSET);
     ASSERT(TLS_SELF_OFFSET_ASM == TLS_SELF_OFFSET);
-#    endif
-#    ifdef HAVE_TLS
+#endif
+#ifdef HAVE_TLS
     /* We create a 1-page segment with an LDT entry for each thread and load its
      * selector into fs/gs.
      * FIXME PR 205276: this whole scheme currently does not check if app is using
@@ -2008,9 +1972,9 @@ os_tls_init(void)
     os_tls->self = os_tls;
     os_tls->tid = get_sys_thread_id();
     os_tls->tls_type = TLS_TYPE_NONE;
-#        ifdef X86
+#    ifdef X86
     os_tls->magic = TLS_MAGIC_VALID;
-#        endif
+#    endif
     /* We save DR's TLS segment base here so that os_get_dr_tls_base() will work
      * even when -no_mangle_app_seg is set.  If -mangle_app_seg is set, this
      * will be overwritten in os_tls_app_seg_init().
@@ -2032,11 +1996,11 @@ os_tls_init(void)
     /* FIXME: this should be a SYSLOG fatal error?  Should fall back on !HAVE_TLS?
      * Should have create_ldt_entry() return failure instead of asserting, then.
      */
-#    else
+#else
     tls_table = (tls_slot_t *)global_heap_alloc(MAX_THREADS *
                                                 sizeof(tls_slot_t) HEAPACCT(ACCT_OTHER));
     memset(tls_table, 0, MAX_THREADS * sizeof(tls_slot_t));
-#    endif
+#endif
     if (!first_thread_tls_initialized) {
         first_thread_tls_initialized = true;
         if (last_thread_tls_exited) /* re-attach */
@@ -2048,22 +2012,22 @@ os_tls_init(void)
 static bool
 should_zero_tls_at_thread_exit()
 {
-#    ifdef X86
+#ifdef X86
     /* i#2089: For a thread w/o CLONE_SIGHAND we cannot handle a fault, so we want to
      * leave &uninit_tls (which was put in place in os_thread_exit()) as long as
      * possible.  For non-detach, that means until the exit.
      */
     return !INTERNAL_OPTION(safe_read_tls_init) || doing_detach;
-#    else
+#else
     return true;
-#    endif
+#endif
 }
 
 /* TLS exit for the current thread who must own local_state. */
 void
 os_tls_thread_exit(local_state_t *local_state)
 {
-#    ifdef HAVE_TLS
+#ifdef HAVE_TLS
     /* We assume (assert below) that local_state_t's start == local_state_extended_t */
     os_local_state_t *os_tls =
         (os_local_state_t *)(((byte *)local_state) - offsetof(os_local_state_t, state));
@@ -2075,7 +2039,7 @@ os_tls_thread_exit(local_state_t *local_state)
     if (should_zero_tls_at_thread_exit()) {
         tls_thread_free(tls_type, index);
 
-#        if defined(X86) && defined(X64)
+#    if defined(X86) && defined(X64)
         if (tls_type == TLS_TYPE_ARCH_PRCTL) {
             /* syscall re-sets gs register so re-clear it */
             if (read_thread_register(SEG_TLS) != 0) {
@@ -2083,7 +2047,7 @@ os_tls_thread_exit(local_state_t *local_state)
                 WRITE_DR_SEG(zero); /* macro needs lvalue! */
             }
         }
-#        endif
+#    endif
     }
 
     /* We already set TLS to &uninit_tls in os_thread_exit() */
@@ -2096,7 +2060,7 @@ os_tls_thread_exit(local_state_t *local_state)
         last_thread_tls_exited = true;
         first_thread_tls_initialized = false; /* for possible re-attach */
     }
-#    endif
+#endif
 }
 
 /* Frees local_state.  If the calling thread is exiting (i.e.,
@@ -2106,15 +2070,15 @@ os_tls_thread_exit(local_state_t *local_state)
 void
 os_tls_exit(local_state_t *local_state, bool other_thread)
 {
-#    ifdef HAVE_TLS
-#        ifdef X86
+#ifdef HAVE_TLS
+#    ifdef X86
     static const ptr_uint_t zero = 0;
-#        endif /* X86 */
+#    endif /* X86 */
     /* We can't read from fs: as we can be called from other threads */
     /* ASSUMPTION: local_state_t is laid out at same start as local_state_extended_t */
     os_local_state_t *os_tls =
         (os_local_state_t *)(((byte *)local_state) - offsetof(os_local_state_t, state));
-#        ifdef X86
+#    ifdef X86
     /* If the MSR is in use, writing to the reg faults.  We rely on it being 0
      * to indicate that.
      */
@@ -2122,7 +2086,7 @@ os_tls_exit(local_state_t *local_state, bool other_thread)
         should_zero_tls_at_thread_exit()) {
         WRITE_DR_SEG(zero); /* macro needs lvalue! */
     }
-#        endif /* X86 */
+#    endif /* X86 */
 
     /* For another thread we can't really make these syscalls so we have to
      * leave it un-cleaned-up.  That's fine if the other thread is exiting:
@@ -2133,10 +2097,10 @@ os_tls_exit(local_state_t *local_state, bool other_thread)
 
     /* We can't free prior to tls_thread_free() in case that routine refs os_tls */
     heap_munmap(os_tls->self, PAGE_SIZE, VMM_SPECIAL_MMAP);
-#    else
+#else
     global_heap_free(tls_table, MAX_THREADS * sizeof(tls_slot_t) HEAPACCT(ACCT_OTHER));
     DELETE_LOCK(tls_lock);
-#    endif
+#endif
 }
 
 static int
@@ -2153,7 +2117,7 @@ os_tls_get_gdt_index(dcontext_t *dcontext)
 void
 os_tls_pre_init(int gdt_index)
 {
-#    ifdef X86
+#ifdef X86
     /* Only set to above 0 for tls_type == TLS_TYPE_GDT */
     if (gdt_index > 0) {
         /* PR 458917: clear gdt slot to avoid leak across exec */
@@ -2167,13 +2131,13 @@ os_tls_pre_init(int gdt_index)
         tls_clear_descriptor(gdt_index);
         ASSERT(ok);
     }
-#    elif defined(ARM)
+#elif defined(ARM)
     /* FIXME i#1551: NYI on ARM */
     ASSERT_NOT_IMPLEMENTED(false);
-#    endif /* X86/ARM */
+#endif /* X86/ARM */
 }
 
-#    ifdef CLIENT_INTERFACE
+#ifdef CLIENT_INTERFACE
 /* Allocates num_slots tls slots aligned with alignment align */
 bool
 os_tls_calloc(OUT uint *offset, uint num_slots, uint alignment)
@@ -2224,7 +2188,7 @@ os_tls_cfree(uint offset, uint num_slots)
     mutex_unlock(&client_tls_lock);
     return ok;
 }
-#    endif
+#endif
 
 /* os_data is a clone_record_t for signal_thread_inherit */
 void
@@ -2245,12 +2209,12 @@ os_thread_init(dcontext_t *dcontext, void *os_data)
     ksynch_init_var(&ostd->terminated);
     ksynch_init_var(&ostd->detached);
 
-#    ifdef RETURN_AFTER_CALL
+#ifdef RETURN_AFTER_CALL
     /* We only need the stack bottom for the initial thread, and due to thread
      * init now preceding vm_areas_init(), we initialize in find_executable_vm_areas()
      */
     ostd->stack_bottom_pc = NULL;
-#    endif
+#endif
 
     ASSIGN_INIT_LOCK_FREE(ostd->suspend_lock, suspend_lock);
 
@@ -2269,14 +2233,14 @@ os_thread_init(dcontext_t *dcontext, void *os_data)
     LOG(THREAD, LOG_THREADS, 1, "TLS priv alt base =" PFX "\n", ostd->priv_alt_tls_base);
     LOG(THREAD, LOG_THREADS, 1, "TLS DynamoRIO base=" PFX "\n", ostd->dr_tls_base);
 
-#    ifdef X86
+#ifdef X86
     if (INTERNAL_OPTION(mangle_app_seg)) {
         ostd->app_thread_areas = heap_alloc(
             dcontext, sizeof(our_modify_ldt_t) * GDT_NUM_TLS_SLOTS HEAPACCT(ACCT_OTHER));
         memcpy(ostd->app_thread_areas, os_tls->os_seg_info.app_thread_areas,
                sizeof(our_modify_ldt_t) * GDT_NUM_TLS_SLOTS);
     }
-#    endif
+#endif
 
     LOG(THREAD, LOG_THREADS, 1, "post-TLS-setup, cur %s base is " PFX "\n",
         IF_X86_ELSE("gs", "tpidruro"),
@@ -2285,11 +2249,11 @@ os_thread_init(dcontext_t *dcontext, void *os_data)
         IF_X86_ELSE("fs", "tpidrurw"),
         get_segment_base(IF_X86_ELSE(SEG_FS, DR_REG_TPIDRURW)));
 
-#    ifdef MACOS
+#ifdef MACOS
     /* XXX: do we need to free/close dcontext->thread_port?  I don't think so. */
     dcontext->thread_port = dynamorio_mach_syscall(MACH_thread_self_trap, 0);
     LOG(THREAD, LOG_ALL, 1, "Mach thread port: %d\n", dcontext->thread_port);
-#    endif
+#endif
 }
 
 /* os_data is a clone_record_t for signal_thread_inherit */
@@ -2324,7 +2288,7 @@ os_thread_exit(dcontext_t *dcontext, bool other_thread)
     ksynch_free_var(&ostd->terminated);
     ksynch_free_var(&ostd->detached);
 
-#    ifdef X86
+#ifdef X86
     if (ostd->clone_tls != NULL) {
         if (!other_thread) {
             /* Avoid faults in is_thread_tls_initialized() */
@@ -2336,19 +2300,19 @@ os_thread_exit(dcontext_t *dcontext, bool other_thread)
                            UNPROTECTED);
         });
     }
-#    endif
+#endif
 
     /* for non-debug we do fast exit path and don't free local heap */
     DODEBUG({
         if (MACHINE_TLS_IS_DR_TLS) {
-#    ifdef X86
+#ifdef X86
             heap_free(dcontext, ostd->app_thread_areas,
                       sizeof(our_modify_ldt_t) * GDT_NUM_TLS_SLOTS HEAPACCT(ACCT_OTHER));
-#    endif
-#    ifdef CLIENT_INTERFACE
+#endif
+#ifdef CLIENT_INTERFACE
             if (INTERNAL_OPTION(private_loader))
                 privload_tls_exit(IF_UNIT_TEST_ELSE(NULL, ostd->priv_lib_tls_base));
-#    endif
+#endif
         }
         heap_free(dcontext, ostd, sizeof(os_thread_data_t) HEAPACCT(ACCT_OTHER));
     });
@@ -2465,7 +2429,7 @@ os_fork_init(dcontext_t *dcontext)
 static void
 os_swap_dr_tls(dcontext_t *dcontext, bool to_app)
 {
-#    ifdef X86
+#ifdef X86
     /* If the option is off, we really should swap it (xref i#107/i#2088 comments
      * in os_swap_context()) but there are few consequences of not doing it, and we
      * have no code set up separate from the i#2089 scheme here.
@@ -2516,7 +2480,7 @@ os_swap_dr_tls(dcontext_t *dcontext, bool to_app)
             os_set_dr_tls_base(dcontext, real_tls, (byte *)real_tls);
         }
     }
-#    endif
+#endif
 }
 
 static void
@@ -2563,24 +2527,24 @@ os_get_dr_tls_base(dcontext_t *dcontext)
 bool
 os_should_swap_state(void)
 {
-#    ifdef X86
+#ifdef X86
     /* -private_loader currently implies -mangle_app_seg, but let's be safe. */
     return (INTERNAL_OPTION(mangle_app_seg) &&
             IF_CLIENT_INTERFACE_ELSE(INTERNAL_OPTION(private_loader), false));
-#    elif defined(AARCHXX)
+#elif defined(AARCHXX)
     /* FIXME i#1582: this should return true, but there is a lot of complexity
      * getting os_switch_seg_to_context() to do the right then when called
      * at main thread init, secondary thread init, early and late injection,
      * and thread exit, since it is fragile with its writes to app TLS.
      */
     return false;
-#    endif
+#endif
 }
 
 bool
 os_using_app_state(dcontext_t *dcontext)
 {
-#    ifdef X86
+#ifdef X86
     /* FIXME: This could be optimized to avoid the syscall by keeping state in
      * the dcontext.
      */
@@ -2588,7 +2552,7 @@ os_using_app_state(dcontext_t *dcontext)
         return (get_segment_base(TLS_REG_LIB) ==
                 os_get_app_tls_base(dcontext, TLS_REG_LIB));
     }
-#    endif
+#endif
     /* We're always in the app state if we're not mangling. */
     return true;
 }
@@ -2611,15 +2575,15 @@ os_swap_context(dcontext_t *dcontext, bool to_app, dr_state_flags_t flags)
 void
 os_swap_context_go_native(dcontext_t *dcontext, dr_state_flags_t flags)
 {
-#    ifdef AARCHXX
+#ifdef AARCHXX
     /* FIXME i#1582: remove this routine once os_should_swap_state()
      * is not disabled and we can actually call
      * os_swap_context_go_native() safely from multiple places.
      */
     os_switch_seg_to_context(dcontext, LIB_SEG_TLS, true /*to app*/);
-#    else
+#else
     os_swap_context(dcontext, true /*to app*/, flags);
-#    endif
+#endif
 }
 
 void
@@ -2704,14 +2668,6 @@ get_process_group_id()
     return dynamorio_syscall(SYS_getpgid, 0);
 }
 
-#endif /* !NOT_DYNAMORIO_CORE_PROPER: around most of file, to exclude preload */
-process_id_t
-get_process_id()
-{
-    return dynamorio_syscall(SYS_getpid, 0);
-}
-#ifndef NOT_DYNAMORIO_CORE_PROPER /* around most of file, to exclude preload */
-
 process_id_t
 get_parent_id(void)
 {
@@ -2721,13 +2677,13 @@ get_parent_id(void)
 thread_id_t
 get_sys_thread_id(void)
 {
-#    ifdef MACOS
+#ifdef MACOS
     if (kernel_thread_groups)
         return dynamorio_syscall(SYS_thread_selfid, 0);
-#    else
+#else
     if (kernel_thread_groups)
         return dynamorio_syscall(SYS_gettid, 0);
-#    endif
+#endif
     return dynamorio_syscall(SYS_getpid, 0);
 }
 
@@ -2763,7 +2719,7 @@ get_tls_thread_id(void)
 dcontext_t *
 get_thread_private_dcontext(void)
 {
-#    ifdef HAVE_TLS
+#ifdef HAVE_TLS
     dcontext_t *dcontext;
     /* We have to check this b/c this is called from __errno_location prior
      * to os_tls_init, as well as after os_tls_exit, and early in a new
@@ -2805,7 +2761,7 @@ get_thread_private_dcontext(void)
     });
     READ_TLS_SLOT_IMM(TLS_DCONTEXT_OFFSET, dcontext);
     return dcontext;
-#    else
+#else
     /* Assumption: no lock needed on a read => no race conditions between
      * reading and writing same tid!  Since both get and set are only for
      * the current thread, they cannot both execute simultaneously for the
@@ -2821,17 +2777,17 @@ get_thread_private_dcontext(void)
         }
     }
     return NULL;
-#    endif
+#endif
 }
 
 /* sets the thread-private dcontext pointer for the calling thread */
 void
 set_thread_private_dcontext(dcontext_t *dcontext)
 {
-#    ifdef HAVE_TLS
+#ifdef HAVE_TLS
     ASSERT(is_thread_tls_allocated());
     WRITE_TLS_SLOT_IMM(TLS_DCONTEXT_OFFSET, dcontext);
-#    else
+#else
     thread_id_t tid = get_thread_id();
     int i;
     bool found = false;
@@ -2865,7 +2821,7 @@ set_thread_private_dcontext(dcontext_t *dcontext)
     }
     mutex_unlock(&tls_lock);
     ASSERT(found);
-#    endif
+#endif
 }
 
 /* replaces old with new
@@ -2874,7 +2830,7 @@ set_thread_private_dcontext(dcontext_t *dcontext)
 static void
 replace_thread_id(thread_id_t old, thread_id_t new)
 {
-#    ifdef HAVE_TLS
+#ifdef HAVE_TLS
     thread_id_t new_tid = new;
     ASSERT(is_thread_tls_initialized());
     DOCHECK(1, {
@@ -2883,7 +2839,7 @@ replace_thread_id(thread_id_t old, thread_id_t new)
         ASSERT(old_tid == old);
     });
     WRITE_TLS_INT_SLOT_IMM(TLS_THREAD_ID_OFFSET, new_tid);
-#    else
+#else
     int i;
     mutex_lock(&tls_lock);
     for (i = 0; i < MAX_THREADS; i++) {
@@ -2893,42 +2849,9 @@ replace_thread_id(thread_id_t old, thread_id_t new)
         }
     }
     mutex_unlock(&tls_lock);
-#    endif
+#endif
 }
 
-#endif /* !NOT_DYNAMORIO_CORE_PROPER */
-
-/* translate permission string to platform independent protection bits */
-uint
-permstr_to_memprot(const char *const perm)
-{
-    uint mem_prot = 0;
-    if (perm == NULL || *perm == '\0')
-        return mem_prot;
-    if (perm[2] == 'x')
-        mem_prot |= MEMPROT_EXEC;
-    if (perm[1] == 'w')
-        mem_prot |= MEMPROT_WRITE;
-    if (perm[0] == 'r')
-        mem_prot |= MEMPROT_READ;
-    return mem_prot;
-}
-
-/* translate platform independent protection bits to native flags */
-uint
-memprot_to_osprot(uint prot)
-{
-    uint mmap_prot = 0;
-    if (TEST(MEMPROT_EXEC, prot))
-        mmap_prot |= PROT_EXEC;
-    if (TEST(MEMPROT_READ, prot))
-        mmap_prot |= PROT_READ;
-    if (TEST(MEMPROT_WRITE, prot))
-        mmap_prot |= PROT_WRITE;
-    return mmap_prot;
-}
-
-#ifndef NOT_DYNAMORIO_CORE_PROPER
 /* translate native flags to platform independent protection bits */
 static inline uint
 osprot_to_memprot(uint prot)
@@ -2942,7 +2865,6 @@ osprot_to_memprot(uint prot)
         mem_prot |= MEMPROT_WRITE;
     return mem_prot;
 }
-#endif
 
 /* returns osprot flags preserving all native protection flags except
  * for RWX, which are replaced according to memprot */
@@ -2963,46 +2885,6 @@ mprotect_syscall(byte *p, size_t size, uint prot)
     return dynamorio_syscall(SYS_mprotect, 3, p, size, prot);
 }
 
-bool
-mmap_syscall_succeeded(byte *retval)
-{
-    ptr_int_t result = (ptr_int_t)retval;
-    /* libc interprets up to -PAGE_SIZE as an error, and you never know if
-     * some weird errno will be used by say vmkernel (xref PR 365331)
-     */
-    bool fail = (result < 0 && result >= -PAGE_SIZE);
-    ASSERT_CURIOSITY(!fail ||
-                     IF_VMX86(result == -ENOENT ||) IF_VMX86(result == -ENOSPC ||)
-                             result == -EBADF ||
-                     result == -EACCES || result == -EINVAL || result == -ETXTBSY ||
-                     result == -EAGAIN || result == -ENOMEM || result == -ENODEV ||
-                     result == -EFAULT || result == -EPERM);
-    return !fail;
-}
-
-/* N.B.: offs should be in pages for 32-bit Linux */
-static inline byte *
-mmap_syscall(byte *addr, size_t len, ulong prot, ulong flags, ulong fd, ulong offs)
-{
-#if defined(MACOS) && !defined(X64)
-    return (byte *)(ptr_int_t)dynamorio_syscall(
-        SYS_mmap, 7, addr, len, prot, flags, fd,
-        /* represent 64-bit arg as 2 32-bit args */
-        offs, 0);
-#else
-    return (byte *)(ptr_int_t)dynamorio_syscall(
-        IF_MACOS_ELSE(SYS_mmap, IF_X64_ELSE(SYS_mmap, SYS_mmap2)), 6, addr, len, prot,
-        flags, fd, offs);
-#endif
-}
-
-static inline long
-munmap_syscall(byte *addr, size_t len)
-{
-    return dynamorio_syscall(SYS_munmap, 2, addr, len);
-}
-
-#ifndef NOT_DYNAMORIO_CORE_PROPER
 /* free memory allocated from os_raw_mem_alloc */
 bool
 os_raw_mem_free(void *p, size_t size, uint flags, heap_error_code_t *error_code)
@@ -3052,7 +2934,7 @@ os_raw_mem_alloc(void *preferred, size_t size, uint prot, uint flags,
     return p;
 }
 
-#    ifdef LINUX
+#ifdef LINUX
 void
 init_emulated_brk(app_pc exe_end)
 {
@@ -3109,9 +2991,9 @@ emulate_app_brk(dcontext_t *dcontext, byte *new_val)
         handle_app_brk(dcontext, app_brk_map, old_brk, app_brk_cur);
     return app_brk_cur;
 }
-#    endif /* LINUX */
+#endif /* LINUX */
 
-#    if defined(CLIENT_INTERFACE) && defined(LINUX)
+#if defined(CLIENT_INTERFACE) && defined(LINUX)
 DR_API
 /* XXX: could add dr_raw_mem_realloc() instead of dr_raw_mremap() -- though there
  * is no realloc for Windows: supposed to reserve yourself and then commit in
@@ -3164,7 +3046,7 @@ dr_raw_brk(void *new_address)
         }
     }
 }
-#    endif /* CLIENT_INTERFACE && LINUX */
+#endif /* CLIENT_INTERFACE && LINUX */
 
 /* caller is required to handle thread synchronization and to update dynamo vm areas */
 void
@@ -3194,7 +3076,7 @@ os_heap_reserve(void *preferred, size_t size, heap_error_code_t *error_code,
 {
     void *p;
     uint prot = PROT_NONE;
-#    ifdef VMX86_SERVER
+#ifdef VMX86_SERVER
     /* PR 365331: we need to be in the mmap_text region for code cache and
      * gencode (PROT_EXEC).
      */
@@ -3213,7 +3095,7 @@ os_heap_reserve(void *preferred, size_t size, heap_error_code_t *error_code,
      */
     if (executable)
         prot = PROT_EXEC;
-#    endif
+#endif
     /* should only be used on aligned pieces */
     ASSERT(size > 0 && ALIGNED(size, PAGE_SIZE));
     ASSERT(error_code != NULL);
@@ -3245,17 +3127,17 @@ os_heap_reserve(void *preferred, size_t size, heap_error_code_t *error_code,
         *error_code = HEAP_ERROR_SUCCESS;
     }
     LOG(GLOBAL, LOG_HEAP, 2, "os_heap_reserve: %d bytes @ " PFX "\n", size, p);
-#    ifdef VMX86_SERVER
+#ifdef VMX86_SERVER
     /* PR 365331: ensure our memory is all in the mmap_text region */
     ASSERT(!os_in_vmkernel_userworld() || !executable ||
            ((byte *)p >= os_vmk_mmap_text_start() &&
             ((byte *)p) + size <= os_vmk_mmap_text_end()));
-#    endif
-#    if defined(ANDROID) && defined(DEBUG)
+#endif
+#if defined(ANDROID) && defined(DEBUG)
     /* We don't label in release to be more transparent */
     dynamorio_syscall(SYS_prctl, 5, PR_SET_VMA, PR_SET_VMA_ANON_NAME, p, size,
                       "DynamoRIO-internal");
-#    endif
+#endif
     return p;
 }
 
@@ -3308,7 +3190,7 @@ os_heap_reserve_in_region(void *start, void *end, size_t size,
         return os_heap_reserve(NULL, size, error_code, executable);
 
         /* loop to handle races */
-#    define RESERVE_IN_REGION_MAX_ITERS 128
+#define RESERVE_IN_REGION_MAX_ITERS 128
     while (find_free_memory_in_region(start, end, size, &try_start, &try_end)) {
         /* If there's space we'd prefer the end, to avoid the common case of
          * a large binary + heap at attach where we're likely to reserve
@@ -3406,24 +3288,24 @@ os_heap_get_commit_limit(size_t *commit_used, size_t *commit_limit)
 void
 os_thread_yield()
 {
-#    ifdef MACOS
+#ifdef MACOS
     /* XXX i#1291: use raw syscall instead */
     swtch_pri(0);
-#    else
+#else
     dynamorio_syscall(SYS_sched_yield, 0);
-#    endif
+#endif
 }
 
 bool
 thread_signal(process_id_t pid, thread_id_t tid, int signum)
 {
-#    ifdef MACOS
+#ifdef MACOS
     /* FIXME i#58: this takes in a thread port.  Need to map thread id to port.
      * Need to figure out whether we support raw Mach threads w/o pthread on top.
      */
     ASSERT_NOT_IMPLEMENTED(false);
     return false;
-#    else
+#else
     /* FIXME: for non-NPTL use SYS_kill */
     /* Note that the pid is equivalent to the thread group id.
      * However, we can have threads sharing address space but not pid
@@ -3431,13 +3313,13 @@ thread_signal(process_id_t pid, thread_id_t tid, int signum)
      * use the pid of the target thread, not our pid.
      */
     return (dynamorio_syscall(SYS_tgkill, 3, pid, tid, signum) == 0);
-#    endif
+#endif
 }
 
 static bool
 known_thread_signal(thread_record_t *tr, int signum)
 {
-#    ifdef MACOS
+#ifdef MACOS
     ptr_int_t res;
     if (tr->dcontext == NULL)
         return FALSE;
@@ -3445,26 +3327,26 @@ known_thread_signal(thread_record_t *tr, int signum)
     LOG(THREAD_GET, LOG_ALL, 3, "%s: signal %d to port %d => %ld\n", __FUNCTION__, signum,
         tr->dcontext->thread_port, res);
     return res == 0;
-#    else
+#else
     return thread_signal(tr->pid, tr->id, signum);
-#    endif
+#endif
 }
 
 void
 os_thread_sleep(uint64 milliseconds)
 {
-#    ifdef MACOS
+#ifdef MACOS
     semaphore_t sem = MACH_PORT_NULL;
     int res;
-#    else
+#else
     struct timespec remain;
     int count = 0;
-#    endif
+#endif
     struct timespec req;
     req.tv_sec = (milliseconds / 1000);
     /* docs say can go up to 1000000000, but doesn't work on FC9 */
     req.tv_nsec = (milliseconds % 1000) * 1000000;
-#    ifdef MACOS
+#ifdef MACOS
     if (sem == MACH_PORT_NULL) {
         DEBUG_DECLARE(kern_return_t res =)
         semaphore_create(mach_task_self(), &sem, SYNC_POLICY_FIFO, 0);
@@ -3476,7 +3358,7 @@ os_thread_sleep(uint64 milliseconds)
     if (res == -EINTR) {
         /* FIXME i#58: figure out how much time elapsed and re-wait */
     }
-#    else
+#else
     /* FIXME: if we need accurate sleeps in presence of itimers we should
      * be using SYS_clock_nanosleep w/ an absolute time instead of relative
      */
@@ -3498,7 +3380,7 @@ os_thread_sleep(uint64 milliseconds)
         }
         req = remain;
     }
-#    endif
+#endif
 }
 
 bool
@@ -3544,7 +3426,7 @@ os_thread_suspend(thread_record_t *tr)
          * assumes there will not be a retry until the target reaches the suspend
          * point.  Xref i#2779.
          */
-#    define SUSPEND_DEBUG_TIMEOUT_MS 5000
+#define SUSPEND_DEBUG_TIMEOUT_MS 5000
         if (ksynch_wait(&ostd->suspended, 0, SUSPEND_DEBUG_TIMEOUT_MS) == -ETIMEDOUT) {
             ASSERT_CURIOSITY(false && "failed to suspend thread in 5s");
         }
@@ -3727,8 +3609,8 @@ is_thread_currently_native(thread_record_t *tr)
             (tr->dcontext != NULL && tr->dcontext->currently_stopped));
 }
 
-#    ifdef CLIENT_SIDELINE /* PR 222812: tied to sideline usage */
-#        ifdef LINUX       /* XXX i#58: just until we have Mac support */
+#ifdef CLIENT_SIDELINE /* PR 222812: tied to sideline usage */
+#    ifdef LINUX       /* XXX i#58: just until we have Mac support */
 static void
 client_thread_run(void)
 {
@@ -3764,7 +3646,7 @@ client_thread_run(void)
     block_cleanup_and_terminate(dcontext, SYS_exit, 0, 0, false /*just thread*/,
                                 IF_MACOS_ELSE(dcontext->thread_port, 0), 0);
 }
-#        endif
+#    endif
 
 /* i#41/PR 222812: client threads
  * * thread must have dcontext since many API routines require one and we
@@ -3780,7 +3662,7 @@ client_thread_run(void)
 DR_API bool
 dr_create_client_thread(void (*func)(void *param), void *arg)
 {
-#        ifdef LINUX
+#    ifdef LINUX
     dcontext_t *dcontext = get_thread_private_dcontext();
     byte *xsp;
     /* We do not pass SIGCHLD since don't want signal to parent and don't support
@@ -3806,7 +3688,7 @@ dr_create_client_thread(void (*func)(void *param), void *arg)
     /* i#501 switch to app's tls before creating client thread */
     if (IF_CLIENT_INTERFACE_ELSE(INTERNAL_OPTION(private_loader), false))
         os_switch_lib_tls(dcontext, true /*to app*/);
-#            if defined(X86) && !defined(X64)
+#        if defined(X86) && !defined(X64)
     /* For the TCB we simply share the parent's.  On Linux we could just inherit
      * the same selector but not for VMX86_SERVER so we specify for both for
      * 32-bit.  Most of the fields are pthreads-specific and we assume the ones
@@ -3821,7 +3703,7 @@ dr_create_client_thread(void (*func)(void *param), void *arg)
             __FUNCTION__, index);
         return false;
     }
-#            endif
+#        endif
     LOG(THREAD, LOG_ALL, 1, "dr_create_client_thread xsp=" PFX " dstack=" PFX "\n", xsp,
         get_clone_record_dstack(crec));
     thread_id_t newpid =
@@ -3839,23 +3721,23 @@ dr_create_client_thread(void (*func)(void *param), void *arg)
         return false;
     }
     return true;
-#        else
+#    else
     ASSERT_NOT_IMPLEMENTED(false); /* FIXME i#58: implement on Mac */
     return false;
-#        endif
+#    endif
 }
-#    endif /* CLIENT_SIDELINE PR 222812: tied to sideline usage */
+#endif /* CLIENT_SIDELINE PR 222812: tied to sideline usage */
 
 int
 get_num_processors(void)
 {
     static uint num_cpu = 0; /* cached value */
     if (!num_cpu) {
-#    ifdef MACOS
+#ifdef MACOS
         DEBUG_DECLARE(bool ok =)
         sysctl_query(CTL_HW, HW_NCPU, &num_cpu, sizeof(num_cpu));
         ASSERT(ok);
-#    else
+#else
         /* We used to use get_nprocs_conf, but that's in libc, so now we just
          * look at the /sys filesystem ourselves, which is what glibc does.
          */
@@ -3872,7 +3754,7 @@ get_num_processors(void)
         }
         os_close(cpu_dir);
         num_cpu = local_num_cpus;
-#    endif
+#endif
         ASSERT(num_cpu);
     }
     return num_cpu;
@@ -3884,51 +3766,51 @@ get_num_processors(void)
  * with -no_private_loader, so this should never happen.
  */
 
-#    if defined(CLIENT_INTERFACE) || defined(HOT_PATCHING_INTERFACE)
+#if defined(CLIENT_INTERFACE) || defined(HOT_PATCHING_INTERFACE)
 shlib_handle_t
 load_shared_library(const char *name, bool reachable)
 {
-#        ifdef STATIC_LIBRARY
+#    ifdef STATIC_LIBRARY
     if (os_files_same(name, get_application_name())) {
         /* The private loader falls back to dlsym() and friends for modules it
          * doesn't recognize, so this works without disabling the private loader.
          */
         return dlopen(NULL, RTLD_LAZY); /* Gets a handle to the exe. */
     }
-#        endif
+#    endif
     /* We call locate_and_load_private_library() to support searching for
      * a pathless name.
      */
     if (IF_CLIENT_INTERFACE_ELSE(INTERNAL_OPTION(private_loader), false))
         return (shlib_handle_t)locate_and_load_private_library(name, reachable);
-#        if defined(STATIC_LIBRARY) || defined(MACOS)
+#    if defined(STATIC_LIBRARY) || defined(MACOS)
     ASSERT(!DYNAMO_OPTION(early_inject));
     return dlopen(name, RTLD_LAZY);
-#        else
+#    else
     /* -no_private_loader is no longer supported in our default builds.
      * If we want it for hybrid mode we should add a new build param and include
      * the libdl calls here under that param.
      */
     ASSERT_NOT_REACHED();
     return NULL;
-#        endif
-}
 #    endif
+}
+#endif
 
-#    if defined(CLIENT_INTERFACE)
+#if defined(CLIENT_INTERFACE)
 shlib_routine_ptr_t
 lookup_library_routine(shlib_handle_t lib, const char *name)
 {
     if (IF_CLIENT_INTERFACE_ELSE(INTERNAL_OPTION(private_loader), false)) {
         return (shlib_routine_ptr_t)get_private_library_address((app_pc)lib, name);
     }
-#        if defined(STATIC_LIBRARY) || defined(MACOS)
+#    if defined(STATIC_LIBRARY) || defined(MACOS)
     ASSERT(!DYNAMO_OPTION(early_inject));
     return dlsym(lib, name);
-#        else
+#    else
     ASSERT_NOT_REACHED(); /* -no_private_loader is no longer supported: see above */
     return NULL;
-#        endif
+#    endif
 }
 
 void
@@ -3937,14 +3819,14 @@ unload_shared_library(shlib_handle_t lib)
     if (IF_CLIENT_INTERFACE_ELSE(INTERNAL_OPTION(private_loader), false)) {
         unload_private_library(lib);
     } else {
-#        if defined(STATIC_LIBRARY) || defined(MACOS)
+#    if defined(STATIC_LIBRARY) || defined(MACOS)
         ASSERT(!DYNAMO_OPTION(early_inject));
         if (!DYNAMO_OPTION(avoid_dlclose)) {
             dlclose(lib);
         }
-#        else
+#    else
         ASSERT_NOT_REACHED(); /* -no_private_loader is no longer supported: see above  */
-#        endif
+#    endif
     }
 }
 
@@ -3955,16 +3837,16 @@ shared_library_error(char *buf, int maxlen)
     if (IF_CLIENT_INTERFACE_ELSE(INTERNAL_OPTION(private_loader), false)) {
         err = "error in private loader";
     } else {
-#        if defined(STATIC_LIBRARY) || defined(MACOS)
+#    if defined(STATIC_LIBRARY) || defined(MACOS)
         ASSERT(!DYNAMO_OPTION(early_inject));
         err = dlerror();
         if (err == NULL) {
             err = "dlerror returned NULL";
         }
-#        else
+#    else
         ASSERT_NOT_REACHED(); /* -no_private_loader is no longer supported */
         err = "unknown error";
-#        endif
+#    endif
     }
     strncpy(buf, err, maxlen - 1);
     buf[maxlen - 1] = '\0'; /* strncpy won't put on trailing null if maxes out */
@@ -4001,237 +3883,14 @@ shared_library_bounds(IN shlib_handle_t lib, IN byte *addr, IN const char *name,
     }
     return (memquery_library_bounds(name, start, end, NULL, 0, NULL, 0) > 0);
 }
-#    endif /* defined(CLIENT_INTERFACE) */
+#endif /* defined(CLIENT_INTERFACE) */
 
-#endif /* !NOT_DYNAMORIO_CORE_PROPER: around most of file, to exclude preload */
-
-/* FIXME - not available in 2.0 or earlier kernels, not really an issue since no one
- * should be running anything that old. */
-int
-llseek_syscall(int fd, int64 offset, int origin, int64 *result)
-{
-#if defined(X64) || defined(MACOS)
-#    ifndef X64
-    /* 2 slots for 64-bit arg */
-    *result = dynamorio_syscall(SYS_lseek, 4, fd, (uint)(offset & 0xFFFFFFFF),
-                                (uint)((offset >> 32) & 0xFFFFFFFF), origin);
-#    else
-    *result = dynamorio_syscall(SYS_lseek, 3, fd, offset, origin);
-#    endif
-    return ((*result > 0) ? 0 : (int)*result);
-#else
-    return dynamorio_syscall(SYS__llseek, 5, fd, (uint)((offset >> 32) & 0xFFFFFFFF),
-                             (uint)(offset & 0xFFFFFFFF), result, origin);
-#endif
-}
-
-static ptr_int_t
-dynamorio_syscall_stat(const char *fname, struct stat64 *st)
-{
-#ifdef SYSNUM_STAT
-    return dynamorio_syscall(SYSNUM_STAT, 2, fname, st);
-#else
-    return dynamorio_syscall(SYS_fstatat, 4, AT_FDCWD, fname, st, 0);
-#endif
-}
-
-bool
-os_file_exists(const char *fname, bool is_dir)
-{
-    /* _LARGEFILE64_SOURCE should make libc struct match kernel (see top of file) */
-    struct stat64 st;
-    ptr_int_t res = dynamorio_syscall_stat(fname, &st);
-    if (res != 0) {
-        LOG(THREAD_GET, LOG_SYSCALLS, 2, "%s failed: " PIFX "\n", __func__, res);
-        return false;
-    }
-    return (!is_dir || S_ISDIR(st.st_mode));
-}
-
-/* Returns true if two paths point to the same file.  Follows symlinks.
- */
-bool
-os_files_same(const char *path1, const char *path2)
-{
-    struct stat64 st1, st2;
-    ptr_int_t res = dynamorio_syscall_stat(path1, &st1);
-    if (res != 0) {
-        LOG(THREAD_GET, LOG_SYSCALLS, 2, "%s failed: " PIFX "\n", __func__, res);
-        return false;
-    }
-    res = dynamorio_syscall_stat(path2, &st2);
-    if (res != 0) {
-        LOG(THREAD_GET, LOG_SYSCALLS, 2, "%s failed: " PIFX "\n", __func__, res);
-        return false;
-    }
-    return st1.st_ino == st2.st_ino;
-}
-
-bool
-os_get_file_size(const char *file, uint64 *size)
-{
-    /* _LARGEFILE64_SOURCE should make libc struct match kernel (see top of file) */
-    struct stat64 st;
-    ptr_int_t res = dynamorio_syscall_stat(file, &st);
-    if (res != 0) {
-        LOG(THREAD_GET, LOG_SYSCALLS, 2, "%s failed: " PIFX "\n", __func__, res);
-        return false;
-    }
-    ASSERT(size != NULL);
-    *size = st.st_size;
-    return true;
-}
-
-bool
-os_get_file_size_by_handle(file_t fd, uint64 *size)
-{
-    /* _LARGEFILE64_SOURCE should make libc struct match kernel (see top of file) */
-    struct stat64 st;
-    ptr_int_t res = dynamorio_syscall(SYSNUM_FSTAT, 2, fd, &st);
-    if (res != 0) {
-        LOG(THREAD_GET, LOG_SYSCALLS, 2, "%s failed: " PIFX "\n", __func__, res);
-        return false;
-    }
-    ASSERT(size != NULL);
-    *size = st.st_size;
-    return true;
-}
-
-/* created directory will be owned by effective uid,
- * Note a symbolic link will never be followed.
- */
-bool
-os_create_dir(const char *fname, create_directory_flags_t create_dir_flags)
-{
-    bool require_new = TEST(CREATE_DIR_REQUIRE_NEW, create_dir_flags);
-#ifdef SYS_mkdir
-    int rc = dynamorio_syscall(SYS_mkdir, 2, fname, S_IRWXU | S_IRWXG);
-#else
-    int rc = dynamorio_syscall(SYS_mkdirat, 3, AT_FDCWD, fname, S_IRWXU | S_IRWXG);
-#endif
-    ASSERT(create_dir_flags == CREATE_DIR_REQUIRE_NEW ||
-           create_dir_flags == CREATE_DIR_ALLOW_EXISTING);
-    return (rc == 0 || (!require_new && rc == -EEXIST));
-}
-
-bool
-os_delete_dir(const char *name)
-{
-#ifdef SYS_rmdir
-    return (dynamorio_syscall(SYS_rmdir, 1, name) == 0);
-#else
-    return (dynamorio_syscall(SYS_unlinkat, 3, AT_FDCWD, name, AT_REMOVEDIR) == 0);
-#endif
-}
-
-int
-open_syscall(const char *file, int flags, int mode)
-{
-    ASSERT(file != NULL);
-#ifdef SYS_open
-    return dynamorio_syscall(SYSNUM_NO_CANCEL(SYS_open), 3, file, flags, mode);
-#else
-    return dynamorio_syscall(SYSNUM_NO_CANCEL(SYS_openat), 4, AT_FDCWD, file, flags,
-                             mode);
-#endif
-}
-
-int
-close_syscall(int fd)
-{
-    return dynamorio_syscall(SYSNUM_NO_CANCEL(SYS_close), 1, fd);
-}
-
-int
-dup_syscall(int fd)
-{
-    return dynamorio_syscall(SYS_dup, 1, fd);
-}
-
-ssize_t
-read_syscall(int fd, void *buf, size_t nbytes)
-{
-    return dynamorio_syscall(SYSNUM_NO_CANCEL(SYS_read), 3, fd, buf, nbytes);
-}
-
-ssize_t
-write_syscall(int fd, const void *buf, size_t nbytes)
-{
-    return dynamorio_syscall(SYSNUM_NO_CANCEL(SYS_write), 3, fd, buf, nbytes);
-}
-
-#ifndef NOT_DYNAMORIO_CORE_PROPER
 static int
 fcntl_syscall(int fd, int cmd, long arg)
 {
     return dynamorio_syscall(SYSNUM_NO_CANCEL(SYS_fcntl), 3, fd, cmd, arg);
 }
-#endif /* !NOT_DYNAMORIO_CORE_PROPER */
 
-/* not easily accessible in header files */
-#ifndef O_LARGEFILE
-#    ifdef X64
-/* not needed */
-#        define O_LARGEFILE 0
-#    else
-#        define O_LARGEFILE 0100000
-#    endif
-#endif
-
-/* we assume that opening for writing wants to create file.
- * we also assume that nobody calling this is creating a persistent
- * file: for that, use os_open_protected() to avoid leaking on exec
- * and to separate from the app's files.
- */
-file_t
-os_open(const char *fname, int os_open_flags)
-{
-    int res;
-    int flags = 0;
-    if (TEST(OS_OPEN_ALLOW_LARGE, os_open_flags))
-        flags |= O_LARGEFILE;
-    if (TEST(OS_OPEN_WRITE_ONLY, os_open_flags))
-        res = open_syscall(fname, flags | O_WRONLY, 0);
-    else if (!TEST(OS_OPEN_WRITE, os_open_flags))
-        res = open_syscall(fname, flags | O_RDONLY, 0);
-    else {
-        res = open_syscall(
-            fname,
-            flags | O_RDWR | O_CREAT |
-                (TEST(OS_OPEN_APPEND, os_open_flags)
-                     ?
-                     /* Currently we only support either appending
-                      * or truncating, just like Windows and the client
-                      * interface.  If we end up w/ a use case that wants
-                      * neither it could open append and then seek; if we do
-                      * add OS_TRUNCATE or sthg we'll need to add it to
-                      * any current writers who don't set OS_OPEN_REQUIRE_NEW.
-                      */
-                     O_APPEND
-                     : O_TRUNC) |
-                (TEST(OS_OPEN_REQUIRE_NEW, os_open_flags) ? O_EXCL : 0),
-            S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP);
-    }
-    if (res < 0)
-        return INVALID_FILE;
-
-    return res;
-}
-
-file_t
-os_open_directory(const char *fname, int os_open_flags)
-{
-    /* no special handling */
-    return os_open(fname, os_open_flags);
-}
-
-void
-os_close(file_t f)
-{
-    close_syscall(f);
-}
-
-#ifndef NOT_DYNAMORIO_CORE_PROPER
 /* dups curfd to a private fd.
  * returns -1 if unsuccessful.
  */
@@ -4290,11 +3949,11 @@ fd_table_add(file_t fd, uint flags)
                          (void *)(ptr_uint_t)(flags | OS_OPEN_RESERVED));
         TABLE_RWLOCK(fd_table, write, unlock);
     } else {
-#    ifdef DEBUG
+#ifdef DEBUG
         num_fd_add_pre_heap++;
         /* we add main_logfile in os_init() */
         ASSERT(num_fd_add_pre_heap == 1 && "only main_logfile should come here");
-#    endif
+#endif
     }
 }
 
@@ -4357,7 +4016,7 @@ os_close_protected(file_t f)
 bool
 os_get_current_dir(char *buf, size_t bufsz)
 {
-#    ifdef MACOS
+#ifdef MACOS
     static char noheap_buf[MAXPATHLEN];
     bool res = false;
     file_t fd = os_open(".", OS_OPEN_READ);
@@ -4380,110 +4039,27 @@ cwd_error:
         global_heap_free(fcntl_buf, MAXPATHLEN HEAPACCT(ACCT_OTHER));
     os_close(fd);
     return res;
-#    else
+#else
     return (dynamorio_syscall(SYS_getcwd, 2, buf, bufsz) > 0);
-#    endif
+#endif
 }
-#endif /* !NOT_DYNAMORIO_CORE_PROPER */
 
-#ifndef NOT_DYNAMORIO_CORE_PROPER /* so drinject can use drdecode's copy */
 ssize_t
 os_write(file_t f, const void *buf, size_t count)
 {
     return write_syscall(f, buf, count);
 }
-#endif /* !NOT_DYNAMORIO_CORE_PROPER */
 
-ssize_t
-os_read(file_t f, void *buf, size_t count)
-{
-    return read_syscall(f, buf, count);
-}
-
-void
-os_flush(file_t f)
-{
-    /* we're not using FILE*, so there is no buffering */
-}
-
-/* seek current file position to offset bytes from origin, return true if successful */
-bool
-os_seek(file_t f, int64 offset, int origin)
-{
-    int64 result;
-    int ret = 0;
-
-    ret = llseek_syscall(f, offset, origin, &result);
-
-    return (ret == 0);
-}
-
-/* return the current file position, -1 on failure */
-int64
-os_tell(file_t f)
-{
-    int64 result = -1;
-    int ret = 0;
-
-    ret = llseek_syscall(f, 0, SEEK_CUR, &result);
-
-    if (ret != 0)
-        return -1;
-
-    return result;
-}
-
-bool
-os_delete_file(const char *name)
-{
-#ifdef SYS_unlink
-    return (dynamorio_syscall(SYS_unlink, 1, name) == 0);
-#else
-    return (dynamorio_syscall(SYS_unlinkat, 3, AT_FDCWD, name, 0) == 0);
-#endif
-}
-
-bool
-os_rename_file(const char *orig_name, const char *new_name, bool replace)
-{
-    ptr_int_t res;
-    if (!replace) {
-        /* SYS_rename replaces so we must test beforehand => could have race */
-        /* _LARGEFILE64_SOURCE should make libc struct match kernel (see top of file) */
-        struct stat64 st;
-        ptr_int_t res = dynamorio_syscall_stat(new_name, &st);
-        if (res == 0)
-            return false;
-        else if (res != -ENOENT) {
-            LOG(THREAD_GET, LOG_SYSCALLS, 2, "%s stat failed: " PIFX "\n", __func__, res);
-            return false;
-        }
-    }
-#ifdef SYS_rename
-    res = dynamorio_syscall(SYS_rename, 2, orig_name, new_name);
-#else
-    res = dynamorio_syscall(SYS_renameat, 4, AT_FDCWD, orig_name, AT_FDCWD, new_name);
-#endif
-    if (res != 0) {
-        LOG(THREAD_GET, LOG_SYSCALLS, 2, "%s \"%s\" to \"%s\" failed: " PIFX "\n",
-            __func__, orig_name, new_name, res);
-    }
-    return (res == 0);
-}
-
-bool
-os_delete_mapped_file(const char *filename)
-{
-    return os_delete_file(filename);
-}
-
+/* There are enough differences vs the shared drlibc_os.c version that we override
+ * it here.  We use a loop to ensure reachability for the core.
+ */
 byte *
 os_map_file(file_t f, size_t *size INOUT, uint64 offs, app_pc addr, uint prot,
             map_flags_t map_flags)
 {
     int flags;
     byte *map;
-#if defined(X64) && !defined(NOT_DYNAMORIO_CORE_PROPER)
+#if defined(X64)
     bool loop = false;
     uint iters = 0;
 #    define MAX_MMAP_LOOP_ITERS 100
@@ -4498,7 +4074,7 @@ os_map_file(file_t f, size_t *size INOUT, uint64 offs, app_pc addr, uint prot,
 #else
     flags = TEST(MAP_FILE_COPY_ON_WRITE, map_flags) ? MAP_PRIVATE : MAP_SHARED;
 #endif
-#if defined(X64) && !defined(NOT_DYNAMORIO_CORE_PROPER)
+#if defined(X64)
     /* Allocate memory from reachable range for image: or anything (pcache
      * in particular): for low 4GB, easiest to just pass MAP_32BIT (which is
      * low 2GB, but good enough).
@@ -4512,8 +4088,7 @@ os_map_file(file_t f, size_t *size INOUT, uint64 offs, app_pc addr, uint prot,
         flags |= MAP_ANONYMOUS;
     if (TEST(MAP_FILE_FIXED, map_flags))
         flags |= MAP_FIXED;
-        /* Reachability is not supported for drinjectlib */
-#if defined(X64) && !defined(NOT_DYNAMORIO_CORE_PROPER)
+#if defined(X64)
     if (!TEST(MAP_32BIT, flags) && TEST(MAP_FILE_REACHABLE, map_flags)) {
         vmcode_get_reachable_region(&region_start, &region_end);
         /* addr need not be NULL: we'll use it if it's in the region */
@@ -4532,7 +4107,7 @@ os_map_file(file_t f, size_t *size INOUT, uint64 offs, app_pc addr, uint prot,
             LOG(THREAD_GET, LOG_SYSCALLS, 2, "%s failed: " PIFX "\n", __func__, map);
             map = NULL;
         }
-#if defined(X64) && !defined(NOT_DYNAMORIO_CORE_PROPER)
+#if defined(X64)
         else if (loop && (map < region_start || map + *size > region_end)) {
             /* Try again: probably a race.  Hopefully our notion of "there's a free
              * region big enough" matches the kernel's, else we'll loop forever
@@ -4554,16 +4129,6 @@ os_map_file(file_t f, size_t *size INOUT, uint64 offs, app_pc addr, uint prot,
 #endif
     return map;
 }
-
-bool
-os_unmap_file(byte *map, size_t size)
-{
-    long res = munmap_syscall(map, size);
-    return (res == 0);
-}
-
-/* around most of file, to exclude preload */
-#if !defined(NOT_DYNAMORIO_CORE_PROPER) || defined(STANDALONE_UNIT_TEST)
 
 bool
 os_get_disk_free_space(/*IN*/ file_t file_handle,
@@ -4591,7 +4156,7 @@ os_get_disk_free_space(/*IN*/ file_t file_handle,
     return true;
 }
 
-#    ifdef LINUX
+#ifdef LINUX
 static bool
 symlink_is_self_exe(const char *path)
 {
@@ -4600,9 +4165,9 @@ symlink_is_self_exe(const char *path)
      * (though it could also be under DR we have no simple way to obtain
      * its actual app path).
      */
-#        define SELF_LEN_LEADER 6  /* "/proc/" */
-#        define SELF_LEN_TRAILER 4 /* "/exe" */
-#        define SELF_LEN_MAX 18
+#    define SELF_LEN_LEADER 6  /* "/proc/" */
+#    define SELF_LEN_TRAILER 4 /* "/exe" */
+#    define SELF_LEN_MAX 18
     size_t len = strlen(path);
     if (strcmp(path, "/proc/self/exe") == 0)
         return true;
@@ -4619,7 +4184,7 @@ symlink_is_self_exe(const char *path)
     }
     return false;
 }
-#    endif
+#endif
 
 void
 exit_process_syscall(long status)
@@ -4640,15 +4205,15 @@ exit_process_syscall(long status)
 void
 exit_thread_syscall(long status)
 {
-#    ifdef MACOS
+#ifdef MACOS
     mach_port_t thread_port = dynamorio_mach_syscall(MACH_thread_self_trap, 0);
     /* FIXME i#1403: on MacOS we fail to free the app's stack: we need to pass it to
      * bsdthread_terminate.
      */
     dynamorio_syscall(SYSNUM_EXIT_THREAD, 4, 0, 0, thread_port, 0);
-#    else
+#else
     dynamorio_syscall(SYSNUM_EXIT_THREAD, 1, status);
-#    endif
+#endif
 }
 
 /* FIXME: this one will not be easily internationalizable
@@ -4798,8 +4363,6 @@ is_user_address(byte *pc)
     return true;
 }
 
-#endif /* !NOT_DYNAMORIO_CORE_PROPER */
-
 /* change protections on memory region starting at pc of length length
  * this does not update the all memory area info
  */
@@ -4843,22 +4406,20 @@ os_set_protection(byte *pc, size_t length, uint prot /*MEMPROT_*/)
     return true;
 }
 
-#ifndef NOT_DYNAMORIO_CORE_PROPER
-
 /* change protections on memory region starting at pc of length length */
 bool
 set_protection(byte *pc, size_t length, uint prot /*MEMPROT_*/)
 {
     if (os_set_protection(pc, length, prot) == false)
         return false;
-#    ifndef HAVE_MEMINFO_QUERY
+#ifndef HAVE_MEMINFO_QUERY
     else {
         app_pc start_page = (app_pc)PAGE_START(pc);
         uint num_bytes = ALIGN_FORWARD(length + (pc - start_page), PAGE_SIZE);
         memcache_update_locked(start_page, start_page + num_bytes, prot,
                                -1 /*type unchanged*/, true /*exists*/);
     }
-#    endif
+#endif
     return true;
 }
 
@@ -4882,41 +4443,41 @@ make_writable(byte *pc, size_t size)
      * this is crucial on modern linux kernels which refuse to mark stack +x.
      */
     if (!is_in_dynamo_dll(pc) /*avoid allmem assert*/ &&
-#    ifdef STATIC_LIBRARY
+#ifdef STATIC_LIBRARY
         /* FIXME i#975: is_in_dynamo_dll() is always false for STATIC_LIBRARY,
          * but we can't call get_memory_info() until allmem is initialized.  Our
          * uses before then are for patching x86.asm, which is OK.
          */
         IF_NO_MEMQUERY(memcache_initialized() &&)
-#    endif
+#endif
             get_memory_info(pc, NULL, NULL, &prot))
         prot |= PROT_WRITE;
 
     ASSERT(start_page == pc && ALIGN_FORWARD(size, PAGE_SIZE) == size);
-#    ifdef IA32_ON_IA64
+#ifdef IA32_ON_IA64
     LOG(THREAD_GET, LOG_VMAREAS, 1, "protection change not supported on IA64\n");
     LOG(THREAD_GET, LOG_VMAREAS, 3,
         "attempted make_writable: pc " PFX " -> " PFX "-" PFX "\n", pc, start_page,
         start_page + prot_size);
-#    else
+#else
     res = mprotect_syscall((void *)start_page, prot_size, prot);
     LOG(THREAD_GET, LOG_VMAREAS, 3, "make_writable: pc " PFX " -> " PFX "-" PFX " %d\n",
         pc, start_page, start_page + prot_size, res);
     ASSERT(res == 0);
     if (res != 0)
         return false;
-#    endif
+#endif
     STATS_INC(protection_change_calls);
     STATS_ADD(protection_change_pages, size / PAGE_SIZE);
 
-#    ifndef HAVE_MEMINFO_QUERY
+#ifndef HAVE_MEMINFO_QUERY
     /* update all_memory_areas list with the protection change */
     if (memcache_initialized()) {
         memcache_update_locked(start_page, start_page + prot_size,
                                osprot_to_memprot(prot), -1 /*type unchanged*/,
                                true /*exists*/);
     }
-#    endif
+#endif
     return true;
 }
 
@@ -4940,13 +4501,13 @@ make_unwritable(byte *pc, size_t size)
      * this is crucial on modern linux kernels which refuse to mark stack +x.
      */
     if (!is_in_dynamo_dll(pc) /*avoid allmem assert*/ &&
-#    ifdef STATIC_LIBRARY
+#ifdef STATIC_LIBRARY
         /* FIXME i#975: is_in_dynamo_dll() is always false for STATIC_LIBRARY,
          * but we can't call get_memory_info() until allmem is initialized.  Our
          * uses before then are for patching x86.asm, which is OK.
          */
         IF_NO_MEMQUERY(memcache_initialized() &&)
-#    endif
+#endif
             get_memory_info(pc, NULL, NULL, &prot))
         prot &= ~PROT_WRITE;
 
@@ -4954,26 +4515,26 @@ make_unwritable(byte *pc, size_t size)
     /* inc stats before making unwritable, in case messing w/ data segment */
     STATS_INC(protection_change_calls);
     STATS_ADD(protection_change_pages, size / PAGE_SIZE);
-#    ifdef IA32_ON_IA64
+#ifdef IA32_ON_IA64
     LOG(THREAD_GET, LOG_VMAREAS, 1, "protection change not supported on IA64\n");
     LOG(THREAD_GET, LOG_VMAREAS, 3,
         "attempted make_writable: pc " PFX " -> " PFX "-" PFX "\n", pc, start_page,
         start_page + prot_size);
-#    else
+#else
     res = mprotect_syscall((void *)start_page, prot_size, prot);
     LOG(THREAD_GET, LOG_VMAREAS, 3, "make_unwritable: pc " PFX " -> " PFX "-" PFX "\n",
         pc, start_page, start_page + prot_size);
     ASSERT(res == 0);
 
-#        ifndef HAVE_MEMINFO_QUERY
+#    ifndef HAVE_MEMINFO_QUERY
     /* update all_memory_areas list with the protection change */
     if (memcache_initialized()) {
         memcache_update_locked(start_page, start_page + prot_size,
                                osprot_to_memprot(prot), -1 /*type unchanged*/,
                                false /*!exists*/);
     }
-#        endif
 #    endif
+#endif
 }
 
 /****************************************************************************/
@@ -4996,7 +4557,7 @@ make_unwritable(byte *pc, size_t size)
 int
 os_normalized_sysnum(int num_raw, instr_t *gateway, dcontext_t *dcontext)
 {
-#    ifdef MACOS
+#ifdef MACOS
     /* The x64 encoding indicates the syscall type in the top 8 bits.
      * We drop the 0x2000000 for BSD so we can use the SYS_ enum constants.
      * That leaves 0x1000000 for Mach and 0x3000000 for Machdep.
@@ -5021,94 +4582,94 @@ os_normalized_sysnum(int num_raw, instr_t *gateway, dcontext_t *dcontext)
             }
         }
     }
-#        ifdef X64
+#    ifdef X64
     if (num_raw >> 24 == 0x2)
         return (int)(num_raw & 0xffffff); /* Drop BSD bit */
     else
         num = (int)num_raw; /* Keep Mach and Machdep bits */
-#        else
+#    else
     if ((ptr_int_t)num_raw < 0) /* Mach syscall */
         return (SYSCALL_NUM_MARKER_MACH | -(int)num_raw);
     else {
         /* Bottom 16 bits are the number, top are arg size. */
         num = (int)(num_raw & 0xffff);
     }
-#        endif
+#    endif
     if (interrupt == 0x81)
         num |= SYSCALL_NUM_MARKER_MACH;
     else if (interrupt == 0x82)
         num |= SYSCALL_NUM_MARKER_MACHDEP;
     return num;
-#    else
+#else
     return num_raw;
-#    endif
+#endif
 }
 
 static bool
 ignorable_system_call_normalized(int num)
 {
     switch (num) {
-#    if defined(SYS_exit_group)
+#if defined(SYS_exit_group)
     case SYS_exit_group:
-#    endif
+#endif
     case SYS_exit:
-#    ifdef MACOS
+#ifdef MACOS
     case SYS_bsdthread_terminate:
-#    endif
-#    ifdef LINUX
+#endif
+#ifdef LINUX
     case SYS_brk:
-#        ifdef SYS_uselib
+#    ifdef SYS_uselib
     case SYS_uselib:
-#        endif
 #    endif
-#    if defined(X64) || !defined(ARM)
+#endif
+#if defined(X64) || !defined(ARM)
     case SYS_mmap:
-#    endif
-#    if !defined(X64) && !defined(MACOS)
+#endif
+#if !defined(X64) && !defined(MACOS)
     case SYS_mmap2:
-#    endif
+#endif
     case SYS_munmap:
-#    ifdef LINUX
+#ifdef LINUX
     case SYS_mremap:
-#    endif
+#endif
     case SYS_mprotect:
-#    ifdef ANDROID
+#ifdef ANDROID
     case SYS_prctl:
-#    endif
+#endif
     case SYS_execve:
-#    ifdef LINUX
+#ifdef LINUX
     case SYS_clone:
-#    elif defined(MACOS)
+#elif defined(MACOS)
     case SYS_bsdthread_create:
     case SYS_posix_spawn:
-#    endif
-#    ifdef SYS_fork
+#endif
+#ifdef SYS_fork
     case SYS_fork:
-#    endif
-#    ifdef SYS_vfork
+#endif
+#ifdef SYS_vfork
     case SYS_vfork:
-#    endif
+#endif
     case SYS_kill:
-#    if defined(SYS_tkill)
+#if defined(SYS_tkill)
     case SYS_tkill:
-#    endif
-#    if defined(SYS_tgkill)
+#endif
+#if defined(SYS_tgkill)
     case SYS_tgkill:
-#    endif
-#    if defined(LINUX) && !defined(X64) && !defined(ARM)
+#endif
+#if defined(LINUX) && !defined(X64) && !defined(ARM)
     case SYS_signal:
-#    endif
-#    ifdef MACOS
+#endif
+#ifdef MACOS
     case SYS_sigsuspend_nocancel:
-#    endif
-#    if !defined(X64) || defined(MACOS)
+#endif
+#if !defined(X64) || defined(MACOS)
     case SYS_sigaction:
     case SYS_sigsuspend:
     case SYS_sigpending:
     case SYS_sigreturn:
     case SYS_sigprocmask:
-#    endif
-#    ifdef LINUX
+#endif
+#ifdef LINUX
     case SYS_rt_sigreturn:
     case SYS_rt_sigaction:
     case SYS_rt_sigprocmask:
@@ -5116,80 +4677,80 @@ ignorable_system_call_normalized(int num)
     case SYS_rt_sigtimedwait:
     case SYS_rt_sigqueueinfo:
     case SYS_rt_sigsuspend:
-#        ifdef SYS_signalfd
+#    ifdef SYS_signalfd
     case SYS_signalfd:
-#        endif
-    case SYS_signalfd4:
 #    endif
+    case SYS_signalfd4:
+#endif
     case SYS_sigaltstack:
-#    if defined(LINUX) && !defined(X64) && !defined(ARM)
+#if defined(LINUX) && !defined(X64) && !defined(ARM)
     case SYS_sgetmask:
     case SYS_ssetmask:
-#    endif
+#endif
     case SYS_setitimer:
     case SYS_getitimer:
-#    ifdef MACOS
+#ifdef MACOS
     case SYS_close_nocancel:
-#    endif
+#endif
     case SYS_close:
-#    ifdef SYS_dup2
+#ifdef SYS_dup2
     case SYS_dup2:
-#    endif
-#    ifdef LINUX
+#endif
+#ifdef LINUX
     case SYS_dup3:
-#    endif
-#    ifdef MACOS
+#endif
+#ifdef MACOS
     case SYS_fcntl_nocancel:
-#    endif
+#endif
     case SYS_fcntl:
-#    if defined(X64) || !defined(ARM)
+#if defined(X64) || !defined(ARM)
     case SYS_getrlimit:
-#    endif
-#    if defined(LINUX) && !defined(X64)
+#endif
+#if defined(LINUX) && !defined(X64)
     case SYS_ugetrlimit:
-#    endif
+#endif
     case SYS_setrlimit:
-#    ifdef LINUX
+#ifdef LINUX
     case SYS_prlimit64:
-#    endif
-#    if defined(LINUX) && defined(X86)
+#endif
+#if defined(LINUX) && defined(X86)
     /* i#784: app may have behavior relying on SIGALRM */
     case SYS_alarm:
-#    endif
+#endif
         /* i#107: syscall might change/query app's seg memory
          * need stop app from clobbering our GDT slot.
          */
-#    if defined(LINUX) && defined(X86) && defined(X64)
+#if defined(LINUX) && defined(X86) && defined(X64)
     case SYS_arch_prctl:
-#    endif
-#    if defined(LINUX) && defined(X86)
+#endif
+#if defined(LINUX) && defined(X86)
     case SYS_set_thread_area:
     case SYS_get_thread_area:
         /* FIXME: we might add SYS_modify_ldt later. */
-#    endif
-#    if defined(LINUX) && defined(ARM)
+#endif
+#if defined(LINUX) && defined(ARM)
     /* syscall changes app's thread register */
     case SYS_set_tls:
     case SYS_cacheflush:
-#    endif
-#    if defined(LINUX)
+#endif
+#if defined(LINUX)
     /* syscalls change procsigmask */
     case SYS_pselect6:
     case SYS_ppoll:
     case SYS_epoll_pwait:
-#    endif
+#endif
         return false;
-#    ifdef LINUX
-#        ifdef SYS_readlink
+#ifdef LINUX
+#    ifdef SYS_readlink
     case SYS_readlink:
-#        endif
-    case SYS_readlinkat: return !DYNAMO_OPTION(early_inject);
 #    endif
+    case SYS_readlinkat: return !DYNAMO_OPTION(early_inject);
+#endif
     default:
-#    ifdef VMX86_SERVER
+#ifdef VMX86_SERVER
         if (is_vmkuw_sysnum(num))
             return vmkuw_ignorable_system_call(num);
-#    endif
+#endif
         return true;
     }
 }
@@ -5210,30 +4771,12 @@ typedef struct {
     unsigned long offset;
 } mmap_arg_struct_t;
 
-#endif /* !NOT_DYNAMORIO_CORE_PROPER: around most of file, to exclude preload */
-
-const reg_id_t syscall_regparms[MAX_SYSCALL_ARGS] = {
-#ifdef X86
-#    ifdef X64
-    DR_REG_RDI, DR_REG_RSI,
-    DR_REG_RDX, DR_REG_R10, /* RCX goes here in normal x64 calling contention. */
-    DR_REG_R8,  DR_REG_R9
-#    else
-    DR_REG_EBX, DR_REG_ECX, DR_REG_EDX, DR_REG_ESI, DR_REG_EDI, DR_REG_EBP
-#    endif /* 64/32-bit */
-#elif defined(AARCHXX)
-    DR_REG_R0, DR_REG_R1, DR_REG_R2, DR_REG_R3, DR_REG_R4, DR_REG_R5,
-#endif /* X86/ARM */
-};
-
-#ifndef NOT_DYNAMORIO_CORE_PROPER
-
 static inline reg_t *
 sys_param_addr(dcontext_t *dcontext, int num)
 {
     /* we force-inline get_mcontext() and so don't take it as a param */
     priv_mcontext_t *mc = get_mcontext(dcontext);
-#    if defined(X86) && defined(X64)
+#if defined(X86) && defined(X64)
     switch (num) {
     case 0: return &mc->xdi;
     case 1: return &mc->xsi;
@@ -5243,8 +4786,8 @@ sys_param_addr(dcontext_t *dcontext, int num)
     case 5: return &mc->r9;
     default: CLIENT_ASSERT(false, "invalid system call parameter number");
     }
-#    else
-#        ifdef MACOS
+#else
+#    ifdef MACOS
     /* XXX: if we don't end up using dcontext->sys_was_int here, we could
      * make that field Linux-only.
      */
@@ -5252,7 +4795,7 @@ sys_param_addr(dcontext_t *dcontext, int num)
      * (regardless of whether using a sysenter or int gateway).
      */
     return ((reg_t *)mc->esp) + 1 /*retaddr*/ + num;
-#        endif
+#    endif
     /* even for vsyscall where ecx (syscall) or esp (sysenter) are saved into
      * ebp, the original parameter registers are not yet changed pre-syscall,
      * except for ebp, which is pushed on the stack:
@@ -5277,13 +4820,13 @@ sys_param_addr(dcontext_t *dcontext, int num)
     case 5:
         return IF_X86_ELSE((dcontext->sys_was_int ? &mc->xbp : ((reg_t *)mc->xsp)),
                            &mc->r5);
-#        ifdef ARM
+#    ifdef ARM
     /* AArch32 supposedly has 7 args in some cases. */
     case 6: return &mc->r6;
-#        endif
+#    endif
     default: CLIENT_ASSERT(false, "invalid system call parameter number");
     }
-#    endif
+#endif
     return 0;
 }
 
@@ -5302,7 +4845,7 @@ set_syscall_param(dcontext_t *dcontext, int param_num, reg_t new_value)
 static inline bool
 syscall_successful(priv_mcontext_t *mc, int normalized_sysnum)
 {
-#    ifdef MACOS
+#ifdef MACOS
     if (TEST(SYSCALL_NUM_MARKER_MACH, normalized_sysnum)) {
         /* XXX: Mach syscalls vary (for some KERN_SUCCESS=0 is success,
          * for others that return mach_port_t 0 is failure (I think?).
@@ -5311,15 +4854,15 @@ syscall_successful(priv_mcontext_t *mc, int normalized_sysnum)
         return ((ptr_int_t)MCXT_SYSCALL_RES(mc) >= 0);
     } else
         return !TEST(EFLAGS_CF, mc->eflags);
-#    else
+#else
     if (normalized_sysnum == IF_X64_ELSE(SYS_mmap, SYS_mmap2) ||
-#        if !defined(ARM) && !defined(X64)
+#    if !defined(ARM) && !defined(X64)
         normalized_sysnum == SYS_mmap ||
-#        endif
+#    endif
         normalized_sysnum == SYS_mremap)
         return mmap_syscall_succeeded((byte *)MCXT_SYSCALL_RES(mc));
     return ((ptr_int_t)MCXT_SYSCALL_RES(mc) >= 0);
-#    endif
+#endif
 }
 
 /* For non-Mac, this does nothing to indicate "success": you can pass -errno.
@@ -5332,12 +4875,12 @@ set_success_return_val(dcontext_t *dcontext, reg_t val)
 {
     /* since always coming from dispatch now, only need to set mcontext */
     priv_mcontext_t *mc = get_mcontext(dcontext);
-#    ifdef MACOS
+#ifdef MACOS
     /* On MacOS, success is determined by CF, except for Mach syscalls, but
      * there it doesn't hurt to set CF.
      */
     mc->eflags &= ~(EFLAGS_CF);
-#    endif
+#endif
     MCXT_SYSCALL_RES(mc) = val;
 }
 
@@ -5346,16 +4889,16 @@ static inline void
 set_failure_return_val(dcontext_t *dcontext, uint errno_val)
 {
     priv_mcontext_t *mc = get_mcontext(dcontext);
-#    ifdef MACOS
+#ifdef MACOS
     /* On MacOS, success is determined by CF, and errno is positive */
     mc->eflags |= EFLAGS_CF;
     MCXT_SYSCALL_RES(mc) = errno_val;
-#    else
+#else
     MCXT_SYSCALL_RES(mc) = -(int)errno_val;
-#    endif
+#endif
 }
 
-#    ifdef CLIENT_INTERFACE
+#ifdef CLIENT_INTERFACE
 DR_API
 reg_t
 dr_syscall_get_param(void *drcontext, int param_num)
@@ -5489,36 +5032,36 @@ dr_syscall_invoke_another(void *drcontext)
                   "event");
     LOG(THREAD, LOG_SYSCALLS, 2, "invoking additional syscall on client request\n");
     dcontext->client_data->invoke_another_syscall = true;
-#        ifdef X86
+#    ifdef X86
     if (get_syscall_method() == SYSCALL_METHOD_SYSENTER) {
         priv_mcontext_t *mc = get_mcontext(dcontext);
         /* restore xbp to xsp */
         mc->xbp = mc->xsp;
     }
-#        endif /* X86 */
+#    endif /* X86 */
     /* for x64 we don't need to copy xcx into r10 b/c we use r10 as our param */
 }
-#    endif /* CLIENT_INTERFACE */
+#endif /* CLIENT_INTERFACE */
 
 static inline bool
 is_thread_create_syscall_helper(ptr_uint_t sysnum, ptr_uint_t flags)
 {
-#    ifdef MACOS
+#ifdef MACOS
     /* XXX i#1403: we need earlier injection to intercept
      * bsdthread_register in order to capture workqueue threads.
      */
     return (sysnum == SYS_bsdthread_create || sysnum == SYS_vfork);
-#    else
-#        ifdef SYS_vfork
+#else
+#    ifdef SYS_vfork
     if (sysnum == SYS_vfork)
         return true;
-#        endif
-#        ifdef LINUX
+#    endif
+#    ifdef LINUX
     if (sysnum == SYS_clone && TEST(CLONE_VM, flags))
         return true;
-#        endif
-    return false;
 #    endif
+    return false;
+#endif
 }
 
 bool
@@ -5539,11 +5082,11 @@ was_thread_create_syscall(dcontext_t *dcontext)
 bool
 is_sigreturn_syscall_number(int sysnum)
 {
-#    ifdef MACOS
+#ifdef MACOS
     return sysnum == SYS_sigreturn;
-#    else
+#else
     return (IF_NOT_X64(sysnum == SYS_sigreturn ||) sysnum == SYS_rt_sigreturn);
-#    endif
+#endif
 }
 
 bool
@@ -5618,7 +5161,7 @@ static const char *const env_to_propagate[] = {
     /* these will only be propagated if they exist */
     DYNAMORIO_VAR_CONFIGDIR,
 };
-#    define NUM_ENV_TO_PROPAGATE (sizeof(env_to_propagate) / sizeof(env_to_propagate[0]))
+#define NUM_ENV_TO_PROPAGATE (sizeof(env_to_propagate) / sizeof(env_to_propagate[0]))
 
 /* Called at pre-SYS_execve to append DR vars in the target process env vars list.
  * For late injection via libdrpreload, we call this for *all children, because
@@ -5986,9 +5529,9 @@ handle_execve(dcontext_t *dcontext)
     char *inject_library_path;
     char rununder_buf[16]; /* just an integer printed in ascii */
     bool app_specific, from_env, rununder_on;
-#    if defined(LINUX) || defined(DEBUG)
+#if defined(LINUX) || defined(DEBUG)
     const char **argv;
-#    endif
+#endif
 
     if (DYNAMO_OPTION(follow_children) && DYNAMO_OPTION(early_inject)) {
         int ret = handle_execve_script(dcontext);
@@ -5997,18 +5540,18 @@ handle_execve(dcontext_t *dcontext)
     }
 
     fname = (char *)sys_param(dcontext, 0);
-#    if defined(LINUX) || defined(DEBUG)
+#if defined(LINUX) || defined(DEBUG)
     argv = (const char **)sys_param(dcontext, 1);
-#    endif
+#endif
 
-#    ifdef LINUX
+#ifdef LINUX
     if (DYNAMO_OPTION(early_inject) && symlink_is_self_exe(fname)) {
         /* i#907: /proc/self/exe points at libdynamorio.so.  Make sure we run
          * the right thing here.
          */
         fname = get_application_name();
     }
-#    endif
+#endif
 
     LOG(GLOBAL, LOG_ALL, 1,
         "\n---------------------------------------------------------------------------"
@@ -6052,12 +5595,12 @@ handle_execve(dcontext_t *dcontext)
      */
     mark_thread_execve(dcontext->thread_record, true);
 
-#    ifdef STATIC_LIBRARY
+#ifdef STATIC_LIBRARY
     /* no way we can inject, we just lose control */
     SYSLOG_INTERNAL_WARNING("WARNING: static DynamoRIO library, losing control on "
                             "execve");
     return 0;
-#    endif
+#endif
 
     /* Issue 20: handle cross-architecture execve */
     /* Xref alternate solution i#145: use dual paths on
@@ -6092,7 +5635,7 @@ handle_execve(dcontext_t *dcontext)
         dcontext->sys_param1 = 0;
     }
 
-#    ifdef LINUX
+#ifdef LINUX
     /* We have to be accurate with expect_to_fail as we cannot come back
      * and fail the syscall once the kernel execs DR!
      */
@@ -6126,7 +5669,7 @@ handle_execve(dcontext_t *dcontext)
         dcontext->sys_param3 = 0; /* no restore in post */
         dcontext->sys_param4 = 0; /* no restore in post */
     }
-#    endif
+#endif
 
     /* we need to clean up the .1config file here.  if the execve fails,
      * we'll just live w/o dynamic option re-read.
@@ -6145,11 +5688,11 @@ handle_execve_post(dcontext_t *dcontext)
      */
     char **old_envp = (char **)dcontext->sys_param0;
     char **new_envp = (char **)dcontext->sys_param1;
-#    ifdef STATIC_LIBRARY
+#ifdef STATIC_LIBRARY
     /* nothing to clean up */
     return;
-#    endif
-#    ifdef LINUX
+#endif
+#ifdef LINUX
     if (dcontext->sys_param4 != 0) {
         /* restore original /proc/.../exe */
         *sys_param_addr(dcontext, 0) = dcontext->sys_param4;
@@ -6159,7 +5702,7 @@ handle_execve_post(dcontext_t *dcontext)
             argv[0] = (const char *)dcontext->sys_param4;
         }
     }
-#    endif
+#endif
     if (new_envp != NULL) {
         int i;
         LOG(THREAD, LOG_SYSCALLS, 2, "\tcleaning up our env vars\n");
@@ -6210,10 +5753,10 @@ cleanup_after_vfork_execve(dcontext_t *dcontext)
 static void
 set_stdfile_fileno(stdfile_t **stdfile, file_t file_no)
 {
-#    ifdef STDFILE_FILENO
+#ifdef STDFILE_FILENO
     (*stdfile)->STDFILE_FILENO = file_no;
-#    else
-#        warning stdfile_t is opaque; DynamoRIO will not set fds of libc FILEs.
+#else
+#    warning stdfile_t is opaque; DynamoRIO will not set fds of libc FILEs.
     /* i#1973: musl libc support (and potentially other non-glibcs) */
     /* only called by handle_close_pre(), so warning is specific to that. */
     SYSLOG_INTERNAL_WARNING_ONCE(
@@ -6221,7 +5764,7 @@ set_stdfile_fileno(stdfile_t **stdfile, file_t file_no)
         "this platform. Client usage of stdio.h stdin, stdout, or stderr may "
         "no longer work as expected, because the app is closing the UNIX fds "
         "backing these.");
-#    endif
+#endif
 }
 
 /* returns whether to execute syscall */
@@ -6307,11 +5850,11 @@ handle_close_pre(dcontext_t *dcontext)
  * is currently in a syscall handler.
  * Alternatively for sysenter we could set app_sysenter_instr_addr for Linux.
  */
-#    define SYSCALL_PC(dc)                                                              \
-        ((get_syscall_method() == SYSCALL_METHOD_INT ||                                 \
-          get_syscall_method() == SYSCALL_METHOD_SYSCALL)                               \
-             ? (ASSERT(SYSCALL_LENGTH == INT_LENGTH), POST_SYSCALL_PC(dc) - INT_LENGTH) \
-             : (vsyscall_syscall_end_pc - SYSENTER_LENGTH))
+#define SYSCALL_PC(dc)                                                              \
+    ((get_syscall_method() == SYSCALL_METHOD_INT ||                                 \
+      get_syscall_method() == SYSCALL_METHOD_SYSCALL)                               \
+         ? (ASSERT(SYSCALL_LENGTH == INT_LENGTH), POST_SYSCALL_PC(dc) - INT_LENGTH) \
+         : (vsyscall_syscall_end_pc - SYSENTER_LENGTH))
 
 static void
 handle_exit(dcontext_t *dcontext)
@@ -6403,11 +5946,12 @@ handle_exit(dcontext_t *dcontext)
                                 sys_param(dcontext, 2), sys_param(dcontext, 3));
 }
 
-#    if defined(LINUX) && defined(X86) /* XXX i#58: just until we have Mac support */
+#if defined(LINUX) && defined(X86) /* XXX i#58: just until we have Mac support \
+                                    */
 static bool
 os_set_app_thread_area(dcontext_t *dcontext, our_modify_ldt_t *user_desc)
 {
-#        ifdef X86
+#    ifdef X86
     int i;
     os_thread_data_t *ostd = dcontext->os_field;
     our_modify_ldt_t *desc = (our_modify_ldt_t *)ostd->app_thread_areas;
@@ -6460,17 +6004,17 @@ os_set_app_thread_area(dcontext_t *dcontext, our_modify_ldt_t *user_desc)
         GDT_SELECTOR(user_desc->entry_number) != read_thread_register(SEG_TLS) &&
         GDT_SELECTOR(user_desc->entry_number) != read_thread_register(LIB_SEG_TLS))
         return false;
-#        elif defined(ARM)
+#    elif defined(ARM)
     /* FIXME i#1551: NYI on ARM */
     ASSERT_NOT_IMPLEMENTED(false);
-#        endif /* X86/ARM */
+#    endif /* X86/ARM */
     return true;
 }
 
 static bool
 os_get_app_thread_area(dcontext_t *dcontext, our_modify_ldt_t *user_desc)
 {
-#        ifdef X86
+#    ifdef X86
     os_thread_data_t *ostd = (os_thread_data_t *)dcontext->os_field;
     our_modify_ldt_t *desc = (our_modify_ldt_t *)ostd->app_thread_areas;
     int i = user_desc->entry_number - tls_min_index();
@@ -6478,13 +6022,13 @@ os_get_app_thread_area(dcontext_t *dcontext, our_modify_ldt_t *user_desc)
         return false;
     if (desc[i].seg_not_present == 1)
         return false;
-#        elif defined(ARM)
+#    elif defined(ARM)
     /* FIXME i#1551: NYI on ARM */
     ASSERT_NOT_IMPLEMENTED(false);
-#        endif /* X86/ARM */
+#    endif /* X86/ARM */
     return true;
 }
-#    endif
+#endif
 
 /* This function is used for switch lib tls segment on creating thread.
  * We switch to app's lib tls seg before thread creation system call, i.e.
@@ -6504,7 +6048,7 @@ os_switch_lib_tls(dcontext_t *dcontext, bool to_app)
     return os_switch_seg_to_context(dcontext, LIB_SEG_TLS, to_app);
 }
 
-#    ifdef X86
+#ifdef X86
 /* dcontext can be NULL if !to_app */
 static bool
 os_switch_seg_to_base(dcontext_t *dcontext, os_local_state_t *os_tls, reg_id_t seg,
@@ -6515,7 +6059,7 @@ os_switch_seg_to_base(dcontext_t *dcontext, os_local_state_t *os_tls, reg_id_t s
     ASSERT(IF_X86_ELSE((seg == SEG_FS || seg == SEG_GS),
                        (seg == DR_REG_TPIDRURW || DR_REG_TPIDRURO)));
     switch (os_tls->tls_type) {
-#        ifdef X64
+#    ifdef X64
     case TLS_TYPE_ARCH_PRCTL: {
         res = tls_set_fs_gs_segment_base(os_tls->tls_type, seg, base, NULL);
         ASSERT(res);
@@ -6530,7 +6074,7 @@ os_switch_seg_to_base(dcontext_t *dcontext, os_local_state_t *os_tls, reg_id_t s
         }
         break;
     }
-#        endif
+#    endif
     case TLS_TYPE_GDT: {
         our_modify_ldt_t desc;
         uint index;
@@ -6612,13 +6156,13 @@ os_set_dr_tls_base(dcontext_t *dcontext, os_local_state_t *tls, byte *base)
     }
     return os_switch_seg_to_base(dcontext, tls, SEG_TLS, false, base);
 }
-#    endif /* X86 */
+#endif /* X86 */
 
 static bool
 os_switch_seg_to_context(dcontext_t *dcontext, reg_id_t seg, bool to_app)
 {
     os_local_state_t *os_tls = get_os_tls_from_dc(dcontext);
-#    ifdef X86
+#ifdef X86
     app_pc base;
     /* we can only update the executing thread's segment (i#920) */
     ASSERT_MESSAGE(CHKLVL_ASSERTS + 1 /*expensive*/, "can only act on executing thread",
@@ -6633,7 +6177,7 @@ os_switch_seg_to_context(dcontext_t *dcontext, reg_id_t seg, bool to_app)
         base = os_get_priv_tls_base(dcontext, seg);
     }
     return os_switch_seg_to_base(dcontext, os_tls, seg, to_app, base);
-#    elif defined(AARCHXX)
+#elif defined(AARCHXX)
     bool res = false;
     os_thread_data_t *ostd = (os_thread_data_t *)dcontext->os_field;
     ASSERT(INTERNAL_OPTION(private_loader));
@@ -6683,11 +6227,11 @@ os_switch_seg_to_context(dcontext_t *dcontext, reg_id_t seg, bool to_app)
     LOG(THREAD, LOG_LOADER, 2, "%s %s: set_tls swap success=%d for thread " TIDFMT "\n",
         __FUNCTION__, to_app ? "to app" : "to DR", res, get_thread_id());
     return res;
-#    elif defined(AARCH64)
+#elif defined(AARCH64)
     (void)os_tls;
     ASSERT_NOT_IMPLEMENTED(false); /* FIXME i#1569 */
     return false;
-#    endif /* X86/ARM/AARCH64 */
+#endif /* X86/ARM/AARCH64 */
 }
 
 /* System call interception: put any special handling here
@@ -6736,7 +6280,7 @@ pre_system_call(dcontext_t *dcontext)
     });
     LOG(THREAD, LOG_SYSCALLS, 2, "system call %d\n", dcontext->sys_num);
 
-#    if defined(LINUX) && defined(X86)
+#if defined(LINUX) && defined(X86)
     /* PR 313715: If we fail to hook the vsyscall page (xref PR 212570, PR 288330)
      * we fall back on int, but we have to tweak syscall param #5 (ebp)
      * Once we have PR 288330 we can remove this.
@@ -6752,19 +6296,19 @@ pre_system_call(dcontext_t *dcontext)
                        mc->xbp = 0;
                    });
     }
-#    endif
+#endif
 
     switch (dcontext->sys_num) {
 
     case SYSNUM_EXIT_PROCESS:
-#    if defined(LINUX) && VMX86_SERVER
+#if defined(LINUX) && VMX86_SERVER
         if (os_in_vmkernel_32bit()) {
             /* on esx 3.5 => ENOSYS, so wait for SYS_exit */
             LOG(THREAD, LOG_SYSCALLS, 2, "on esx35 => ignoring exitgroup\n");
             DODEBUG({ dcontext->expect_last_syscall_to_fail = true; });
             break;
         }
-#    endif
+#endif
         /* fall-through */
     case SYSNUM_EXIT_THREAD: {
         handle_exit(dcontext);
@@ -6774,7 +6318,7 @@ pre_system_call(dcontext_t *dcontext)
     /****************************************************************************/
     /* MEMORY REGIONS */
 
-#    if defined(LINUX) && !defined(X64) && !defined(ARM)
+#if defined(LINUX) && !defined(X64) && !defined(ARM)
     case SYS_mmap: {
         /* in /usr/src/linux/arch/i386/kernel/sys_i386.c:
            asmlinkage int old_mmap(struct mmap_arg_struct_t *arg)
@@ -6808,7 +6352,7 @@ pre_system_call(dcontext_t *dcontext)
         dcontext->sys_param0 = (reg_t)arg;
         break;
     }
-#    endif
+#endif
     case IF_MACOS_ELSE(SYS_mmap, IF_X64_ELSE(SYS_mmap, SYS_mmap2)): {
         /* in /usr/src/linux/arch/i386/kernel/sys_i386.c:
            asmlinkage long sys_mmap2(unsigned long addr, unsigned long len,
@@ -6895,14 +6439,14 @@ pre_system_call(dcontext_t *dcontext)
                                 false /* don't own thread_initexit_lock */,
                                 true /* image, FIXME: though not necessarily */);
         /* FIXME: case 4983 use is_elf_so_header() */
-#    ifndef HAVE_MEMINFO_QUERY
+#ifndef HAVE_MEMINFO_QUERY
         memcache_lock();
         memcache_remove(addr, addr + len);
         memcache_unlock();
-#    endif
+#endif
         break;
     }
-#    ifdef LINUX
+#ifdef LINUX
     case SYS_mremap: {
         /* in /usr/src/linux/mm/mmap.c:
            asmlinkage unsigned long sys_mremap(unsigned long addr,
@@ -6937,7 +6481,7 @@ pre_system_call(dcontext_t *dcontext)
         });
         break;
     }
-#    endif
+#endif
     case SYS_mprotect: {
         /* in /usr/src/linux/mm/mprotect.c:
            asmlinkage long sys_mprotect(unsigned long start, uint len,
@@ -6996,7 +6540,7 @@ pre_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    ifdef ANDROID
+#ifdef ANDROID
     case SYS_prctl:
         dcontext->sys_param0 = sys_param(dcontext, 0);
         dcontext->sys_param1 = sys_param(dcontext, 1);
@@ -7004,8 +6548,8 @@ pre_system_call(dcontext_t *dcontext)
         dcontext->sys_param3 = sys_param(dcontext, 3);
         dcontext->sys_param4 = sys_param(dcontext, 4);
         break;
-#    endif
-#    ifdef LINUX
+#endif
+#ifdef LINUX
     case SYS_brk: {
         if (DYNAMO_OPTION(emulate_brk)) {
             /* i#1004: emulate brk via a separate mmap */
@@ -7023,7 +6567,7 @@ pre_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#        ifdef SYS_uselib
+#    ifdef SYS_uselib
     case SYS_uselib: {
         /* Used to get the kernel to load a share library (legacy system call).
          * Was primarily used when statically linking to dynamically loaded shared
@@ -7033,13 +6577,13 @@ pre_system_call(dcontext_t *dcontext)
         ASSERT_NOT_IMPLEMENTED(false);
         break;
     }
-#        endif
 #    endif
+#endif
 
     /****************************************************************************/
     /* SPAWNING */
 
-#    ifdef LINUX
+#ifdef LINUX
     case SYS_clone: {
         /* in /usr/src/linux/arch/i386/kernel/process.c
          * 32-bit params: flags, newsp, ptid, tls, ctid
@@ -7085,7 +6629,7 @@ pre_system_call(dcontext_t *dcontext)
             os_fork_pre(dcontext);
         break;
     }
-#    elif defined(MACOS)
+#elif defined(MACOS)
     case SYS_bsdthread_create: {
         /* XXX i#1403: we need earlier injection to intercept
          * bsdthread_register in order to capture workqueue threads.
@@ -7111,9 +6655,9 @@ pre_system_call(dcontext_t *dcontext)
         ASSERT_NOT_IMPLEMENTED(false);
         break;
     }
-#    endif
+#endif
 
-#    ifdef SYS_vfork
+#ifdef SYS_vfork
     case SYS_vfork: {
         /* treat as if sys_clone with flags just as sys_vfork does */
         /* in /usr/src/linux/arch/i386/kernel/process.c */
@@ -7130,24 +6674,24 @@ pre_system_call(dcontext_t *dcontext)
          */
         IF_LINUX(ASSERT(is_thread_create_syscall(dcontext)));
         dcontext->sys_param1 = mc->xsp; /* for restoring in parent */
-#        ifdef MACOS
+#    ifdef MACOS
         create_clone_record(dcontext, (reg_t *)&mc->xsp, NULL, NULL);
-#        else
+#    else
         create_clone_record(dcontext, (reg_t *)&mc->xsp /*child uses parent sp*/);
-#        endif
+#    endif
         os_clone_pre(dcontext);
         os_new_thread_pre();
         break;
     }
-#    endif
+#endif
 
-#    ifdef SYS_fork
+#ifdef SYS_fork
     case SYS_fork: {
         LOG(THREAD, LOG_SYSCALLS, 2, "syscall: fork\n");
         os_fork_pre(dcontext);
         break;
     }
-#    endif
+#endif
 
     case SYS_execve: {
         int ret = handle_execve(dcontext);
@@ -7194,7 +6738,7 @@ pre_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    if defined(LINUX) && !defined(X64)
+#if defined(LINUX) && !defined(X64)
     case SYS_sigaction: { /* 67 */
         /* sys_sigaction(int sig, const struct old_sigaction *act,
          *               struct old_sigaction *oact)
@@ -7218,8 +6762,8 @@ pre_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    endif
-#    if defined(LINUX) && !defined(X64)
+#endif
+#if defined(LINUX) && !defined(X64)
     case SYS_sigreturn: { /* 119 */
         /* in /usr/src/linux/arch/i386/kernel/signal.c:
            asmlinkage int sys_sigreturn(unsigned long __unused)
@@ -7231,8 +6775,8 @@ pre_system_call(dcontext_t *dcontext)
          */
         break;
     }
-#    endif
-#    ifdef LINUX
+#endif
+#ifdef LINUX
     case SYS_rt_sigreturn: { /* 173 */
         /* in /usr/src/linux/arch/i386/kernel/signal.c:
            asmlinkage int sys_rt_sigreturn(unsigned long __unused)
@@ -7241,8 +6785,8 @@ pre_system_call(dcontext_t *dcontext)
         /* see comment for SYS_sigreturn on return val */
         break;
     }
-#    endif
-#    ifdef MACOS
+#endif
+#ifdef MACOS
     case SYS_sigreturn: {
         /* int sigreturn(struct ucontext *uctx, int infostyle) */
         execute_syscall = handle_sigreturn(dcontext, (void *)sys_param(dcontext, 0),
@@ -7250,7 +6794,7 @@ pre_system_call(dcontext_t *dcontext)
         /* see comment for SYS_sigreturn on return val */
         break;
     }
-#    endif
+#endif
     case SYS_sigaltstack: { /* 186 */
         /* in /usr/src/linux/arch/i386/kernel/signal.c:
            asmlinkage int
@@ -7290,9 +6834,9 @@ pre_system_call(dcontext_t *dcontext)
             set_success_return_val(dcontext, 0);
         break;
     }
-#    ifdef MACOS
+#ifdef MACOS
     case SYS_sigsuspend_nocancel:
-#    endif
+#endif
     case IF_MACOS_ELSE(SYS_sigsuspend, SYS_rt_sigsuspend): { /* 179 */
         /* in /usr/src/linux/kernel/signal.c:
            asmlinkage int
@@ -7302,10 +6846,10 @@ pre_system_call(dcontext_t *dcontext)
                           (size_t)sys_param(dcontext, 1));
         break;
     }
-#    ifdef LINUX
-#        ifdef SYS_signalfd
+#ifdef LINUX
+#    ifdef SYS_signalfd
     case SYS_signalfd: /* 282/321 */
-#        endif
+#    endif
     case SYS_signalfd4: { /* 289 */
         /* int signalfd (int fd, const sigset_t *mask, size_t sizemask) */
         /* int signalfd4(int fd, const sigset_t *mask, size_t sizemask, int flags) */
@@ -7313,11 +6857,11 @@ pre_system_call(dcontext_t *dcontext)
         dcontext->sys_param0 = sys_param(dcontext, 0);
         dcontext->sys_param1 = sys_param(dcontext, 1);
         dcontext->sys_param2 = sys_param(dcontext, 2);
-#        ifdef SYS_signalfd
+#    ifdef SYS_signalfd
         if (dcontext->sys_num == SYS_signalfd)
             dcontext->sys_param3 = 0;
         else
-#        endif
+#    endif
             dcontext->sys_param3 = sys_param(dcontext, 3);
         new_result = handle_pre_signalfd(
             dcontext, (int)dcontext->sys_param0, (kernel_sigset_t *)dcontext->sys_param1,
@@ -7327,7 +6871,7 @@ pre_system_call(dcontext_t *dcontext)
         set_success_return_val(dcontext, new_result);
         break;
     }
-#    endif
+#endif
     case SYS_kill: { /* 37 */
         /* in /usr/src/linux/kernel/signal.c:
          * asmlinkage long sys_kill(int pid, int sig)
@@ -7343,7 +6887,7 @@ pre_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    if defined(SYS_tkill)
+#if defined(SYS_tkill)
     case SYS_tkill: { /* 238 */
         /* in /usr/src/linux/kernel/signal.c:
          * asmlinkage long sys_tkill(int pid, int sig)
@@ -7357,8 +6901,8 @@ pre_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    endif
-#    if defined(SYS_tgkill)
+#endif
+#if defined(SYS_tgkill)
     case SYS_tgkill: { /* 270 */
         /* in /usr/src/linux/kernel/signal.c:
          * asmlinkage long sys_tgkill(int tgid, int pid, int sig)
@@ -7381,7 +6925,7 @@ pre_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    endif
+#endif
     case SYS_setitimer: /* 104 */
         dcontext->sys_param0 = sys_param(dcontext, 0);
         dcontext->sys_param1 = sys_param(dcontext, 1);
@@ -7394,14 +6938,14 @@ pre_system_call(dcontext_t *dcontext)
         dcontext->sys_param0 = sys_param(dcontext, 0);
         dcontext->sys_param1 = sys_param(dcontext, 1);
         break;
-#    if defined(LINUX) && defined(X86)
+#if defined(LINUX) && defined(X86)
     case SYS_alarm: /* 27 on x86 and 37 on x64 */
         dcontext->sys_param0 = sys_param(dcontext, 0);
         handle_pre_alarm(dcontext, (unsigned int)dcontext->sys_param0);
         break;
-#    endif
-#    if 0
-#        ifndef X64
+#endif
+#if 0
+#    ifndef X64
     case SYS_signal: {         /* 48 */
         /* in /usr/src/linux/kernel/signal.c:
            asmlinkage unsigned long
@@ -7423,29 +6967,29 @@ pre_system_call(dcontext_t *dcontext)
          */
         break;
     }
-#        endif
-#    else
+#    endif
+#else
         /* until we've implemented them, keep down here to get warning: */
-#        if defined(LINUX) && !defined(X64)
-#            ifndef ARM
+#    if defined(LINUX) && !defined(X64)
+#        ifndef ARM
     case SYS_signal:
-#            endif
+#        endif
     case SYS_sigsuspend:
     case SYS_sigprocmask:
-#        endif
 #    endif
+#endif
 
-#    if defined(LINUX) && !defined(X64)
+#if defined(LINUX) && !defined(X64)
     case SYS_sigpending: /* 73 */
-#        ifndef ARM
+#    ifndef ARM
     case SYS_sgetmask: /* 68 */
     case SYS_ssetmask: /* 69 */
-#        endif
 #    endif
-#    ifdef LINUX
+#endif
+#ifdef LINUX
     case SYS_rt_sigtimedwait: /* 177 */
     case SYS_rt_sigqueueinfo: /* 178 */
-#    endif
+#endif
     case IF_MACOS_ELSE(SYS_sigpending, SYS_rt_sigpending): { /* 176 */
         /* FIXME i#92: handle all of these syscalls! */
         LOG(THREAD, LOG_ASYNCH | LOG_SYSCALLS, 1,
@@ -7454,7 +6998,7 @@ pre_system_call(dcontext_t *dcontext)
                                      dcontext->sys_num);
         break;
     }
-#    ifdef LINUX
+#ifdef LINUX
     case SYS_ppoll: {
         kernel_sigset_t *sigmask = (kernel_sigset_t *)sys_param(dcontext, 3);
         size_t sizemask = (size_t)sys_param(dcontext, 4);
@@ -7553,7 +7097,7 @@ pre_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    endif
+#endif
 
     /****************************************************************************/
     /* FILES */
@@ -7561,19 +7105,19 @@ pre_system_call(dcontext_t *dcontext)
      * it's not worth monitoring all syscalls that take in fds from affecting ours.
      */
 
-#    ifdef MACOS
+#ifdef MACOS
     case SYS_close_nocancel:
-#    endif
+#endif
     case SYS_close: {
         execute_syscall = handle_close_pre(dcontext);
-#    ifdef LINUX
+#ifdef LINUX
         if (execute_syscall)
             signal_handle_close(dcontext, (file_t)sys_param(dcontext, 0));
-#    endif
+#endif
         break;
     }
 
-#    ifdef SYS_dup2
+#ifdef SYS_dup2
     case SYS_dup2:
         IF_LINUX(case SYS_dup3:)
         {
@@ -7591,11 +7135,11 @@ pre_system_call(dcontext_t *dcontext)
             }
             break;
         }
-#    endif
+#endif
 
-#    ifdef MACOS
+#ifdef MACOS
     case SYS_fcntl_nocancel:
-#    endif
+#endif
     case SYS_fcntl: {
         int cmd = (int)sys_param(dcontext, 1);
         long arg = (long)sys_param(dcontext, 2);
@@ -7616,12 +7160,12 @@ pre_system_call(dcontext_t *dcontext)
         break;
     }
 
-#    if defined(X64) || !defined(ARM) || defined(MACOS)
+#if defined(X64) || !defined(ARM) || defined(MACOS)
     case SYS_getrlimit:
-#    endif
-#    if defined(LINUX) && !defined(X64)
+#endif
+#if defined(LINUX) && !defined(X64)
     case SYS_ugetrlimit:
-#    endif
+#endif
         /* save for post */
         dcontext->sys_param0 = sys_param(dcontext, 0); /* resource */
         dcontext->sys_param1 = sys_param(dcontext, 1); /* rlimit */
@@ -7630,11 +7174,11 @@ pre_system_call(dcontext_t *dcontext)
     case SYS_setrlimit: {
         int resource = (int)sys_param(dcontext, 0);
         if (resource == RLIMIT_NOFILE && DYNAMO_OPTION(steal_fds) > 0) {
-#    if !defined(ARM) && !defined(X64) && !defined(MACOS)
+#if !defined(ARM) && !defined(X64) && !defined(MACOS)
             struct compat_rlimit rlim;
-#    else
+#else
             struct rlimit rlim;
-#    endif
+#endif
             if (!safe_read((void *)sys_param(dcontext, 1), sizeof(rlim), &rlim)) {
                 LOG(THREAD, LOG_SYSCALLS, 2, "\treturning EFAULT to app for prlimit64\n");
                 set_failure_return_val(dcontext, EFAULT);
@@ -7663,7 +7207,7 @@ pre_system_call(dcontext_t *dcontext)
         break;
     }
 
-#    ifdef LINUX
+#ifdef LINUX
     case SYS_prlimit64:
         /* save for post */
         dcontext->sys_param0 = sys_param(dcontext, 0); /* pid */
@@ -7722,12 +7266,12 @@ pre_system_call(dcontext_t *dcontext)
             execute_syscall = false;
         }
         break;
-#    endif
+#endif
 
-#    ifdef LINUX
-#        ifdef SYS_readlink
+#ifdef LINUX
+#    ifdef SYS_readlink
     case SYS_readlink:
-#        endif
+#    endif
     case SYS_readlinkat:
         if (DYNAMO_OPTION(early_inject)) {
             dcontext->sys_param0 = sys_param(dcontext, 0);
@@ -7740,15 +7284,15 @@ pre_system_call(dcontext_t *dcontext)
 
         /* i#107 syscalls that might change/query app's segment */
 
-#        if defined(X86) && defined(X64)
+#    if defined(X86) && defined(X64)
     case SYS_arch_prctl: {
         /* we handle arch_prctl in post_syscall */
         dcontext->sys_param0 = sys_param(dcontext, 0);
         dcontext->sys_param1 = sys_param(dcontext, 1);
         break;
     }
-#        endif
-#        ifdef X86
+#    endif
+#    ifdef X86
     case SYS_set_thread_area: {
         our_modify_ldt_t desc;
         if (INTERNAL_OPTION(mangle_app_seg) &&
@@ -7777,8 +7321,8 @@ pre_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#        endif /* X86 */
-#        ifdef ARM
+#    endif /* X86 */
+#    ifdef ARM
     case SYS_set_tls: {
         LOG(THREAD, LOG_VMAREAS | LOG_SYSCALLS, 2, "syscall: set_tls " PFX "\n",
             sys_param(dcontext, 0));
@@ -7808,33 +7352,33 @@ pre_system_call(dcontext_t *dcontext)
                                     false /*don't force synchall*/);
         break;
     }
-#        endif /* ARM */
-#    elif defined(MACOS)
+#    endif /* ARM */
+#elif defined(MACOS)
     /* FIXME i#58: handle i386_{get,set}_ldt and thread_fast_set_cthread_self64 */
-#    endif
+#endif
 
-#    ifdef DEBUG
-#        ifdef MACOS
+#ifdef DEBUG
+#    ifdef MACOS
     case SYS_open_nocancel:
-#        endif
-#        ifdef SYS_open
+#    endif
+#    ifdef SYS_open
     case SYS_open: {
         dcontext->sys_param0 = sys_param(dcontext, 0);
         break;
     }
-#        endif
 #    endif
+#endif
 
     default: {
-#    ifdef LINUX
+#ifdef LINUX
         execute_syscall = handle_restartable_region_syscall_pre(dcontext);
-#    endif
-#    ifdef VMX86_SERVER
+#endif
+#ifdef VMX86_SERVER
         if (is_vmkuw_sysnum(dcontext->sys_num)) {
             execute_syscall = vmkuw_pre_system_call(dcontext);
             break;
         }
-#    endif
+#endif
         break;
     }
 
@@ -7941,13 +7485,13 @@ mmap_check_for_module_overlap(app_pc base, size_t size, bool readable, uint64 in
         });
     }
     os_get_module_info_unlock();
-#    ifdef ANDROID
+#ifdef ANDROID
     /* i#1860: we need to keep looking for the segment with .dynamic as Android's
      * loader does not map the whole file up front.
      */
     if (ma != NULL && at_map && readable)
         os_module_update_dynamic_info(base, size, at_map);
-#    endif
+#endif
     return ma != NULL;
 }
 
@@ -8019,10 +7563,10 @@ os_add_new_app_module(dcontext_t *dcontext, bool at_map, app_pc base, size_t siz
         }
     }
     memquery_iterator_stop(&iter);
-#    ifdef HAVE_MEMINFO
+#ifdef HAVE_MEMINFO
     /* barring weird races we should find this map except */
     ASSERT_CURIOSITY(found_map);
-#    else  /* HAVE_MEMINFO */
+#else  /* HAVE_MEMINFO */
     /* Without /proc/maps or other memory querying interface available at
      * library map time, there is no way to find out the name of the file
      * that was mapped, thus its inode isn't available either.
@@ -8038,7 +7582,7 @@ os_add_new_app_module(dcontext_t *dcontext, bool at_map, app_pc base, size_t siz
      * Once PR 235433 is implemented in visor then fix memquery_iterator*() to
      * use vsi to find out page protection info, file name & inode.
      */
-#    endif /* HAVE_MEMINFO */
+#endif /* HAVE_MEMINFO */
     /* XREF 307599 on rounding module end to the next PAGE boundary */
     if (found_map) {
         module_list_add(base, ALIGN_FORWARD(mod_size, PAGE_SIZE), at_map, filename,
@@ -8077,11 +7621,11 @@ process_mmap(dcontext_t *dcontext, app_pc base, size_t size, uint prot,
 {
     bool image = false;
     uint memprot = osprot_to_memprot(prot);
-#    ifdef ANDROID
+#ifdef ANDROID
     /* i#1861: avoid merging file-backed w/ anon regions */
     if (!TEST(MAP_ANONYMOUS, flags))
         memprot |= MEMPROT_HAS_COMMENT;
-#    endif
+#endif
 
     LOG(THREAD, LOG_SYSCALLS, 4, "process_mmap(" PFX "," PFX ",0x%x,%s,%s)\n", base, size,
         flags, memprot_string(memprot), map_type);
@@ -8125,7 +7669,7 @@ process_mmap(dcontext_t *dcontext, app_pc base, size_t size, uint prot,
                 * read, so pass size=0 to use a safe_read.
                 */
                module_is_header(base, 0)) {
-#    ifdef ANDROID
+#ifdef ANDROID
         /* The Android loader's initial all-segment-covering mmap is anonymous */
         dr_mem_info_t info;
         if (query_memory_ex_from_os((byte *)ALIGN_FORWARD(base + size, PAGE_SIZE),
@@ -8138,7 +7682,7 @@ process_mmap(dcontext_t *dcontext, app_pc base, size_t size, uint prot,
                                   /* pass segment size, not whole module size */
                                   size, memprot);
         } else
-#    endif
+#endif
             if (module_is_partial_map(base, size, memprot)) {
             /* i#1240: App might read first page of ELF header using mmap, which
              * might accidentally be treated as a module load. Heuristically
@@ -8167,7 +7711,7 @@ process_mmap(dcontext_t *dcontext, app_pc base, size_t size, uint prot,
     LOG(THREAD, LOG_SYSCALLS, 4, "\t app_mem_alloc -- DONE\n");
 }
 
-#    ifdef LINUX
+#ifdef LINUX
 /* Call right after the system call.
  * i#173: old_prot and old_type should be from before the system call
  */
@@ -8233,7 +7777,7 @@ handle_app_brk(dcontext_t *dcontext, byte *lowest_brk /*if known*/, byte *old_br
     }
     IF_NO_MEMQUERY(memcache_handle_app_brk(lowest_brk, old_brk, new_brk));
 }
-#    endif
+#endif
 
 /* This routine is *not* called is pre_system_call() returns false to skip
  * the syscall.
@@ -8264,7 +7808,7 @@ post_system_call(dcontext_t *dcontext)
     old_whereami = dcontext->whereami;
     dcontext->whereami = DR_WHERE_SYSCALL_HANDLER;
 
-#    if defined(LINUX) && defined(X86)
+#if defined(LINUX) && defined(X86)
     /* PR 313715: restore xbp since for some vsyscall sequences that use
      * the syscall instruction its value is needed:
      *   0xffffe400 <__kernel_vsyscall+0>:       push   %ebp
@@ -8279,23 +7823,23 @@ post_system_call(dcontext_t *dcontext)
     if (should_syscall_method_be_sysenter() && !dcontext->sys_was_int) {
         mc->xbp = dcontext->sys_xbp;
     }
-#    endif
+#endif
 
     /* handle fork, try to do it early before too much logging occurs */
     if (false
-#    ifdef SYS_fork
+#ifdef SYS_fork
         || sysnum ==
             SYS_fork
-#    endif
+#endif
                 IF_LINUX(
                     || (sysnum == SYS_clone && !TEST(CLONE_VM, dcontext->sys_param0)))) {
         if (result == 0) {
             /* we're the child */
             thread_id_t child = get_sys_thread_id();
-#    ifdef DEBUG
+#ifdef DEBUG
             thread_id_t parent = get_parent_id();
             SYSLOG_INTERNAL_INFO("-- parent %d forked child %d --", parent, child);
-#    endif
+#endif
             /* first, fix TLS of dcontext */
             ASSERT(parent != 0);
             /* change parent pid to our pid */
@@ -8327,11 +7871,11 @@ post_system_call(dcontext_t *dcontext)
         /****************************************************************************/
         /* MEMORY REGIONS */
 
-#    ifdef DEBUG
-#        ifdef MACOS
+#ifdef DEBUG
+#    ifdef MACOS
     case SYS_open_nocancel:
-#        endif
-#        ifdef SYS_open
+#    endif
+#    ifdef SYS_open
     case SYS_open: {
         if (success) {
             /* useful for figuring out what module was loaded that then triggers
@@ -8342,12 +7886,12 @@ post_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#        endif
 #    endif
+#endif
 
-#    if defined(LINUX) && !defined(X64) && !defined(ARM)
+#if defined(LINUX) && !defined(X64) && !defined(ARM)
     case SYS_mmap:
-#    endif
+#endif
     case IF_MACOS_ELSE(SYS_mmap, IF_X64_ELSE(SYS_mmap, SYS_mmap2)): {
         uint flags;
         DEBUG_DECLARE(const char *map_type;)
@@ -8355,7 +7899,7 @@ post_system_call(dcontext_t *dcontext)
         base = (app_pc)MCXT_SYSCALL_RES(mc); /* For mmap, it's NOT arg->addr! */
         /* mmap isn't simply a user-space wrapper for mmap2. It's called
          * directly when dynamically loading an SO, i.e., dlopen(). */
-#    ifdef LINUX /* MacOS success is in CF */
+#ifdef LINUX /* MacOS success is in CF */
         success = mmap_syscall_succeeded((app_pc)result);
         /* The syscall either failed OR the retcode is less than the
          * largest uint value of any errno and the addr returned is
@@ -8364,12 +7908,12 @@ post_system_call(dcontext_t *dcontext)
         ASSERT_CURIOSITY(
             !success ||
             ((app_pc)result < (app_pc)(ptr_int_t)-0x1000 && ALIGNED(base, PAGE_SIZE)));
-#    else
+#else
         ASSERT_CURIOSITY(!success || ALIGNED(base, PAGE_SIZE));
-#    endif
+#endif
         if (!success)
             goto exit_post_system_call;
-#    if defined(LINUX) && !defined(X64) && !defined(ARM)
+#if defined(LINUX) && !defined(X64) && !defined(ARM)
         if (sysnum == SYS_mmap) {
             /* The syscall succeeded so the read of 'arg' should be
              * safe. */
@@ -8379,14 +7923,14 @@ post_system_call(dcontext_t *dcontext)
             flags = (uint)arg->flags;
             DEBUG_DECLARE(map_type = "mmap";)
         } else {
-#    endif
+#endif
             size = (size_t)dcontext->sys_param1;
             prot = (uint)dcontext->sys_param2;
             flags = (uint)dcontext->sys_param3;
             DEBUG_DECLARE(map_type = IF_X64_ELSE("mmap2", "mmap");)
-#    if defined(LINUX) && !defined(X64) && !defined(ARM)
+#if defined(LINUX) && !defined(X64) && !defined(ARM)
         }
-#    endif
+#endif
         process_mmap(dcontext, base, size, prot, flags _IF_DEBUG(map_type));
         break;
     }
@@ -8427,7 +7971,7 @@ post_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    ifdef LINUX
+#ifdef LINUX
     case SYS_mremap: {
         app_pc old_base = (app_pc)dcontext->sys_param0;
         size_t old_size = (size_t)dcontext->sys_param1;
@@ -8453,12 +7997,12 @@ post_system_call(dcontext_t *dcontext)
             goto exit_post_system_call;
         break;
     }
-#    endif
+#endif
     case SYS_mprotect: {
         base = (app_pc)dcontext->sys_param0;
         size = dcontext->sys_param1;
         prot = dcontext->sys_param2;
-#    ifdef VMX86_SERVER
+#ifdef VMX86_SERVER
         /* PR 475111: workaround for PR 107872 */
         if (os_in_vmkernel_userworld() && result == -EBUSY && prot == PROT_NONE) {
             result = mprotect_syscall(base, size, PROT_READ);
@@ -8470,7 +8014,7 @@ post_system_call(dcontext_t *dcontext)
                 (int)result);
             SYSLOG_INTERNAL_WARNING_ONCE("re-doing mprotect for PR 475111, PR 107872");
         }
-#    endif
+#endif
         /* FIXME i#143: we need to tweak the returned oldprot for
          * writable areas we've made read-only
          */
@@ -8513,7 +8057,7 @@ post_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    ifdef ANDROID
+#ifdef ANDROID
     case SYS_prctl: {
         int code = (int)dcontext->sys_param0;
         int subcode = (ulong)dcontext->sys_param1;
@@ -8536,8 +8080,8 @@ post_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    endif
-#    ifdef LINUX
+#endif
+#ifdef LINUX
     case SYS_brk: {
         /* i#91/PR 396352: need to watch SYS_brk to maintain all_memory_areas.
          * This code should work regardless of whether syscall failed
@@ -8548,7 +8092,7 @@ post_system_call(dcontext_t *dcontext)
         app_pc new_brk = (app_pc)result;
         DEBUG_DECLARE(app_pc req_brk = (app_pc)dcontext->sys_param0;);
         ASSERT(!DYNAMO_OPTION(emulate_brk)); /* shouldn't get here */
-#        ifdef DEBUG
+#    ifdef DEBUG
         if (DYNAMO_OPTION(early_inject) &&
             req_brk != NULL /* Ignore calls that don't increase brk. */) {
             DO_ONCE({
@@ -8557,16 +8101,16 @@ post_system_call(dcontext_t *dcontext)
                                  "allocation failed with -early_inject");
             });
         }
-#        endif
+#    endif
         handle_app_brk(dcontext, NULL, old_brk, new_brk);
         break;
     }
-#    endif
+#endif
 
     /****************************************************************************/
     /* SPAWNING -- fork mostly handled above */
 
-#    ifdef LINUX
+#ifdef LINUX
     case SYS_clone: {
         /* in /usr/src/linux/arch/i386/kernel/process.c */
         LOG(THREAD, LOG_SYSCALLS, 2, "syscall: clone returned " PFX "\n",
@@ -8583,7 +8127,7 @@ post_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    elif defined(MACOS) && !defined(X64)
+#elif defined(MACOS) && !defined(X64)
     case SYS_bsdthread_create: {
         /* restore stack values we clobbered */
         ASSERT(*sys_param_addr(dcontext, 0) == (reg_t)new_bsdthread_intercept);
@@ -8591,17 +8135,17 @@ post_system_call(dcontext_t *dcontext)
         *sys_param_addr(dcontext, 1) = dcontext->sys_param1;
         break;
     }
-#    endif
+#endif
 
-#    ifdef SYS_fork
+#ifdef SYS_fork
     case SYS_fork: {
         LOG(THREAD, LOG_SYSCALLS, 2, "syscall: fork returned " PFX "\n",
             MCXT_SYSCALL_RES(mc));
         break;
     }
-#    endif
+#endif
 
-#    ifdef SYS_vfork
+#ifdef SYS_vfork
     case SYS_vfork: {
         LOG(THREAD, LOG_SYSCALLS, 2, "syscall: vfork returned " PFX "\n",
             MCXT_SYSCALL_RES(mc));
@@ -8624,7 +8168,7 @@ post_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    endif
+#endif
 
     case SYS_execve: {
         /* if we get here it means execve failed (doesn't return on success) */
@@ -8663,7 +8207,7 @@ post_system_call(dcontext_t *dcontext)
             goto exit_post_system_call;
         break;
     }
-#    if defined(LINUX) && !defined(X64)
+#if defined(LINUX) && !defined(X64)
     case SYS_sigaction: { /* 67 */
         int sig = (int)dcontext->sys_param0;
         const old_sigaction_t *act = (const old_sigaction_t *)dcontext->sys_param1;
@@ -8676,7 +8220,7 @@ post_system_call(dcontext_t *dcontext)
             goto exit_post_system_call;
         break;
     }
-#    endif
+#endif
     case IF_MACOS_ELSE(SYS_sigprocmask, SYS_rt_sigprocmask): { /* 175 */
         /* in /usr/src/linux/kernel/signal.c:
            asmlinkage long
@@ -8689,9 +8233,9 @@ post_system_call(dcontext_t *dcontext)
             (kernel_sigset_t *)dcontext->sys_param2, (size_t)dcontext->sys_param3);
         break;
     }
-#    if defined(LINUX) && !defined(X64)
+#if defined(LINUX) && !defined(X64)
     case SYS_sigreturn: /* 119 */
-#    endif
+#endif
     case IF_MACOS_ELSE(SYS_sigreturn, SYS_rt_sigreturn): /* 173 */
         /* there is no return value: it's just the value of eax, so avoid
          * assert below
@@ -8708,12 +8252,12 @@ post_system_call(dcontext_t *dcontext)
         handle_post_getitimer(dcontext, success, (int)dcontext->sys_param0,
                               (struct itimerval *)dcontext->sys_param1);
         break;
-#    if defined(LINUX) && defined(X86)
+#if defined(LINUX) && defined(X86)
     case SYS_alarm: /* 27 on x86 and 37 on x64 */
         handle_post_alarm(dcontext, success, (unsigned int)dcontext->sys_param0);
         break;
-#    endif
-#    if defined(LINUX) && defined(X86) && defined(X64)
+#endif
+#if defined(LINUX) && defined(X86) && defined(X64)
     case SYS_arch_prctl: {
         if (success && INTERNAL_OPTION(mangle_app_seg)) {
             tls_handle_post_arch_prctl(dcontext, dcontext->sys_param0,
@@ -8721,8 +8265,8 @@ post_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    endif
-#    ifdef LINUX
+#endif
+#ifdef LINUX
     case SYS_ppoll: {
         handle_post_extended_syscall_sigmasks(dcontext, success);
         set_syscall_param(dcontext, 3, dcontext->sys_param3);
@@ -8746,28 +8290,28 @@ post_system_call(dcontext_t *dcontext)
         set_syscall_param(dcontext, 4, dcontext->sys_param4);
         break;
     }
-#    endif
+#endif
 
     /****************************************************************************/
     /* FILES */
 
-#    ifdef SYS_dup2
+#ifdef SYS_dup2
     case SYS_dup2: IF_LINUX(case SYS_dup3:) {
-#        ifdef LINUX
+#    ifdef LINUX
             if (success) {
                 signal_handle_dup(dcontext, (file_t)sys_param(dcontext, 1),
                                   (file_t)result);
             }
-#        endif
+#    endif
             break;
         }
-#    endif
+#endif
 
-#    ifdef MACOS
+#ifdef MACOS
     case SYS_fcntl_nocancel:
-#    endif
+#endif
     case SYS_fcntl: {
-#    ifdef LINUX /* Linux-only since only for signalfd */
+#ifdef LINUX /* Linux-only since only for signalfd */
         if (success) {
             file_t fd = (long)dcontext->sys_param0;
             int cmd = (int)dcontext->sys_param1;
@@ -8775,7 +8319,7 @@ post_system_call(dcontext_t *dcontext)
                 signal_handle_dup(dcontext, fd, (file_t)result);
         }
         break;
-#    endif
+#endif
     }
 
     case IF_MACOS_ELSE(SYS_getrlimit, IF_X64_ELSE(SYS_getrlimit, SYS_ugetrlimit)): {
@@ -8790,7 +8334,7 @@ post_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    if !defined(ARM) && !defined(X64) && !defined(MACOS)
+#if !defined(ARM) && !defined(X64) && !defined(MACOS)
     /* Old struct w/ smaller fields */
     case SYS_getrlimit: {
         int resource = dcontext->sys_param0;
@@ -8803,9 +8347,9 @@ post_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    endif
+#endif
 
-#    ifdef LINUX
+#ifdef LINUX
     case SYS_prlimit64: {
         int resource = dcontext->sys_param1;
         rlimit64_t *rlim = (rlimit64_t *)dcontext->sys_param3;
@@ -8816,12 +8360,12 @@ post_system_call(dcontext_t *dcontext)
         }
         break;
     }
-#    endif
+#endif
 
-#    ifdef LINUX
-#        ifdef SYS_readlink
+#ifdef LINUX
+#    ifdef SYS_readlink
     case SYS_readlink:
-#        endif
+#    endif
     case SYS_readlinkat:
         if (success && DYNAMO_OPTION(early_inject)) {
             bool is_at = (sysnum == SYS_readlinkat);
@@ -8846,18 +8390,18 @@ post_system_call(dcontext_t *dcontext)
             }
         }
         break;
-#    endif
+#endif
 
     default:
-#    ifdef LINUX
+#ifdef LINUX
         handle_restartable_region_syscall_post(dcontext, success);
-#    endif
-#    ifdef VMX86_SERVER
+#endif
+#ifdef VMX86_SERVER
         if (is_vmkuw_sysnum(sysnum)) {
             vmkuw_post_system_call(dcontext);
             break;
         }
-#    endif
+#endif
         break;
 
     } /* switch */
@@ -8881,7 +8425,7 @@ post_system_call(dcontext_t *dcontext)
 
 exit_post_system_call:
 
-#    ifdef CLIENT_INTERFACE
+#ifdef CLIENT_INTERFACE
     /* The instrument_post_syscall should be called after DR finishes all
      * its operations, since DR needs to know the real syscall results,
      * and any changes made by the client are simply to fool the app.
@@ -8890,7 +8434,7 @@ exit_post_system_call:
      */
     /* after restore of xbp so client sees it as though was sysenter */
     instrument_post_syscall(dcontext, sysnum);
-#    endif
+#endif
 
     dcontext->whereami = old_whereami;
 }
@@ -8911,14 +8455,14 @@ get_dynamo_library_bounds(void)
     app_pc check_start, check_end;
     char *libdir;
     const char *dynamorio_libname;
-#    ifdef STATIC_LIBRARY
+#ifdef STATIC_LIBRARY
     /* We don't know our image name, so look up our bounds with an internal
      * address.
      */
     dynamorio_libname = NULL;
     check_start = (app_pc)&get_dynamo_library_bounds;
-#    else /* !STATIC_LIBRARY */
-#        ifdef LINUX
+#else /* !STATIC_LIBRARY */
+#    ifdef LINUX
     /* PR 361594: we get our bounds from linker-provided symbols.
      * Note that referencing the value of these symbols will crash:
      * always use the address only.
@@ -8926,20 +8470,20 @@ get_dynamo_library_bounds(void)
     extern int dynamorio_so_start, dynamorio_so_end;
     dynamo_dll_start = (app_pc)&dynamorio_so_start;
     dynamo_dll_end = (app_pc)ALIGN_FORWARD(&dynamorio_so_end, PAGE_SIZE);
-#        elif defined(MACOS)
+#    elif defined(MACOS)
     dynamo_dll_start = module_dynamorio_lib_base();
-#        endif
+#    endif
     check_start = dynamo_dll_start;
-#    endif /* STATIC_LIBRARY */
+#endif /* STATIC_LIBRARY */
 
     static char dynamorio_libname_buf[MAXIMUM_PATH];
     res = memquery_library_bounds(NULL, &check_start, &check_end, dynamorio_library_path,
                                   BUFFER_SIZE_ELEMENTS(dynamorio_library_path),
                                   dynamorio_libname_buf,
                                   BUFFER_SIZE_ELEMENTS(dynamorio_libname_buf));
-#    ifndef STATIC_LIBRARY
+#ifndef STATIC_LIBRARY
     dynamorio_libname = IF_UNIT_TEST_ELSE(UNIT_TEST_EXE_NAME, dynamorio_libname_buf);
-#    endif /* STATIC_LIBRARY */
+#endif /* STATIC_LIBRARY */
 
     LOG(GLOBAL, LOG_VMAREAS, 1, PRODUCT_NAME " library path: %s\n",
         dynamorio_library_path);
@@ -8949,15 +8493,15 @@ get_dynamo_library_bounds(void)
     LOG(GLOBAL, LOG_VMAREAS, 1, PRODUCT_NAME " library file path: %s\n",
         dynamorio_library_filepath);
     NULL_TERMINATE_BUFFER(dynamorio_library_filepath);
-#    if !defined(STATIC_LIBRARY) && defined(LINUX)
+#if !defined(STATIC_LIBRARY) && defined(LINUX)
     ASSERT(check_start == dynamo_dll_start && check_end == dynamo_dll_end);
-#    elif defined(MACOS)
+#elif defined(MACOS)
     ASSERT(check_start == dynamo_dll_start);
     dynamo_dll_end = check_end;
-#    else
+#else
     dynamo_dll_start = check_start;
     dynamo_dll_end = check_end;
-#    endif
+#endif
     LOG(GLOBAL, LOG_VMAREAS, 1, "DR library bounds: " PFX " to " PFX "\n",
         dynamo_dll_start, dynamo_dll_end);
     ASSERT(res > 0);
@@ -8998,7 +8542,7 @@ get_dynamorio_library_path(void)
     return dynamorio_library_filepath;
 }
 
-#    ifdef LINUX
+#ifdef LINUX
 /* Get full path+name of executable file from /proc/self/exe.  Returns an empty
  * string on error.
  * FIXME i#47: This will return DR's path when using early injection.
@@ -9008,9 +8552,9 @@ read_proc_self_exe(bool ignore_cache)
 {
     static char exepath[MAXIMUM_PATH];
     static bool tried = false;
-#        ifdef MACOS
+#    ifdef MACOS
     ASSERT_NOT_IMPLEMENTED(false);
-#        endif
+#    endif
     if (!tried || ignore_cache) {
         tried = true;
         /* assume we have /proc/self/exe symlink: could add HAVE_PROC_EXE
@@ -9024,26 +8568,26 @@ read_proc_self_exe(bool ignore_cache)
         ASSERT(len > 0);
         NULL_TERMINATE_BUFFER(exepath);
         /* i#960: readlink does not null terminate, so we do it. */
-#        ifdef SYS_readlink
+#    ifdef SYS_readlink
         res = dynamorio_syscall(SYS_readlink, 3, exepath, exepath,
                                 BUFFER_SIZE_ELEMENTS(exepath) - 1);
-#        else
+#    else
         res = dynamorio_syscall(SYS_readlinkat, 4, AT_FDCWD, exepath, exepath,
                                 BUFFER_SIZE_ELEMENTS(exepath) - 1);
-#        endif
+#    endif
         ASSERT(res < BUFFER_SIZE_ELEMENTS(exepath));
         exepath[MAX(res, 0)] = '\0';
         NULL_TERMINATE_BUFFER(exepath);
     }
     return exepath;
 }
-#    endif /* LINUX */
+#endif /* LINUX */
 
 app_pc
 get_application_base(void)
 {
     if (executable_start == NULL) {
-#    ifdef HAVE_MEMINFO
+#ifdef HAVE_MEMINFO
         /* Haven't done find_executable_vm_areas() yet so walk maps ourselves */
         const char *name = get_application_name();
         if (name != NULL && name[0] != '\0') {
@@ -9052,9 +8596,9 @@ get_application_base(void)
                                     NULL, 0);
             ASSERT(count > 0 && executable_start != NULL);
         }
-#    else
+#else
         /* We have to fail.  Should we dl_iterate this early? */
-#    endif
+#endif
     }
     return executable_start;
 }
@@ -9087,26 +8631,26 @@ get_image_entry()
     return image_entry_point;
 }
 
-#    ifdef DEBUG
+#ifdef DEBUG
 void
 mem_stats_snapshot()
 {
     /* FIXME: NYI */
 }
-#    endif
+#endif
 
 bool
 is_in_dynamo_dll(app_pc pc)
 {
     ASSERT(dynamo_dll_start != NULL);
-#    ifdef VMX86_SERVER
+#ifdef VMX86_SERVER
     /* We want to consider vmklib as part of the DR lib for allowing
      * execution (_init calls os_in_vmkernel_classic()) and for
      * reporting crashes as our fault
      */
     if (vmk_in_vmklib(pc))
         return true;
-#    endif
+#endif
     return (pc >= dynamo_dll_start && pc < dynamo_dll_end);
 }
 
@@ -9139,7 +8683,7 @@ get_dynamorio_dll_preferred_base()
 static void
 found_vsyscall_page(memquery_iter_t *iter _IF_DEBUG(OUT const char **map_type))
 {
-#    ifndef X64
+#ifndef X64
     /* We assume no vsyscall page for x64; thus, checking the
      * hardcoded address shouldn't have any false positives.
      */
@@ -9166,7 +8710,7 @@ found_vsyscall_page(memquery_iter_t *iter _IF_DEBUG(OUT const char **map_type))
         vsyscall_page_start = iter->vm_start;
     LOG(GLOBAL, LOG_VMAREAS, 1, "found vdso/vsyscall pages @ " PFX " %s\n",
         vsyscall_page_start, iter->comment);
-#    else
+#else
     /* i#172
      * fix bugs for OS where vdso page is set unreadable as below
      * ffffffffff600000-ffffffffffe00000 ---p 00000000 00:00 0 [vdso]
@@ -9186,23 +8730,23 @@ found_vsyscall_page(memquery_iter_t *iter _IF_DEBUG(OUT const char **map_type))
     else if (strncmp(iter->comment, VSYSCALL_REGION_MAPS_NAME,
                      strlen(VSYSCALL_REGION_MAPS_NAME)) == 0)
         vsyscall_page_start = iter->vm_start;
-#    endif
+#endif
 }
 
 int
 os_walk_address_space(memquery_iter_t *iter, bool add_modules)
 {
     int count = 0;
-#    ifdef MACOS
+#ifdef MACOS
     app_pc shared_start, shared_end;
     bool have_shared = module_dyld_shared_region(&shared_start, &shared_end);
-#    endif
-#    ifdef RETURN_AFTER_CALL
+#endif
+#ifdef RETURN_AFTER_CALL
     dcontext_t *dcontext = get_thread_private_dcontext();
     os_thread_data_t *ostd = (os_thread_data_t *)dcontext->os_field;
-#    endif
+#endif
 
-#    ifndef HAVE_MEMINFO_QUERY
+#ifndef HAVE_MEMINFO_QUERY
     /* We avoid tracking the innards of vmheap for all_memory_areas by
      * adding a single no-access region for the whole vmheap.
      * Queries from heap routines use _from_os.
@@ -9220,11 +8764,11 @@ os_walk_address_space(memquery_iter_t *iter, bool add_modules)
         memcache_update_locked(our_heap_start, our_heap_end, MEMPROT_NONE,
                                DR_MEMTYPE_DATA, false /*!exists*/);
     }
-#    endif
+#endif
 
-#    ifndef HAVE_MEMINFO
+#ifndef HAVE_MEMINFO
     count = find_vm_areas_via_probe();
-#    else
+#else
     while (memquery_iterator_next(iter)) {
         bool image = false;
         size_t size = iter->vm_end - iter->vm_start;
@@ -9257,7 +8801,7 @@ os_walk_address_space(memquery_iter_t *iter, bool add_modules)
             /* i#479, hide private module and match Windows's behavior */
             LOG(GLOBAL, LOG_VMAREAS, 2, PFX "-" PFX " skipping: internal DR region\n",
                 iter->vm_start, iter->vm_end);
-#        ifdef MACOS
+#    ifdef MACOS
         } else if (have_shared && iter->vm_start >= shared_start &&
                    iter->vm_start < shared_end) {
             /* Skip modules we happen to find inside the dyld shared cache,
@@ -9265,7 +8809,7 @@ os_walk_address_space(memquery_iter_t *iter, bool add_modules)
              * in module_walk_dyld_list instead.
              */
             image = true;
-#        endif
+#    endif
         } else if (strncmp(iter->comment, VSYSCALL_PAGE_MAPS_NAME,
                            strlen(VSYSCALL_PAGE_MAPS_NAME)) == 0 ||
                    IF_X64_ELSE(strncmp(iter->comment, VSYSCALL_REGION_MAPS_NAME,
@@ -9306,12 +8850,12 @@ os_walk_address_space(memquery_iter_t *iter, bool add_modules)
                 "\t" PFX "-" PFX "%s inode=" UINT64_FORMAT_STRING " name=%s\n",
                 iter->vm_start, iter->vm_end, TEST(MEMPROT_EXEC, iter->prot) ? " +x" : "",
                 iter->inode, iter->comment);
-#        ifdef LINUX
+#    ifdef LINUX
             /* Mapped images should have inodes, except for cases where an anon
              * map is placed on top (i#2566)
              */
             ASSERT_CURIOSITY(iter->inode != 0 || iter->comment[0] == '\0');
-#        endif
+#    endif
             ASSERT_CURIOSITY(iter->offset == 0); /* first map shouldn't have offset */
             /* Get size by walking the program headers.  This includes .bss. */
             if (module_walk_program_headers(iter->vm_start, size, false,
@@ -9329,7 +8873,7 @@ os_walk_address_space(memquery_iter_t *iter, bool add_modules)
 
             if (add_modules) {
                 /* look for executable */
-#        ifdef LINUX
+#    ifdef LINUX
                 exec_match = get_application_name();
                 if (exec_match != NULL && exec_match[0] != '\0')
                     found_exec = (strcmp(iter->comment, exec_match) == 0);
@@ -9337,7 +8881,7 @@ os_walk_address_space(memquery_iter_t *iter, bool add_modules)
                 if (!found_exec && executable_start != NULL &&
                     executable_start == iter->vm_start)
                     found_exec = true;
-#        else
+#    else
                 /* We don't have a nice normalized name: it can have ./ or ../ inside
                  * it.  But, we can distinguish an exe from a lib here, even for PIE,
                  * so we go with that plus a basename comparison.
@@ -9352,7 +8896,7 @@ os_walk_address_space(memquery_iter_t *iter, bool add_modules)
                         iter_basename++;
                     found_exec = (strcmp(iter_basename, exec_match) == 0);
                 }
-#        endif
+#    endif
                 if (found_exec) {
                     if (executable_start == NULL)
                         executable_start = iter->vm_start;
@@ -9369,18 +8913,18 @@ os_walk_address_space(memquery_iter_t *iter, bool add_modules)
                 module_list_add(iter->vm_start, mod_first_end - mod_base, false,
                                 iter->comment, iter->inode);
 
-#        ifdef MACOS
+#    ifdef MACOS
                 /* look for dyld */
                 if (strcmp(iter->comment, "/usr/lib/dyld") == 0)
                     module_walk_dyld_list(iter->vm_start);
-#        endif
+#    endif
             }
         } else if (iter->inode != 0) {
             DODEBUG({ map_type = "Mapped File"; });
         }
 
         /* add all regions (incl. dynamo_areas and stack) to all_memory_areas */
-#        ifndef HAVE_MEMINFO_QUERY
+#    ifndef HAVE_MEMINFO_QUERY
         /* Don't add if we're using one single vmheap entry. */
         if (iter->vm_start < our_heap_start || iter->vm_end > our_heap_end) {
             LOG(GLOBAL, LOG_VMAREAS, 4,
@@ -9390,7 +8934,7 @@ os_walk_address_space(memquery_iter_t *iter, bool add_modules)
                                    image ? DR_MEMTYPE_IMAGE : DR_MEMTYPE_DATA,
                                    false /*!exists*/);
         }
-#        endif
+#    endif
 
         /* FIXME: best if we could pass every region to vmareas, but
          * it has no way of determining if this is a stack b/c we don't have
@@ -9403,17 +8947,17 @@ os_walk_address_space(memquery_iter_t *iter, bool add_modules)
             count++;
         }
     }
-#    endif /* !HAVE_MEMINFO */
+#endif /* !HAVE_MEMINFO */
 
-#    ifndef HAVE_MEMINFO_QUERY
+#ifndef HAVE_MEMINFO_QUERY
     DOLOG(4, LOG_VMAREAS, memcache_print(GLOBAL, "init: all memory areas:\n"););
-#    endif
+#endif
 
-#    ifdef RETURN_AFTER_CALL
+#ifdef RETURN_AFTER_CALL
     /* Find the bottom of the stack of the initial (native) entry */
     ostd->stack_bottom_pc = find_stack_bottom();
     LOG(THREAD, LOG_ALL, 1, "Stack bottom pc = " PFX "\n", ostd->stack_bottom_pc);
-#    endif
+#endif
 
     /* now that we've walked memory print all modules */
     LOG(GLOBAL, LOG_VMAREAS, 2, "Module list after memory walk\n");
@@ -9450,7 +8994,7 @@ find_executable_vm_areas(void)
 int
 find_dynamo_library_vm_areas(void)
 {
-#    ifndef STATIC_LIBRARY
+#ifndef STATIC_LIBRARY
     /* We didn't add inside get_dynamo_library_bounds b/c it was called pre-alloc.
      * We don't bother to break down the sub-regions.
      * Assumption: we don't need to have the protection flags for DR sub-regions.
@@ -9460,11 +9004,11 @@ find_dynamo_library_vm_areas(void)
     add_dynamo_vm_area(get_dynamorio_dll_start(), get_dynamorio_dll_end(),
                        MEMPROT_READ | MEMPROT_WRITE | MEMPROT_EXEC,
                        true /* from image */ _IF_DEBUG(dynamorio_library_filepath));
-#    endif
-#    ifdef VMX86_SERVER
+#endif
+#ifdef VMX86_SERVER
     if (os_in_vmkernel_userworld())
         vmk_add_vmklib_to_dynamo_areas();
-#    endif
+#endif
     return 1;
 }
 
@@ -9510,7 +9054,7 @@ get_stack_bounds(dcontext_t *dcontext, byte **base, byte **top)
     return true;
 }
 
-#    ifdef RETURN_AFTER_CALL
+#ifdef RETURN_AFTER_CALL
 initial_call_stack_status_t
 at_initial_stack_bottom(dcontext_t *dcontext, app_pc target_pc)
 {
@@ -9533,17 +9077,17 @@ at_initial_stack_bottom(dcontext_t *dcontext, app_pc target_pc)
         }
     }
 }
-#    endif /* RETURN_AFTER_CALL */
+#endif /* RETURN_AFTER_CALL */
 
 /* Uses our cached data structures (if in use, else raw query) to retrieve memory info */
 bool
 query_memory_ex(const byte *pc, OUT dr_mem_info_t *out_info)
 {
-#    ifdef HAVE_MEMINFO_QUERY
+#ifdef HAVE_MEMINFO_QUERY
     return query_memory_ex_from_os(pc, out_info);
-#    else
+#else
     return memcache_query_memory(pc, out_info);
-#    endif
+#endif
 }
 
 bool
@@ -9633,7 +9177,7 @@ deadlock_avoidance_unlock(mutex_t *lock, bool ownable);
 void
 mutex_wait_contended_lock(mutex_t *lock _IF_CLIENT_INTERFACE(priv_mcontext_t *mc))
 {
-#    ifdef CLIENT_INTERFACE
+#ifdef CLIENT_INTERFACE
     dcontext_t *dcontext = get_thread_private_dcontext();
     bool set_client_safe_for_synch =
         ((dcontext != NULL) && IS_CLIENT_THREAD(dcontext) &&
@@ -9647,7 +9191,7 @@ mutex_wait_contended_lock(mutex_t *lock _IF_CLIENT_INTERFACE(priv_mcontext_t *mc
         ASSERT(!set_client_safe_for_synch);
         *get_mcontext(dcontext) = *mc;
     }
-#    endif
+#endif
 
     /* i#96/PR 295561: use futex(2) if available */
     if (ksynch_kernel_support()) {
@@ -9656,42 +9200,42 @@ mutex_wait_contended_lock(mutex_t *lock _IF_CLIENT_INTERFACE(priv_mcontext_t *mc
          * the next unlocker will call mutex_notify_released_lock().
          */
         ptr_int_t res;
-#    ifndef LINUX /* we actually don't use this for Linux: see below */
+#ifndef LINUX /* we actually don't use this for Linux: see below */
         KSYNCH_TYPE *event = mutex_get_contended_event(lock);
         ASSERT(event != NULL && ksynch_var_initialized(event));
-#    endif
+#endif
         while (atomic_exchange_int(&lock->lock_requests, LOCK_CONTENDED_STATE) !=
                LOCK_FREE_STATE) {
-#    ifdef CLIENT_INTERFACE
+#ifdef CLIENT_INTERFACE
             if (set_client_safe_for_synch)
                 dcontext->client_data->client_thread_safe_for_synch = true;
             if (mc != NULL)
                 set_synch_state(dcontext, THREAD_SYNCH_VALID_MCONTEXT);
-#    endif
+#endif
 
                 /* Unfortunately the synch semantics are different for Linux vs Mac.
                  * We have to use lock_requests as the futex to avoid waiting if
                  * lock_requests changes, while on Mac the underlying synch prevents
                  * a wait there.
                  */
-#    ifdef LINUX
+#ifdef LINUX
             /* We'll abort the wait if lock_requests has changed at all.
              * We can't have a series of changes that result in no apparent
              * change w/o someone acquiring the lock, b/c
              * mutex_notify_released_lock() sets lock_requests to LOCK_FREE_STATE.
              */
             res = ksynch_wait(&lock->lock_requests, LOCK_CONTENDED_STATE, 0);
-#    else
+#else
             res = ksynch_wait(event, 0, 0);
-#    endif
+#endif
             if (res != 0 && res != -EWOULDBLOCK)
                 os_thread_yield();
-#    ifdef CLIENT_INTERFACE
+#ifdef CLIENT_INTERFACE
             if (set_client_safe_for_synch)
                 dcontext->client_data->client_thread_safe_for_synch = false;
             if (mc != NULL)
                 set_synch_state(dcontext, THREAD_SYNCH_NONE);
-#    endif
+#endif
 
             /* we don't care whether properly woken (res==0), var mismatch
              * (res==-EWOULDBLOCK), or error: regardless, someone else
@@ -9703,29 +9247,29 @@ mutex_wait_contended_lock(mutex_t *lock _IF_CLIENT_INTERFACE(priv_mcontext_t *mc
         atomic_dec_and_test(&lock->lock_requests);
 
         while (!mutex_trylock(lock)) {
-#    ifdef CLIENT_INTERFACE
+#ifdef CLIENT_INTERFACE
             if (set_client_safe_for_synch)
                 dcontext->client_data->client_thread_safe_for_synch = true;
             if (mc != NULL)
                 set_synch_state(dcontext, THREAD_SYNCH_VALID_MCONTEXT);
-#    endif
+#endif
 
             os_thread_yield();
-#    ifdef CLIENT_INTERFACE
+#ifdef CLIENT_INTERFACE
             if (set_client_safe_for_synch)
                 dcontext->client_data->client_thread_safe_for_synch = false;
             if (mc != NULL)
                 set_synch_state(dcontext, THREAD_SYNCH_NONE);
-#    endif
+#endif
         }
 
-#    ifdef DEADLOCK_AVOIDANCE
+#ifdef DEADLOCK_AVOIDANCE
         /* HACK: trylock's success causes it to do DEADLOCK_AVOIDANCE_LOCK, so to
          * avoid two in a row (causes assertion on owner) we unlock here
          * In the future we will remove the trylock here and this will go away.
          */
         deadlock_avoidance_unlock(lock, true);
-#    endif
+#endif
     }
 
     return;
@@ -9741,11 +9285,11 @@ mutex_notify_released_lock(mutex_t *lock)
          */
         lock->lock_requests = LOCK_FREE_STATE;
         /* No reason to wake multiple threads: just one */
-#    ifdef LINUX
+#ifdef LINUX
         ksynch_wake(&lock->lock_requests);
-#    else
+#else
         ksynch_wake(&lock->contended_event);
-#    endif
+#endif
     } /* else nothing to do */
 }
 
@@ -9846,9 +9390,9 @@ reset_event(event_t e)
 bool
 wait_for_event(event_t e, int timeout_ms)
 {
-#    ifdef DEBUG
+#ifdef DEBUG
     dcontext_t *dcontext = get_thread_private_dcontext();
-#    endif
+#endif
     uint64 start_time, cur_time;
     if (timeout_ms > 0)
         start_time = query_time_millis();
@@ -9900,23 +9444,23 @@ wait_for_event(event_t e, int timeout_ms)
  * that as the next entry.
  */
 struct linux_dirent {
-#    ifdef SYS_getdents
+#ifdef SYS_getdents
     /* Adapted from struct old_linux_dirent in linux/fs/readdir.c: */
     unsigned long d_ino;
     unsigned long d_off;
     unsigned short d_reclen;
     char d_name[];
-#    else
+#else
     /* Adapted from struct linux_dirent64 in linux/include/linux/dirent.h: */
     uint64 d_ino;
     int64 d_off;
     unsigned short d_reclen;
     unsigned char d_type;
     char d_name[];
-#    endif
+#endif
 };
 
-#    define CURRENT_DIRENT(iter) ((struct linux_dirent *)(&iter->buf[iter->off]))
+#define CURRENT_DIRENT(iter) ((struct linux_dirent *)(&iter->buf[iter->off]))
 
 static void
 os_dir_iterator_start(dir_iterator_t *iter, file_t fd)
@@ -9929,13 +9473,13 @@ os_dir_iterator_start(dir_iterator_t *iter, file_t fd)
 static bool
 os_dir_iterator_next(dir_iterator_t *iter)
 {
-#    ifdef MACOS
+#ifdef MACOS
     /* We can use SYS_getdirentries, but do we even need a dir iterator?
      * On Linux it's only used to enumerate /proc/pid/task.
      */
     ASSERT_NOT_IMPLEMENTED(false);
     return false;
-#    else
+#else
     if (iter->off < iter->end) {
         /* Have existing dents, get the next offset. */
         iter->off += CURRENT_DIRENT(iter)->d_reclen;
@@ -9948,13 +9492,13 @@ os_dir_iterator_next(dir_iterator_t *iter)
          * the example code that this is based on.
          */
         iter->off = 0;
-#        ifdef SYS_getdents
+#    ifdef SYS_getdents
         iter->end =
             dynamorio_syscall(SYS_getdents, 3, iter->fd, iter->buf, sizeof(iter->buf));
-#        else
+#    else
         iter->end =
             dynamorio_syscall(SYS_getdents64, 3, iter->fd, iter->buf, sizeof(iter->buf));
-#        endif
+#    endif
         ASSERT(iter->end <= sizeof(iter->buf));
         if (iter->end <= 0) { /* No more dents, or error. */
             iter->name = NULL;
@@ -9967,7 +9511,7 @@ os_dir_iterator_next(dir_iterator_t *iter)
     }
     iter->name = CURRENT_DIRENT(iter)->d_name;
     return true;
-#    endif
+#endif
 }
 
 /***************************************************************************
@@ -10013,7 +9557,7 @@ os_list_threads(dcontext_t *dcontext, uint *num_threads_out)
 
     ASSERT(num_threads_out != NULL);
 
-#    ifdef MACOS
+#ifdef MACOS
     /* XXX i#58: NYI.
      * We may want SYS_proc_info with PROC_INFO_PID_INFO and PROC_PIDLISTTHREADS,
      * or is that just BSD threads and instead we want process_set_tasks()
@@ -10021,7 +9565,7 @@ os_list_threads(dcontext_t *dcontext, uint *num_threads_out)
      */
     *num_threads_out = 0;
     return NULL;
-#    endif
+#endif
 
     tids =
         HEAP_ARRAY_ALLOC(dcontext, thread_id_t, tids_alloced, ACCT_THREAD_MGT, PROTECTED);
@@ -10196,7 +9740,7 @@ os_take_over_all_unknown_threads(dcontext_t *dcontext)
 bool
 os_thread_re_take_over(void)
 {
-#    ifdef X86
+#ifdef X86
     /* i#2089: is_thread_initialized() will fail for a currently-native app.
      * We bypass the magic field checks here of is_thread_tls_initialized().
      * XXX: should this be inside is_thread_initialized()?  But that may mislead
@@ -10218,7 +9762,7 @@ os_thread_re_take_over(void)
             return true;
         }
     }
-#    endif
+#endif
     return false;
 }
 
@@ -10361,7 +9905,7 @@ os_random_seed(void)
     return seed;
 }
 
-#    ifdef RCT_IND_BRANCH
+#ifdef RCT_IND_BRANCH
 /* Analyze a range in a possibly new module
  * return false if not a code section in a module
  * otherwise returns true and adds all valid targets for rct_ind_branch_check
@@ -10411,17 +9955,17 @@ rct_analyze_module_at_violation(dcontext_t *dcontext, app_pc target_pc)
     return false;
 }
 
-#        ifdef X64
+#    ifdef X64
 bool
 rct_add_rip_rel_addr(dcontext_t *dcontext, app_pc tgt _IF_DEBUG(app_pc src))
 {
     /* FIXME PR 276762: not implemented */
     return false;
 }
-#        endif
-#    endif /* RCT_IND_BRANCH */
+#    endif
+#endif /* RCT_IND_BRANCH */
 
-#    ifdef HOT_PATCHING_INTERFACE
+#ifdef HOT_PATCHING_INTERFACE
 void *
 get_drmarker_hotp_policy_status_table()
 {
@@ -10456,7 +10000,7 @@ insert_jmp_at_tramp_entry(dcontext_t *dcontext, byte *trampoline, byte *target)
 {
     ASSERT_NOT_IMPLEMENTED(false);
 }
-#    endif /* HOT_PATCHING_INTERFACE */
+#endif /* HOT_PATCHING_INTERFACE */
 
 bool
 aslr_is_possible_attack(app_pc target)
@@ -10517,7 +10061,7 @@ os_check_option_compatibility(void)
     return false;
 }
 
-#    ifdef X86_32
+#ifdef X86_32
 /* Emulate uint64 modulo and division by uint32 on ia32.
  * XXX: Does *not* handle 64-bit divisors!
  */
@@ -10625,17 +10169,17 @@ __divdi3(int64 dividend, int64 divisor)
 
 /* __moddi3 is coming from third_party/libgcc for x86 as well as arm. */
 
-#    elif defined(ARM)
+#elif defined(ARM)
 /* i#1566: for ARM, __aeabi versions are used instead of udivdi3 and umoddi3.
  * We link with __aeabi routines from libgcc via third_party/libgcc.
  */
-#    endif /* X86_32 */
+#endif /* X86_32 */
 
 /****************************************************************************
  * Kernel-restartable sequences
  */
 
-#    ifdef LINUX
+#ifdef LINUX
 /* Support for Linux kernel extensions for per-cpu critical regions.
  * Xref https://lwn.net/Articles/649288/
  * Some of this may vary on different kernels.
@@ -10657,9 +10201,9 @@ __divdi3(int64 dividend, int64 divisor)
  *   SYSCALL_DEFINE4(rseq, int, op, long, val1, long, val2, long, val3)
  */
 /* Set operation: app_pc start, app_pc end, app_pc restart */
-#        define RSEQ_SET_CRITICAL 1
+#    define RSEQ_SET_CRITICAL 1
 /* Get operation: app_pc *start, app_pc *end, app_pc *restart */
-#        define RSEQ_GET_CRITICAL 3
+#    define RSEQ_GET_CRITICAL 3
 
 static app_pc app_restart_region_start;
 static app_pc app_restart_region_end;
@@ -10748,121 +10292,14 @@ handle_restartable_region_syscall_post(dcontext_t *dcontext, bool success)
         }
     }
 }
-#    endif /* LINUX */
+#endif /* LINUX */
 
 void
 native_exec_os_init(void)
 {
-#    ifdef LINUX
+#ifdef LINUX
     restartable_region_init();
-#    endif
-}
-
-#endif /* !NOT_DYNAMORIO_CORE_PROPER: around most of file, to exclude preload */
-
-/****************************************************************************
- * Page size discovery and query
- */
-
-/* This variable is only used by os_set_page_size and os_page_size, but those
- * functions may be called before libdynamorio.so has been relocated. So check
- * the disassembly of those functions: there should be no relocations.
- */
-static size_t page_size = 0;
-
-/* Return true if size is a multiple of the page size.
- * XXX: This function may be called when DynamoRIO is in a fragile state, or not
- * yet relocated, so keep this self-contained and do not use global variables or
- * logging.
- */
-static bool
-os_try_page_size(size_t size)
-{
-    byte *addr =
-        mmap_syscall(NULL, size * 2, PROT_NONE, MAP_PRIVATE | MAP_ANONYMOUS, -1, 0);
-    if ((ptr_uint_t)addr >= (ptr_uint_t)-4096) /* mmap failed: should not happen */
-        return false;
-    if (munmap_syscall(addr + size, size) == 0) {
-        /* munmap of top half succeeded: munmap bottom half and return true */
-        munmap_syscall(addr, size);
-        return true;
-    }
-    /* munmap of top half failed: munmap whole region and return false */
-    munmap_syscall(addr, size * 2);
-    return false;
-}
-
-/* Directly determine the granularity of memory allocation using mmap and munmap.
- * This is used as a last resort if the page size is required before it has been
- * discovered in any other way, such as from AT_PAGESZ.
- * XXX: This function may be called when DynamoRIO is in a fragile state, or not
- * yet relocated, so keep this self-contained and do not use global variables or
- * logging.
- */
-static size_t
-os_find_page_size(void)
-{
-    size_t size = 4096;
-    if (os_try_page_size(size)) {
-        /* Try smaller sizes. */
-        for (size /= 2; size > 0; size /= 2) {
-            if (!os_try_page_size(size))
-                return size * 2;
-        }
-    } else {
-        /* Try larger sizes. */
-        for (size *= 2; size * 2 > 0; size *= 2) {
-            if (os_try_page_size(size))
-                return size;
-        }
-    }
-    /* Something went wrong... */
-    return 4096;
-}
-
-static void
-os_set_page_size(size_t size)
-{
-    page_size = size; /* atomic write */
-}
-
-size_t
-os_page_size(void)
-{
-    size_t size = page_size; /* atomic read */
-    if (size == 0) {
-        /* XXX: On Mac OSX we should use sysctl_query on hw.pagesize. */
-        size = os_find_page_size();
-        os_set_page_size(size);
-    }
-    return size;
-}
-
-void
-os_page_size_init(const char **env)
-{
-#if defined(LINUX) && !defined(STATIC_LIBRARY)
-    /* On Linux we get the page size from the auxiliary vector, which is what
-     * the C library typically does for implementing sysconf(_SC_PAGESIZE).
-     * However, for STATIC_LIBRARY, our_environ is not guaranteed to point
-     * at the stack as we're so late, so we do not try to read off the end of it
-     * (i#2122).
-     */
-    size_t size = page_size; /* atomic read */
-    if (size == 0) {
-        ELF_AUXV_TYPE *auxv;
-        /* Skip environment. */
-        while (*env != 0)
-            ++env;
-        /* Look for AT_PAGESZ in the auxiliary vector. */
-        for (auxv = (ELF_AUXV_TYPE *)(env + 1); auxv->a_type != AT_NULL; auxv++) {
-            if (auxv->a_type == AT_PAGESZ) {
-                os_set_page_size(auxv->a_un.a_val);
-                break;
-            }
-        }
-    }
-#endif /* LINUX */
+#endif
 }
 
 /****************************************************************************
