+++ make/style_checks/scripts/rules/pp_indent.tcl
@@ -1,5 +1,6 @@
 #!/usr/bin/tclsh
 # **********************************************************
+# Copyright (c) 2019 Google, Inc. All rights reserved.
 # Copyright (c) 2017 ARM Limited. All rights reserved.
 # **********************************************************
 
@@ -71,8 +72,8 @@ set indent_exempt [list {aarch64/instr_create.h}     {1} \
                         {lib/hotpatch_interface.h}   {1} \
                         {lib/instrument.c}           {1} \
                         {lib/instrument_api.h}       {1} \
-                        {lib/dr_helper.h}            {1} \
-                        {lib/statsx.h}                 {1} \
+                        {drlibc/drlibc.h}            {1} \
+                        {lib/statsx.h}               {1} \
                         {unix/os.c}                  {1} \
                         {unix/signal.c}              {1} \
                         {unix/injector.c}            {1} \
